hopr_bindings

Module hopr_announcements

Modules

hopr_announcements

This module was auto-generated with ethers-rs Abigen. More information at: https://github.com/gakonst/ethers-rs

Structs

AddressAnnouncementFilter

AlreadyInitialized

Custom Error type AlreadyInitialized with signature AlreadyInitialized() and selector 0x0dc149f0

AnnounceCall

Container type for all input parameters for the announce function with signature announce(string) and selector 0xea0a5237

AnnounceSafeCall

Container type for all input parameters for the announceSafe function with signature announceSafe(address,string) and selector 0xfad0e5a2

BindKeysAnnounceCall

Container type for all input parameters for the bindKeysAnnounce function with signature bindKeysAnnounce(bytes32,bytes32,bytes32,string) and selector 0xf884a9cb

BindKeysAnnounceSafeCall

Container type for all input parameters for the bindKeysAnnounceSafe function with signature bindKeysAnnounceSafe(address,bytes32,bytes32,bytes32,string) and selector 0xddb9dfc1

BindKeysCall

Container type for all input parameters for the bindKeys function with signature bindKeys(bytes32,bytes32,bytes32) and selector 0xdbb98d91

BindKeysSafeCall

Container type for all input parameters for the bindKeysSafe function with signature bindKeysSafe(address,bytes32,bytes32,bytes32) and selector 0xcae2b434

ContractNotResponsible

Custom Error type ContractNotResponsible with signature ContractNotResponsible() and selector 0xacd5a823

HoprAnnouncements

InvalidSafeAddress

Custom Error type InvalidSafeAddress with signature InvalidSafeAddress() and selector 0x8e9d7c5e

KeyBindingFilter

MultiSigUninitialized

Custom Error type MultiSigUninitialized with signature MultiSigUninitialized() and selector 0x454a20c8

MulticallCall

Container type for all input parameters for the multicall function with signature multicall(bytes[]) and selector 0xac9650d8

MulticallReturn

Container type for all return fields from the multicall function with signature multicall(bytes[]) and selector 0xac9650d8

RevokeAnnouncementFilter

RevokeCall

Container type for all input parameters for the revoke function with signature revoke() and selector 0xb6549f75

RevokeSafeCall

Container type for all input parameters for the revokeSafe function with signature revokeSafe(address) and selector 0x308c712e

Enums

HoprAnnouncementsCalls

Container type for all of the contract’s call

HoprAnnouncementsErrors

Container type for all of the contract’s custom errors

HoprAnnouncementsEvents

Container type for all of the contract’s events

Statics

HOPRANNOUNCEMENTS_ABI

The parsed JSON ABI of the contract.

HOPRANNOUNCEMENTS_BYTECODE

The bytecode of the contract.

HOPRANNOUNCEMENTS_DEPLOYED_BYTECODE

The deployed bytecode of the contract.