hopr_bindings::hopr_announcements

Module HoprAnnouncements

Generated by the following Solidity interface…

library HoprLedger {
    struct RootStruct {
        bytes28 rootHash;
        uint32 timestamp;
    }
}

interface HoprAnnouncements {
    type KeyId is uint32;
    struct KeyBindingWithSignatureTimestamp {
        bytes32 ed25519_sig_0;
        bytes32 ed25519_sig_1;
        bytes32 ed25519_pub_key;
        address chain_key;
        uint256 boundTimestamp;
    }

    error AddressEmptyCode(address target);
    error AlreadyInitialized();
    error ContractNotResponsible();
    error ERC1967InvalidImplementation(address implementation);
    error ERC1967NonPayable();
    error EmptyMultiaddr();
    error ExistingKeyBinding();
    error FailedCall();
    error InvalidInitialization();
    error InvalidKeyBindingFeeAmount();
    error InvalidSafeAddress();
    error InvalidTokenRecipient();
    error InvalidTokenSender();
    error InvalidTokensReceivedUsage();
    error KeyIdOutOfRange();
    error MultiSigUninitialized();
    error NoNeedToProvideKeyBindingFee();
    error NotInitializing();
    error OwnableInvalidOwner(address owner);
    error OwnableUnauthorizedAccount(address account);
    error UUPSUnauthorizedCallContext();
    error UUPSUnsupportedProxiableUUID(bytes32 slot);
    error WrongToken();
    error ZeroAddress(string reason);
    error ZeroInterval();

    event AddressAnnouncement(address node, string baseMultiaddr);
    event Initialized(uint64 version);
    event KeyBinding(bytes32 ed25519_sig_0, bytes32 ed25519_sig_1, bytes32 ed25519_pub_key, address chain_key, uint256 key_id);
    event KeyBindingFeeUpdate(uint256 newFee, uint256 oldFee);
    event LedgerDomainSeparatorUpdated(bytes32 indexed ledgerDomainSeparator);
    event OwnershipTransferred(address indexed previousOwner, address indexed newOwner);
    event RevokeAnnouncement(address node);
    event Upgraded(address indexed implementation);

    constructor();

    function ERC777_HOOK_BIND_KEY_MAYBE_ANNOUNCE_SIZE() external view returns (uint256);
    function LEDGER_VERSION() external view returns (string memory);
    function SNAPSHOT_INTERVAL() external view returns (uint256);
    function TOKEN() external view returns (address);
    function TOKENS_RECIPIENT_INTERFACE_HASH() external view returns (bytes32);
    function UPGRADE_INTERFACE_VERSION() external view returns (string memory);
    function announce(string memory baseMultiaddr) external;
    function announceSafe(address selfAddress, string memory baseMultiaddr) external;
    function getAllKeyBindings() external view returns (KeyBindingWithSignatureTimestamp[] memory);
    function getKeyBindingCount() external view returns (uint256);
    function getKeyBindingWithKeyId(KeyId keyId) external view returns (KeyBindingWithSignatureTimestamp memory);
    function getKeyIdRange() external pure returns (uint32 minKeyId, uint32 maxKeyId);
    function getKeyIdWithOffchainKey(bytes32 ed25519_pub_key) external view returns (bool, KeyId);
    function getOffchainKeyWithKeyId(KeyId keyId) external view returns (bytes32 ed25519_pub_key);
    function initialize(bytes memory initParams) external;
    function isOffchainKeyBound(bytes32 ed25519_pub_key) external view returns (bool);
    function keyBindingFee() external view returns (uint256);
    function latestRoot() external view returns (HoprLedger.RootStruct memory);
    function latestSnapshotRoot() external view returns (HoprLedger.RootStruct memory);
    function ledgerDomainSeparator() external view returns (bytes32);
    function multiaddrOf(address) external view returns (string memory);
    function owner() external view returns (address);
    function proxiableUUID() external view returns (bytes32);
    function renounceOwnership() external;
    function revoke() external;
    function revokeSafe(address selfAddress) external;
    function tokensReceived(address, address from, address to, uint256 amount, bytes memory userData, bytes memory) external;
    function transferOwnership(address newOwner) external;
    function tryGetKeyBinding(bytes32 ed25519_pub_key) external view returns (bool, KeyId, KeyBindingWithSignatureTimestamp memory);
    function updateKeyBindingFee(uint256 newFee) external;
    function updateLedgerDomainSeparator() external;
    function upgradeToAndCall(address newImplementation, bytes memory data) external payable;
}

…which was generated by the following JSON ABI:

[
  {
    "type": "constructor",
    "inputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "ERC777_HOOK_BIND_KEY_MAYBE_ANNOUNCE_SIZE",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "LEDGER_VERSION",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "string",
        "internalType": "string"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "SNAPSHOT_INTERVAL",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "TOKEN",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "contract IERC777"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "TOKENS_RECIPIENT_INTERFACE_HASH",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "UPGRADE_INTERFACE_VERSION",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "string",
        "internalType": "string"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "announce",
    "inputs": [
      {
        "name": "baseMultiaddr",
        "type": "string",
        "internalType": "string"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "announceSafe",
    "inputs": [
      {
        "name": "selfAddress",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "baseMultiaddr",
        "type": "string",
        "internalType": "string"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "getAllKeyBindings",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "tuple[]",
        "internalType": "struct KeyBindingWithSignatureTimestamp[]",
        "components": [
          {
            "name": "ed25519_sig_0",
            "type": "bytes32",
            "internalType": "bytes32"
          },
          {
            "name": "ed25519_sig_1",
            "type": "bytes32",
            "internalType": "bytes32"
          },
          {
            "name": "ed25519_pub_key",
            "type": "bytes32",
            "internalType": "bytes32"
          },
          {
            "name": "chain_key",
            "type": "address",
            "internalType": "address"
          },
          {
            "name": "boundTimestamp",
            "type": "uint256",
            "internalType": "uint256"
          }
        ]
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getKeyBindingCount",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getKeyBindingWithKeyId",
    "inputs": [
      {
        "name": "keyId",
        "type": "uint32",
        "internalType": "KeyId"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "tuple",
        "internalType": "struct KeyBindingWithSignatureTimestamp",
        "components": [
          {
            "name": "ed25519_sig_0",
            "type": "bytes32",
            "internalType": "bytes32"
          },
          {
            "name": "ed25519_sig_1",
            "type": "bytes32",
            "internalType": "bytes32"
          },
          {
            "name": "ed25519_pub_key",
            "type": "bytes32",
            "internalType": "bytes32"
          },
          {
            "name": "chain_key",
            "type": "address",
            "internalType": "address"
          },
          {
            "name": "boundTimestamp",
            "type": "uint256",
            "internalType": "uint256"
          }
        ]
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getKeyIdRange",
    "inputs": [],
    "outputs": [
      {
        "name": "minKeyId",
        "type": "uint32",
        "internalType": "uint32"
      },
      {
        "name": "maxKeyId",
        "type": "uint32",
        "internalType": "uint32"
      }
    ],
    "stateMutability": "pure"
  },
  {
    "type": "function",
    "name": "getKeyIdWithOffchainKey",
    "inputs": [
      {
        "name": "ed25519_pub_key",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      },
      {
        "name": "",
        "type": "uint32",
        "internalType": "KeyId"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getOffchainKeyWithKeyId",
    "inputs": [
      {
        "name": "keyId",
        "type": "uint32",
        "internalType": "KeyId"
      }
    ],
    "outputs": [
      {
        "name": "ed25519_pub_key",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "initialize",
    "inputs": [
      {
        "name": "initParams",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "isOffchainKeyBound",
    "inputs": [
      {
        "name": "ed25519_pub_key",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "keyBindingFee",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "latestRoot",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "tuple",
        "internalType": "struct HoprLedger.RootStruct",
        "components": [
          {
            "name": "rootHash",
            "type": "bytes28",
            "internalType": "bytes28"
          },
          {
            "name": "timestamp",
            "type": "uint32",
            "internalType": "uint32"
          }
        ]
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "latestSnapshotRoot",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "tuple",
        "internalType": "struct HoprLedger.RootStruct",
        "components": [
          {
            "name": "rootHash",
            "type": "bytes28",
            "internalType": "bytes28"
          },
          {
            "name": "timestamp",
            "type": "uint32",
            "internalType": "uint32"
          }
        ]
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "ledgerDomainSeparator",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "multiaddrOf",
    "inputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "string",
        "internalType": "string"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "owner",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "proxiableUUID",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "renounceOwnership",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "revoke",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "revokeSafe",
    "inputs": [
      {
        "name": "selfAddress",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "tokensReceived",
    "inputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "from",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "to",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "amount",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "userData",
        "type": "bytes",
        "internalType": "bytes"
      },
      {
        "name": "",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "transferOwnership",
    "inputs": [
      {
        "name": "newOwner",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "tryGetKeyBinding",
    "inputs": [
      {
        "name": "ed25519_pub_key",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      },
      {
        "name": "",
        "type": "uint32",
        "internalType": "KeyId"
      },
      {
        "name": "",
        "type": "tuple",
        "internalType": "struct KeyBindingWithSignatureTimestamp",
        "components": [
          {
            "name": "ed25519_sig_0",
            "type": "bytes32",
            "internalType": "bytes32"
          },
          {
            "name": "ed25519_sig_1",
            "type": "bytes32",
            "internalType": "bytes32"
          },
          {
            "name": "ed25519_pub_key",
            "type": "bytes32",
            "internalType": "bytes32"
          },
          {
            "name": "chain_key",
            "type": "address",
            "internalType": "address"
          },
          {
            "name": "boundTimestamp",
            "type": "uint256",
            "internalType": "uint256"
          }
        ]
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "updateKeyBindingFee",
    "inputs": [
      {
        "name": "newFee",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "updateLedgerDomainSeparator",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "upgradeToAndCall",
    "inputs": [
      {
        "name": "newImplementation",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "data",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "outputs": [],
    "stateMutability": "payable"
  },
  {
    "type": "event",
    "name": "AddressAnnouncement",
    "inputs": [
      {
        "name": "node",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      },
      {
        "name": "baseMultiaddr",
        "type": "string",
        "indexed": false,
        "internalType": "string"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Initialized",
    "inputs": [
      {
        "name": "version",
        "type": "uint64",
        "indexed": false,
        "internalType": "uint64"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "KeyBinding",
    "inputs": [
      {
        "name": "ed25519_sig_0",
        "type": "bytes32",
        "indexed": false,
        "internalType": "bytes32"
      },
      {
        "name": "ed25519_sig_1",
        "type": "bytes32",
        "indexed": false,
        "internalType": "bytes32"
      },
      {
        "name": "ed25519_pub_key",
        "type": "bytes32",
        "indexed": false,
        "internalType": "bytes32"
      },
      {
        "name": "chain_key",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      },
      {
        "name": "key_id",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "KeyBindingFeeUpdate",
    "inputs": [
      {
        "name": "newFee",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "oldFee",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "LedgerDomainSeparatorUpdated",
    "inputs": [
      {
        "name": "ledgerDomainSeparator",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "OwnershipTransferred",
    "inputs": [
      {
        "name": "previousOwner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "newOwner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "RevokeAnnouncement",
    "inputs": [
      {
        "name": "node",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Upgraded",
    "inputs": [
      {
        "name": "implementation",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "error",
    "name": "AddressEmptyCode",
    "inputs": [
      {
        "name": "target",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "AlreadyInitialized",
    "inputs": []
  },
  {
    "type": "error",
    "name": "ContractNotResponsible",
    "inputs": []
  },
  {
    "type": "error",
    "name": "ERC1967InvalidImplementation",
    "inputs": [
      {
        "name": "implementation",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "ERC1967NonPayable",
    "inputs": []
  },
  {
    "type": "error",
    "name": "EmptyMultiaddr",
    "inputs": []
  },
  {
    "type": "error",
    "name": "ExistingKeyBinding",
    "inputs": []
  },
  {
    "type": "error",
    "name": "FailedCall",
    "inputs": []
  },
  {
    "type": "error",
    "name": "InvalidInitialization",
    "inputs": []
  },
  {
    "type": "error",
    "name": "InvalidKeyBindingFeeAmount",
    "inputs": []
  },
  {
    "type": "error",
    "name": "InvalidSafeAddress",
    "inputs": []
  },
  {
    "type": "error",
    "name": "InvalidTokenRecipient",
    "inputs": []
  },
  {
    "type": "error",
    "name": "InvalidTokenSender",
    "inputs": []
  },
  {
    "type": "error",
    "name": "InvalidTokensReceivedUsage",
    "inputs": []
  },
  {
    "type": "error",
    "name": "KeyIdOutOfRange",
    "inputs": []
  },
  {
    "type": "error",
    "name": "MultiSigUninitialized",
    "inputs": []
  },
  {
    "type": "error",
    "name": "NoNeedToProvideKeyBindingFee",
    "inputs": []
  },
  {
    "type": "error",
    "name": "NotInitializing",
    "inputs": []
  },
  {
    "type": "error",
    "name": "OwnableInvalidOwner",
    "inputs": [
      {
        "name": "owner",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "OwnableUnauthorizedAccount",
    "inputs": [
      {
        "name": "account",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "UUPSUnauthorizedCallContext",
    "inputs": []
  },
  {
    "type": "error",
    "name": "UUPSUnsupportedProxiableUUID",
    "inputs": [
      {
        "name": "slot",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ]
  },
  {
    "type": "error",
    "name": "WrongToken",
    "inputs": []
  },
  {
    "type": "error",
    "name": "ZeroAddress",
    "inputs": [
      {
        "name": "reason",
        "type": "string",
        "internalType": "string"
      }
    ]
  },
  {
    "type": "error",
    "name": "ZeroInterval",
    "inputs": []
  }
]

Structs

AddressAnnouncement

Event with signature AddressAnnouncement(address,string) and selector 0xc4df5ba16814838ab2618829d68f8623bb897302f24dbdba2279dbe45adb3d14.

AddressEmptyCode

Custom error with signature AddressEmptyCode(address) and selector 0x9996b315.

AlreadyInitialized

Custom error with signature AlreadyInitialized() and selector 0x0dc149f0.

ContractNotResponsible

Custom error with signature ContractNotResponsible() and selector 0xacd5a823.

ERC777_HOOK_BIND_KEY_MAYBE_ANNOUNCE_SIZECall

Function with signature ERC777_HOOK_BIND_KEY_MAYBE_ANNOUNCE_SIZE() and selector 0xe9d3ee95.

ERC777_HOOK_BIND_KEY_MAYBE_ANNOUNCE_SIZEReturn

Container type for the return parameters of the ERC777_HOOK_BIND_KEY_MAYBE_ANNOUNCE_SIZE() function.

ERC1967InvalidImplementation

Custom error with signature ERC1967InvalidImplementation(address) and selector 0x4c9c8ce3.

ERC1967NonPayable

Custom error with signature ERC1967NonPayable() and selector 0xb398979f.

EmptyMultiaddr

Custom error with signature EmptyMultiaddr() and selector 0x09ceac70.

ExistingKeyBinding

Custom error with signature ExistingKeyBinding() and selector 0x0b098de0.

FailedCall

Custom error with signature FailedCall() and selector 0xd6bda275.

HoprAnnouncementsInstance

A HoprAnnouncements instance.

Initialized

Event with signature Initialized(uint64) and selector 0xc7f505b2f371ae2175ee4913f4499e1f2633a7b5936321eed1cdaeb6115181d2.

InvalidInitialization

Custom error with signature InvalidInitialization() and selector 0xf92ee8a9.

InvalidKeyBindingFeeAmount

Custom error with signature InvalidKeyBindingFeeAmount() and selector 0x4732cefd.

InvalidSafeAddress

Custom error with signature InvalidSafeAddress() and selector 0x8e9d7c5e.

InvalidTokenRecipient

Custom error with signature InvalidTokenRecipient() and selector 0xb9c49108.

InvalidTokenSender

Custom error with signature InvalidTokenSender() and selector 0x95276c42.

InvalidTokensReceivedUsage

Custom error with signature InvalidTokensReceivedUsage() and selector 0x69ee6f28.

KeyBinding

Event with signature KeyBinding(bytes32,bytes32,bytes32,address,uint256) and selector 0x11d94653de52fc143887956fe4b5e4120a850b1300ece0533dee6bd8ff91f732.

KeyBindingFeeUpdate

Event with signature KeyBindingFeeUpdate(uint256,uint256) and selector 0x6925945cb4c6b597a361809d5b51107e4e03720404a82483853e91faa176b88e.

KeyBindingWithSignatureTimestamp

KeyId

KeyIdOutOfRange

Custom error with signature KeyIdOutOfRange() and selector 0x212a77fe.

LEDGER_VERSIONCall

Function with signature LEDGER_VERSION() and selector 0xddad1902.

LEDGER_VERSIONReturn

Container type for the return parameters of the LEDGER_VERSION() function.

LedgerDomainSeparatorUpdated

Event with signature LedgerDomainSeparatorUpdated(bytes32) and selector 0xa43fad83920fd09445855e854e73c9c532e17402c9ceb09993a2392843a5bdb9.

MultiSigUninitialized

Custom error with signature MultiSigUninitialized() and selector 0x454a20c8.

NoNeedToProvideKeyBindingFee

Custom error with signature NoNeedToProvideKeyBindingFee() and selector 0x0da27ec4.

NotInitializing

Custom error with signature NotInitializing() and selector 0xd7e6bcf8.

OwnableInvalidOwner

Custom error with signature OwnableInvalidOwner(address) and selector 0x1e4fbdf7.

OwnableUnauthorizedAccount

Custom error with signature OwnableUnauthorizedAccount(address) and selector 0x118cdaa7.

OwnershipTransferred

Event with signature OwnershipTransferred(address,address) and selector 0x8be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e0.

RevokeAnnouncement

Event with signature RevokeAnnouncement(address) and selector 0xa4de30a528becadf82649d1395c0e30dd18ae35b5a96ce71e9295bb14bc9f3bc.

SNAPSHOT_INTERVALCall

Function with signature SNAPSHOT_INTERVAL() and selector 0x6d2beef1.

SNAPSHOT_INTERVALReturn

Container type for the return parameters of the SNAPSHOT_INTERVAL() function.

TOKENCall

Function with signature TOKEN() and selector 0x82bfefc8.

TOKENReturn

Container type for the return parameters of the TOKEN() function.

TOKENS_RECIPIENT_INTERFACE_HASHCall

Function with signature TOKENS_RECIPIENT_INTERFACE_HASH() and selector 0x72581cc0.

TOKENS_RECIPIENT_INTERFACE_HASHReturn

Container type for the return parameters of the TOKENS_RECIPIENT_INTERFACE_HASH() function.

UPGRADE_INTERFACE_VERSIONCall

Function with signature UPGRADE_INTERFACE_VERSION() and selector 0xad3cb1cc.

UPGRADE_INTERFACE_VERSIONReturn

Container type for the return parameters of the UPGRADE_INTERFACE_VERSION() function.

UUPSUnauthorizedCallContext

Custom error with signature UUPSUnauthorizedCallContext() and selector 0xe07c8dba.

UUPSUnsupportedProxiableUUID

Custom error with signature UUPSUnsupportedProxiableUUID(bytes32) and selector 0xaa1d49a4.

Upgraded

Event with signature Upgraded(address) and selector 0xbc7cd75a20ee27fd9adebab32041f755214dbc6bffa90cc0225b39da2e5c2d3b.

WrongToken

Custom error with signature WrongToken() and selector 0xa0f3feea.

ZeroAddress

Custom error with signature ZeroAddress(string) and selector 0xeac0d389.

ZeroInterval

Custom error with signature ZeroInterval() and selector 0x346ff607.

announceCall

Function with signature announce(string) and selector 0xea0a5237.

announceReturn

Container type for the return parameters of the announce(string) function.

announceSafeCall

Function with signature announceSafe(address,string) and selector 0xfad0e5a2.

announceSafeReturn

Container type for the return parameters of the announceSafe(address,string) function.

constructorCall

Constructor`.

getAllKeyBindingsCall

Function with signature getAllKeyBindings() and selector 0x10ab5297.

getAllKeyBindingsReturn

Container type for the return parameters of the getAllKeyBindings() function.

getKeyBindingCountCall

Function with signature getKeyBindingCount() and selector 0xf83e4292.

getKeyBindingCountReturn

Container type for the return parameters of the getKeyBindingCount() function.

getKeyBindingWithKeyIdCall

Function with signature getKeyBindingWithKeyId(uint32) and selector 0xa8b4eec9.

getKeyBindingWithKeyIdReturn

Container type for the return parameters of the getKeyBindingWithKeyId(uint32) function.

getKeyIdRangeCall

Function with signature getKeyIdRange() and selector 0x773b4a33.

getKeyIdRangeReturn

Container type for the return parameters of the getKeyIdRange() function.

getKeyIdWithOffchainKeyCall

Function with signature getKeyIdWithOffchainKey(bytes32) and selector 0x604634c9.

getKeyIdWithOffchainKeyReturn

Container type for the return parameters of the getKeyIdWithOffchainKey(bytes32) function.

getOffchainKeyWithKeyIdCall

Function with signature getOffchainKeyWithKeyId(uint32) and selector 0x4ac3e4f2.

getOffchainKeyWithKeyIdReturn

Container type for the return parameters of the getOffchainKeyWithKeyId(uint32) function.

initializeCall

Function with signature initialize(bytes) and selector 0x439fab91.

initializeReturn

Container type for the return parameters of the initialize(bytes) function.

isOffchainKeyBoundCall

Function with signature isOffchainKeyBound(bytes32) and selector 0xa969635a.

isOffchainKeyBoundReturn

Container type for the return parameters of the isOffchainKeyBound(bytes32) function.

keyBindingFeeCall

Function with signature keyBindingFee() and selector 0xfd153acf.

keyBindingFeeReturn

Container type for the return parameters of the keyBindingFee() function.

latestRootCall

Function with signature latestRoot() and selector 0xd7b0fef1.

latestRootReturn

Container type for the return parameters of the latestRoot() function.

latestSnapshotRootCall

Function with signature latestSnapshotRoot() and selector 0x0df18f94.

latestSnapshotRootReturn

Container type for the return parameters of the latestSnapshotRoot() function.

ledgerDomainSeparatorCall

Function with signature ledgerDomainSeparator() and selector 0xc966c4fe.

ledgerDomainSeparatorReturn

Container type for the return parameters of the ledgerDomainSeparator() function.

multiaddrOfCall

Function with signature multiaddrOf(address) and selector 0x53665aaa.

multiaddrOfReturn

Container type for the return parameters of the multiaddrOf(address) function.

ownerCall

Function with signature owner() and selector 0x8da5cb5b.

ownerReturn

Container type for the return parameters of the owner() function.

proxiableUUIDCall

Function with signature proxiableUUID() and selector 0x52d1902d.

proxiableUUIDReturn

Container type for the return parameters of the proxiableUUID() function.

renounceOwnershipCall

Function with signature renounceOwnership() and selector 0x715018a6.

renounceOwnershipReturn

Container type for the return parameters of the renounceOwnership() function.

revokeCall

Function with signature revoke() and selector 0xb6549f75.

revokeReturn

Container type for the return parameters of the revoke() function.

revokeSafeCall

Function with signature revokeSafe(address) and selector 0x308c712e.

revokeSafeReturn

Container type for the return parameters of the revokeSafe(address) function.

tokensReceivedCall

Function with signature tokensReceived(address,address,address,uint256,bytes,bytes) and selector 0x0023de29.

tokensReceivedReturn

Container type for the return parameters of the tokensReceived(address,address,address,uint256,bytes,bytes) function.

transferOwnershipCall

Function with signature transferOwnership(address) and selector 0xf2fde38b.

transferOwnershipReturn

Container type for the return parameters of the transferOwnership(address) function.

tryGetKeyBindingCall

Function with signature tryGetKeyBinding(bytes32) and selector 0xa2a07756.

tryGetKeyBindingReturn

Container type for the return parameters of the tryGetKeyBinding(bytes32) function.

updateKeyBindingFeeCall

Function with signature updateKeyBindingFee(uint256) and selector 0x244d496e.

updateKeyBindingFeeReturn

Container type for the return parameters of the updateKeyBindingFee(uint256) function.

updateLedgerDomainSeparatorCall

Function with signature updateLedgerDomainSeparator() and selector 0xdc96fd50.

updateLedgerDomainSeparatorReturn

Container type for the return parameters of the updateLedgerDomainSeparator() function.

upgradeToAndCallCall

Function with signature upgradeToAndCall(address,bytes) and selector 0x4f1ef286.

upgradeToAndCallReturn

Container type for the return parameters of the upgradeToAndCall(address,bytes) function.

Enums

HoprAnnouncementsCalls

Container for all the HoprAnnouncements function calls.

HoprAnnouncementsErrors

Container for all the HoprAnnouncements custom errors.

HoprAnnouncementsEvents

Container for all the HoprAnnouncements events.

Statics

BYTECODE

The creation / init bytecode of the contract.

DEPLOYED_BYTECODE

The runtime bytecode of the contract, as deployed on the network.

Functions

deploy

Deploys this contract using the given provider and constructor arguments, if any.

deploy_builder

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

new

Creates a new wrapper around an on-chain HoprAnnouncements contract instance.