hopr_bindings::hopr_announcements_proxy

Module HoprAnnouncementsProxy

Generated by the following Solidity interface…

interface HoprAnnouncementsProxy {
    error AddressEmptyCode(address target);
    error ERC1967InvalidImplementation(address implementation);
    error ERC1967NonPayable();
    error FailedCall();

    event Upgraded(address indexed implementation);

    constructor(address _implementation, bytes data);

    fallback() external payable;

    function implementation() external view returns (address);
}

…which was generated by the following JSON ABI:

[
  {
    "type": "constructor",
    "inputs": [
      {
        "name": "_implementation",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "data",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "fallback",
    "stateMutability": "payable"
  },
  {
    "type": "function",
    "name": "implementation",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "event",
    "name": "Upgraded",
    "inputs": [
      {
        "name": "implementation",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "error",
    "name": "AddressEmptyCode",
    "inputs": [
      {
        "name": "target",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "ERC1967InvalidImplementation",
    "inputs": [
      {
        "name": "implementation",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "ERC1967NonPayable",
    "inputs": []
  },
  {
    "type": "error",
    "name": "FailedCall",
    "inputs": []
  }
]

Structs

AddressEmptyCode

Custom error with signature AddressEmptyCode(address) and selector 0x9996b315.

ERC1967InvalidImplementation

Custom error with signature ERC1967InvalidImplementation(address) and selector 0x4c9c8ce3.

ERC1967NonPayable

Custom error with signature ERC1967NonPayable() and selector 0xb398979f.

FailedCall

Custom error with signature FailedCall() and selector 0xd6bda275.

HoprAnnouncementsProxyInstance

A HoprAnnouncementsProxy instance.

Upgraded

Event with signature Upgraded(address) and selector 0xbc7cd75a20ee27fd9adebab32041f755214dbc6bffa90cc0225b39da2e5c2d3b.

constructorCall

Constructor`.

implementationCall

Function with signature implementation() and selector 0x5c60da1b.

implementationReturn

Container type for the return parameters of the implementation() function.

Enums

HoprAnnouncementsProxyCalls

Container for all the HoprAnnouncementsProxy function calls.

HoprAnnouncementsProxyErrors

Container for all the HoprAnnouncementsProxy custom errors.

HoprAnnouncementsProxyEvents

Container for all the HoprAnnouncementsProxy events.

Statics

BYTECODE

The creation / init bytecode of the contract.

DEPLOYED_BYTECODE

The runtime bytecode of the contract, as deployed on the network.

Functions

deploy

Deploys this contract using the given provider and constructor arguments, if any.

deploy_builder

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

new

Creates a new wrapper around an on-chain HoprAnnouncementsProxy contract instance.