hopr_bindings::hopr_boost::HoprBoost

Enum HoprBoostCalls

pub enum HoprBoostCalls {

    DEFAULT_ADMIN_ROLE(DEFAULT_ADMIN_ROLECall),
    MINTER_ROLE(MINTER_ROLECall),
    approve(approveCall),
    balanceOf(balanceOfCall),
    batchMint(batchMintCall),
    boostOf(boostOfCall),
    getApproved(getApprovedCall),
    getRoleAdmin(getRoleAdminCall),
    getRoleMember(getRoleMemberCall),
    getRoleMemberCount(getRoleMemberCountCall),
    grantRole(grantRoleCall),
    hasRole(hasRoleCall),
    isApprovedForAll(isApprovedForAllCall),
    mint(mintCall),
    name(nameCall),
    ownerOf(ownerOfCall),
    reclaimErc20Tokens(reclaimErc20TokensCall),
    reclaimErc721Tokens(reclaimErc721TokensCall),
    renounceRole(renounceRoleCall),
    revokeRole(revokeRoleCall),
    safeTransferFrom_0(safeTransferFrom_0Call),
    safeTransferFrom_1(safeTransferFrom_1Call),
    setApprovalForAll(setApprovalForAllCall),
    supportsInterface(supportsInterfaceCall),
    symbol(symbolCall),
    tokenByIndex(tokenByIndexCall),
    tokenOfOwnerByIndex(tokenOfOwnerByIndexCall),
    tokenURI(tokenURICall),
    totalSupply(totalSupplyCall),
    transferFrom(transferFromCall),
    typeAt(typeAtCall),
    typeIndexOf(typeIndexOfCall),
    typeOf(typeOfCall),
    updateBaseURI(updateBaseURICall),

}

Container for all the HoprBoost function calls.

Variants

DEFAULT_ADMIN_ROLE(DEFAULT_ADMIN_ROLECall)

MINTER_ROLE(MINTER_ROLECall)

approve(approveCall)

balanceOf(balanceOfCall)

batchMint(batchMintCall)

boostOf(boostOfCall)

getApproved(getApprovedCall)

getRoleAdmin(getRoleAdminCall)

getRoleMember(getRoleMemberCall)

getRoleMemberCount(getRoleMemberCountCall)

grantRole(grantRoleCall)

hasRole(hasRoleCall)

isApprovedForAll(isApprovedForAllCall)

mint(mintCall)

name(nameCall)

ownerOf(ownerOfCall)

reclaimErc20Tokens(reclaimErc20TokensCall)

reclaimErc721Tokens(reclaimErc721TokensCall)

renounceRole(renounceRoleCall)

revokeRole(revokeRoleCall)

safeTransferFrom_0(safeTransferFrom_0Call)

safeTransferFrom_1(safeTransferFrom_1Call)

setApprovalForAll(setApprovalForAllCall)

supportsInterface(supportsInterfaceCall)

symbol(symbolCall)

tokenByIndex(tokenByIndexCall)

tokenOfOwnerByIndex(tokenOfOwnerByIndexCall)

tokenURI(tokenURICall)

totalSupply(totalSupplyCall)

transferFrom(transferFromCall)

typeAt(typeAtCall)

typeIndexOf(typeIndexOfCall)

typeOf(typeOfCall)

updateBaseURI(updateBaseURICall)

Implementations

impl HoprBoostCalls

pub const SELECTORS: &'static [[u8; 4]]

All the selectors of this enum.

Note that the selectors might not be in the same order as the variants. No guarantees are made about the order of the selectors.

Prefer using SolInterface methods instead.

pub const VARIANT_NAMES: &'static [&'static str]

The names of the variants in the same order as SELECTORS.

pub const SIGNATURES: &'static [&'static str]

The signatures in the same order as SELECTORS.

pub fn signature_by_selector(selector: [u8; 4]) -> Option<&'static str>

Returns the signature for the given selector, if known.

pub fn name_by_selector(selector: [u8; 4]) -> Option<&'static str>

Returns the enum variant name for the given selector, if known.