hopr_bindings

Module hopr_capability_permissions

Modules

hopr_capability_permissions

This module was auto-generated with ethers-rs Abigen. More information at: https://github.com/gakonst/ethers-rs

Structs

AddressIsZero

Custom Error type AddressIsZero with signature AddressIsZero() and selector 0x867915ab

ApproveSelectorCall

Container type for all input parameters for the APPROVE_SELECTOR function with signature APPROVE_SELECTOR() and selector 0xa0865839

ApproveSelectorReturn

Container type for all return fields from the APPROVE_SELECTOR function with signature APPROVE_SELECTOR() and selector 0xa0865839

ArrayTooLong

Custom Error type ArrayTooLong with signature ArrayTooLong() and selector 0xbd26cc38

ArraysDifferentLength

Custom Error type ArraysDifferentLength with signature ArraysDifferentLength() and selector 0x74f4d537

CalldataOutOfBounds

Custom Error type CalldataOutOfBounds with signature CalldataOutOfBounds() and selector 0x742638b4

CloseIncomingChannelSelectorCall

Container type for all input parameters for the CLOSE_INCOMING_CHANNEL_SELECTOR function with signature CLOSE_INCOMING_CHANNEL_SELECTOR() and selector 0x678a5efb

CloseIncomingChannelSelectorReturn

Container type for all return fields from the CLOSE_INCOMING_CHANNEL_SELECTOR function with signature CLOSE_INCOMING_CHANNEL_SELECTOR() and selector 0x678a5efb

DefaultPermissionRejected

Custom Error type DefaultPermissionRejected with signature DefaultPermissionRejected() and selector 0x58723037

DelegateCallNotAllowed

Custom Error type DelegateCallNotAllowed with signature DelegateCallNotAllowed() and selector 0x0d89438e

FinalizeOutgoingChannelClosureSelectorCall

Container type for all input parameters for the FINALIZE_OUTGOING_CHANNEL_CLOSURE_SELECTOR function with signature FINALIZE_OUTGOING_CHANNEL_CLOSURE_SELECTOR() and selector 0x8167b67e

FinalizeOutgoingChannelClosureSelectorReturn

Container type for all return fields from the FINALIZE_OUTGOING_CHANNEL_CLOSURE_SELECTOR function with signature FINALIZE_OUTGOING_CHANNEL_CLOSURE_SELECTOR() and selector 0x8167b67e

FunctionSignatureTooShort

Custom Error type FunctionSignatureTooShort with signature FunctionSignatureTooShort() and selector 0x4684c122

FundChannelSelectorCall

Container type for all input parameters for the FUND_CHANNEL_SELECTOR function with signature FUND_CHANNEL_SELECTOR() and selector 0xb677a40f

FundChannelSelectorReturn

Container type for all return fields from the FUND_CHANNEL_SELECTOR function with signature FUND_CHANNEL_SELECTOR() and selector 0xb677a40f

GranularPermissionRejected

Custom Error type GranularPermissionRejected with signature GranularPermissionRejected() and selector 0x864dd1e7

HoprCapabilityPermissions

InitiateOutgoingChannelClosureSelectorCall

Container type for all input parameters for the INITIATE_OUTGOING_CHANNEL_CLOSURE_SELECTOR function with signature INITIATE_OUTGOING_CHANNEL_CLOSURE_SELECTOR() and selector 0x5daa9e24

InitiateOutgoingChannelClosureSelectorReturn

Container type for all return fields from the INITIATE_OUTGOING_CHANNEL_CLOSURE_SELECTOR function with signature INITIATE_OUTGOING_CHANNEL_CLOSURE_SELECTOR() and selector 0x5daa9e24

NoMembership

Custom Error type NoMembership with signature NoMembership() and selector 0xfd8e9f28

NodePermissionRejected

Custom Error type NodePermissionRejected with signature NodePermissionRejected() and selector 0x6eb0315f

ParameterNotAllowed

Custom Error type ParameterNotAllowed with signature ParameterNotAllowed() and selector 0x31e98246

PermissionNotConfigured

Custom Error type PermissionNotConfigured with signature PermissionNotConfigured() and selector 0x46ad4588

RedeemTicketSelectorCall

Container type for all input parameters for the REDEEM_TICKET_SELECTOR function with signature REDEEM_TICKET_SELECTOR() and selector 0x51d2614b

RedeemTicketSelectorReturn

Container type for all return fields from the REDEEM_TICKET_SELECTOR function with signature REDEEM_TICKET_SELECTOR() and selector 0x51d2614b

RevokedTargetFilter

ScopedGranularChannelCapabilityFilter

ScopedGranularSendCapabilityFilter

ScopedGranularTokenCapabilityFilter

ScopedTargetChannelsFilter

ScopedTargetSendFilter

ScopedTargetTokenFilter

SendNotAllowed

Custom Error type SendNotAllowed with signature SendNotAllowed() and selector 0x09e9cd49

SendSelectorCall

Container type for all input parameters for the SEND_SELECTOR function with signature SEND_SELECTOR() and selector 0x7a757115

SendSelectorReturn

Container type for all return fields from the SEND_SELECTOR function with signature SEND_SELECTOR() and selector 0x7a757115

TargetAddressNotAllowed

Custom Error type TargetAddressNotAllowed with signature TargetAddressNotAllowed() and selector 0xef3440ac

TargetIsNotScoped

Custom Error type TargetIsNotScoped with signature TargetIsNotScoped() and selector 0x4a890321

TargetIsScoped

Custom Error type TargetIsScoped with signature TargetIsScoped() and selector 0xe8c07d2a

UnacceptableMultiSendOffset

Custom Error type UnacceptableMultiSendOffset with signature UnacceptableMultiSendOffset() and selector 0x7ed11137

Enums

HoprCapabilityPermissionsCalls

Container type for all of the contract’s call

HoprCapabilityPermissionsErrors

Container type for all of the contract’s custom errors

HoprCapabilityPermissionsEvents

Container type for all of the contract’s events

Statics

HOPRCAPABILITYPERMISSIONS_ABI

The parsed JSON ABI of the contract.

HOPRCAPABILITYPERMISSIONS_BYTECODE

The bytecode of the contract.

HOPRCAPABILITYPERMISSIONS_DEPLOYED_BYTECODE

The deployed bytecode of the contract.