hopr_bindings::hopr_channels

Module HoprChannels

Generated by the following Solidity interface…

library HoprChannelsType {
    type ChannelStatus is uint8;
    type Balance is uint96;
    type ChannelEpoch is uint24;
    type TicketIndex is uint48;
    type Timestamp is uint32;
    type WinProb is uint56;
}

library HoprCrypto {
    struct CompactSignature {
        bytes32 r;
        bytes32 vs;
    }
    struct VRFParameters {
        uint256 vx;
        uint256 vy;
        uint256 s;
        uint256 h;
        uint256 sBx;
        uint256 sBy;
        uint256 hVx;
        uint256 hVy;
    }
}

library HoprLedger {
    struct RootStruct {
        bytes28 rootHash;
        uint32 timestamp;
    }
}

interface HoprChannels {
    struct RedeemableTicket {
        TicketData data;
        HoprCrypto.CompactSignature signature;
        uint256 porSecret;
    }
    struct TicketData {
        bytes32 channelId;
        HoprChannelsType.Balance amount;
        HoprChannelsType.TicketIndex ticketIndex;
        HoprChannelsType.ChannelEpoch epoch;
        HoprChannelsType.WinProb winProb;
    }

    error AddressEmptyCode(address target);
    error AlreadyInitialized();
    error BalanceExceedsGlobalPerChannelAllowance();
    error ContractNotResponsible();
    error ECDSAInvalidSignature();
    error ECDSAInvalidSignatureLength(uint256 length);
    error ECDSAInvalidSignatureS(bytes32 s);
    error FailedCall();
    error InsufficientChannelBalance();
    error InvalidBalance();
    error InvalidCurvePoint();
    error InvalidFieldElement();
    error InvalidNoticePeriod();
    error InvalidPointWitness();
    error InvalidSafeAddress();
    error InvalidTicketIndex();
    error InvalidTicketSignature();
    error InvalidTokenRecipient();
    error InvalidTokensReceivedUsage();
    error InvalidVRFProof();
    error MultiSigUninitialized();
    error NoticePeriodNotDue();
    error SourceEqualsDestination();
    error TicketIsNotAWin();
    error TokenTransferFailed();
    error WrongChannelState(string reason);
    error WrongToken();
    error ZeroAddress(string reason);
    error ZeroInterval();

    event ChannelBalanceDecreased(bytes32 indexed channelId, bytes32 channel);
    event ChannelBalanceIncreased(bytes32 indexed channelId, bytes32 channel);
    event ChannelClosed(bytes32 indexed channelId, bytes32 channel);
    event ChannelOpened(bytes32 indexed channelId, address indexed source, address indexed destination, bytes32 channel);
    event DomainSeparatorUpdated(bytes32 indexed domainSeparator);
    event LedgerDomainSeparatorUpdated(bytes32 indexed ledgerDomainSeparator);
    event OutgoingChannelClosureInitiated(bytes32 indexed channelId, bytes32 channel);
    event TicketRedeemed(bytes32 indexed channelId, bytes32 channel);

    constructor(address _token, HoprChannelsType.Timestamp _noticePeriodChannelClosure, address _safeRegistry);

    function ERC777_HOOK_FUND_CHANNEL_MULTI_SIZE() external view returns (uint256);
    function ERC777_HOOK_FUND_CHANNEL_SIZE() external view returns (uint256);
    function LEDGER_VERSION() external view returns (string memory);
    function MAX_USED_BALANCE() external view returns (HoprChannelsType.Balance);
    function MIN_USED_BALANCE() external view returns (HoprChannelsType.Balance);
    function NOTICE_PERIOD_CHANNEL_CLOSURE() external view returns (HoprChannelsType.Timestamp);
    function SNAPSHOT_INTERVAL() external view returns (uint256);
    function TOKEN() external view returns (address);
    function TOKENS_RECIPIENT_INTERFACE_HASH() external view returns (bytes32);
    function VERSION() external view returns (string memory);
    function _currentBlockTimestamp() external view returns (HoprChannelsType.Timestamp);
    function _getChannelId(address source, address destination) external pure returns (bytes32);
    function _getTicketHash(RedeemableTicket memory redeemable) external view returns (bytes32);
    function _isWinningTicket(bytes32 ticketHash, RedeemableTicket memory redeemable, HoprCrypto.VRFParameters memory params) external pure returns (bool);
    function channelState(bytes32 channelId) external view returns (bytes32 state);
    function channels(bytes32) external view returns (HoprChannelsType.Balance balance, HoprChannelsType.TicketIndex ticketIndex, HoprChannelsType.Timestamp closureTime, HoprChannelsType.ChannelEpoch epoch, HoprChannelsType.ChannelStatus status);
    function closeIncomingChannel(address source) external;
    function closeIncomingChannelSafe(address selfAddress, address source) external;
    function domainSeparator() external view returns (bytes32);
    function finalizeOutgoingChannelClosure(address destination) external;
    function finalizeOutgoingChannelClosureSafe(address selfAddress, address destination) external;
    function fundChannel(address account, HoprChannelsType.Balance amount) external;
    function fundChannelSafe(address selfAddress, address account, HoprChannelsType.Balance amount) external;
    function initiateOutgoingChannelClosure(address destination) external;
    function initiateOutgoingChannelClosureSafe(address selfAddress, address destination) external;
    function latestRoot() external view returns (HoprLedger.RootStruct memory);
    function latestSnapshotRoot() external view returns (HoprLedger.RootStruct memory);
    function ledgerDomainSeparator() external view returns (bytes32);
    function multicall(bytes[] memory data) external returns (bytes[] memory results);
    function redeemTicket(RedeemableTicket memory redeemable, HoprCrypto.VRFParameters memory params) external;
    function redeemTicketSafe(address selfAddress, RedeemableTicket memory redeemable, HoprCrypto.VRFParameters memory params) external;
    function tokensReceived(address, address from, address to, uint256 amount, bytes memory userData, bytes memory) external;
    function updateDomainSeparator() external;
    function updateLedgerDomainSeparator() external;
}

…which was generated by the following JSON ABI:

[
  {
    "type": "constructor",
    "inputs": [
      {
        "name": "_token",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "_noticePeriodChannelClosure",
        "type": "uint32",
        "internalType": "HoprChannelsType.Timestamp"
      },
      {
        "name": "_safeRegistry",
        "type": "address",
        "internalType": "contract HoprNodeSafeRegistry"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "ERC777_HOOK_FUND_CHANNEL_MULTI_SIZE",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "ERC777_HOOK_FUND_CHANNEL_SIZE",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "LEDGER_VERSION",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "string",
        "internalType": "string"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "MAX_USED_BALANCE",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint96",
        "internalType": "HoprChannelsType.Balance"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "MIN_USED_BALANCE",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint96",
        "internalType": "HoprChannelsType.Balance"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "NOTICE_PERIOD_CHANNEL_CLOSURE",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint32",
        "internalType": "HoprChannelsType.Timestamp"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "SNAPSHOT_INTERVAL",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "TOKEN",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "contract IERC20"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "TOKENS_RECIPIENT_INTERFACE_HASH",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "VERSION",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "string",
        "internalType": "string"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "_currentBlockTimestamp",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint32",
        "internalType": "HoprChannelsType.Timestamp"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "_getChannelId",
    "inputs": [
      {
        "name": "source",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "destination",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "pure"
  },
  {
    "type": "function",
    "name": "_getTicketHash",
    "inputs": [
      {
        "name": "redeemable",
        "type": "tuple",
        "internalType": "struct HoprChannels.RedeemableTicket",
        "components": [
          {
            "name": "data",
            "type": "tuple",
            "internalType": "struct HoprChannels.TicketData",
            "components": [
              {
                "name": "channelId",
                "type": "bytes32",
                "internalType": "bytes32"
              },
              {
                "name": "amount",
                "type": "uint96",
                "internalType": "HoprChannelsType.Balance"
              },
              {
                "name": "ticketIndex",
                "type": "uint48",
                "internalType": "HoprChannelsType.TicketIndex"
              },
              {
                "name": "epoch",
                "type": "uint24",
                "internalType": "HoprChannelsType.ChannelEpoch"
              },
              {
                "name": "winProb",
                "type": "uint56",
                "internalType": "HoprChannelsType.WinProb"
              }
            ]
          },
          {
            "name": "signature",
            "type": "tuple",
            "internalType": "struct HoprCrypto.CompactSignature",
            "components": [
              {
                "name": "r",
                "type": "bytes32",
                "internalType": "bytes32"
              },
              {
                "name": "vs",
                "type": "bytes32",
                "internalType": "bytes32"
              }
            ]
          },
          {
            "name": "porSecret",
            "type": "uint256",
            "internalType": "uint256"
          }
        ]
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "_isWinningTicket",
    "inputs": [
      {
        "name": "ticketHash",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "redeemable",
        "type": "tuple",
        "internalType": "struct HoprChannels.RedeemableTicket",
        "components": [
          {
            "name": "data",
            "type": "tuple",
            "internalType": "struct HoprChannels.TicketData",
            "components": [
              {
                "name": "channelId",
                "type": "bytes32",
                "internalType": "bytes32"
              },
              {
                "name": "amount",
                "type": "uint96",
                "internalType": "HoprChannelsType.Balance"
              },
              {
                "name": "ticketIndex",
                "type": "uint48",
                "internalType": "HoprChannelsType.TicketIndex"
              },
              {
                "name": "epoch",
                "type": "uint24",
                "internalType": "HoprChannelsType.ChannelEpoch"
              },
              {
                "name": "winProb",
                "type": "uint56",
                "internalType": "HoprChannelsType.WinProb"
              }
            ]
          },
          {
            "name": "signature",
            "type": "tuple",
            "internalType": "struct HoprCrypto.CompactSignature",
            "components": [
              {
                "name": "r",
                "type": "bytes32",
                "internalType": "bytes32"
              },
              {
                "name": "vs",
                "type": "bytes32",
                "internalType": "bytes32"
              }
            ]
          },
          {
            "name": "porSecret",
            "type": "uint256",
            "internalType": "uint256"
          }
        ]
      },
      {
        "name": "params",
        "type": "tuple",
        "internalType": "struct HoprCrypto.VRFParameters",
        "components": [
          {
            "name": "vx",
            "type": "uint256",
            "internalType": "uint256"
          },
          {
            "name": "vy",
            "type": "uint256",
            "internalType": "uint256"
          },
          {
            "name": "s",
            "type": "uint256",
            "internalType": "uint256"
          },
          {
            "name": "h",
            "type": "uint256",
            "internalType": "uint256"
          },
          {
            "name": "sBx",
            "type": "uint256",
            "internalType": "uint256"
          },
          {
            "name": "sBy",
            "type": "uint256",
            "internalType": "uint256"
          },
          {
            "name": "hVx",
            "type": "uint256",
            "internalType": "uint256"
          },
          {
            "name": "hVy",
            "type": "uint256",
            "internalType": "uint256"
          }
        ]
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "pure"
  },
  {
    "type": "function",
    "name": "channelState",
    "inputs": [
      {
        "name": "channelId",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "outputs": [
      {
        "name": "state",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "channels",
    "inputs": [
      {
        "name": "",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "outputs": [
      {
        "name": "balance",
        "type": "uint96",
        "internalType": "HoprChannelsType.Balance"
      },
      {
        "name": "ticketIndex",
        "type": "uint48",
        "internalType": "HoprChannelsType.TicketIndex"
      },
      {
        "name": "closureTime",
        "type": "uint32",
        "internalType": "HoprChannelsType.Timestamp"
      },
      {
        "name": "epoch",
        "type": "uint24",
        "internalType": "HoprChannelsType.ChannelEpoch"
      },
      {
        "name": "status",
        "type": "uint8",
        "internalType": "enum HoprChannelsType.ChannelStatus"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "closeIncomingChannel",
    "inputs": [
      {
        "name": "source",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "closeIncomingChannelSafe",
    "inputs": [
      {
        "name": "selfAddress",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "source",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "domainSeparator",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "finalizeOutgoingChannelClosure",
    "inputs": [
      {
        "name": "destination",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "finalizeOutgoingChannelClosureSafe",
    "inputs": [
      {
        "name": "selfAddress",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "destination",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "fundChannel",
    "inputs": [
      {
        "name": "account",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "amount",
        "type": "uint96",
        "internalType": "HoprChannelsType.Balance"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "fundChannelSafe",
    "inputs": [
      {
        "name": "selfAddress",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "account",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "amount",
        "type": "uint96",
        "internalType": "HoprChannelsType.Balance"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "initiateOutgoingChannelClosure",
    "inputs": [
      {
        "name": "destination",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "initiateOutgoingChannelClosureSafe",
    "inputs": [
      {
        "name": "selfAddress",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "destination",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "latestRoot",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "tuple",
        "internalType": "struct HoprLedger.RootStruct",
        "components": [
          {
            "name": "rootHash",
            "type": "bytes28",
            "internalType": "bytes28"
          },
          {
            "name": "timestamp",
            "type": "uint32",
            "internalType": "uint32"
          }
        ]
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "latestSnapshotRoot",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "tuple",
        "internalType": "struct HoprLedger.RootStruct",
        "components": [
          {
            "name": "rootHash",
            "type": "bytes28",
            "internalType": "bytes28"
          },
          {
            "name": "timestamp",
            "type": "uint32",
            "internalType": "uint32"
          }
        ]
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "ledgerDomainSeparator",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "multicall",
    "inputs": [
      {
        "name": "data",
        "type": "bytes[]",
        "internalType": "bytes[]"
      }
    ],
    "outputs": [
      {
        "name": "results",
        "type": "bytes[]",
        "internalType": "bytes[]"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "redeemTicket",
    "inputs": [
      {
        "name": "redeemable",
        "type": "tuple",
        "internalType": "struct HoprChannels.RedeemableTicket",
        "components": [
          {
            "name": "data",
            "type": "tuple",
            "internalType": "struct HoprChannels.TicketData",
            "components": [
              {
                "name": "channelId",
                "type": "bytes32",
                "internalType": "bytes32"
              },
              {
                "name": "amount",
                "type": "uint96",
                "internalType": "HoprChannelsType.Balance"
              },
              {
                "name": "ticketIndex",
                "type": "uint48",
                "internalType": "HoprChannelsType.TicketIndex"
              },
              {
                "name": "epoch",
                "type": "uint24",
                "internalType": "HoprChannelsType.ChannelEpoch"
              },
              {
                "name": "winProb",
                "type": "uint56",
                "internalType": "HoprChannelsType.WinProb"
              }
            ]
          },
          {
            "name": "signature",
            "type": "tuple",
            "internalType": "struct HoprCrypto.CompactSignature",
            "components": [
              {
                "name": "r",
                "type": "bytes32",
                "internalType": "bytes32"
              },
              {
                "name": "vs",
                "type": "bytes32",
                "internalType": "bytes32"
              }
            ]
          },
          {
            "name": "porSecret",
            "type": "uint256",
            "internalType": "uint256"
          }
        ]
      },
      {
        "name": "params",
        "type": "tuple",
        "internalType": "struct HoprCrypto.VRFParameters",
        "components": [
          {
            "name": "vx",
            "type": "uint256",
            "internalType": "uint256"
          },
          {
            "name": "vy",
            "type": "uint256",
            "internalType": "uint256"
          },
          {
            "name": "s",
            "type": "uint256",
            "internalType": "uint256"
          },
          {
            "name": "h",
            "type": "uint256",
            "internalType": "uint256"
          },
          {
            "name": "sBx",
            "type": "uint256",
            "internalType": "uint256"
          },
          {
            "name": "sBy",
            "type": "uint256",
            "internalType": "uint256"
          },
          {
            "name": "hVx",
            "type": "uint256",
            "internalType": "uint256"
          },
          {
            "name": "hVy",
            "type": "uint256",
            "internalType": "uint256"
          }
        ]
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "redeemTicketSafe",
    "inputs": [
      {
        "name": "selfAddress",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "redeemable",
        "type": "tuple",
        "internalType": "struct HoprChannels.RedeemableTicket",
        "components": [
          {
            "name": "data",
            "type": "tuple",
            "internalType": "struct HoprChannels.TicketData",
            "components": [
              {
                "name": "channelId",
                "type": "bytes32",
                "internalType": "bytes32"
              },
              {
                "name": "amount",
                "type": "uint96",
                "internalType": "HoprChannelsType.Balance"
              },
              {
                "name": "ticketIndex",
                "type": "uint48",
                "internalType": "HoprChannelsType.TicketIndex"
              },
              {
                "name": "epoch",
                "type": "uint24",
                "internalType": "HoprChannelsType.ChannelEpoch"
              },
              {
                "name": "winProb",
                "type": "uint56",
                "internalType": "HoprChannelsType.WinProb"
              }
            ]
          },
          {
            "name": "signature",
            "type": "tuple",
            "internalType": "struct HoprCrypto.CompactSignature",
            "components": [
              {
                "name": "r",
                "type": "bytes32",
                "internalType": "bytes32"
              },
              {
                "name": "vs",
                "type": "bytes32",
                "internalType": "bytes32"
              }
            ]
          },
          {
            "name": "porSecret",
            "type": "uint256",
            "internalType": "uint256"
          }
        ]
      },
      {
        "name": "params",
        "type": "tuple",
        "internalType": "struct HoprCrypto.VRFParameters",
        "components": [
          {
            "name": "vx",
            "type": "uint256",
            "internalType": "uint256"
          },
          {
            "name": "vy",
            "type": "uint256",
            "internalType": "uint256"
          },
          {
            "name": "s",
            "type": "uint256",
            "internalType": "uint256"
          },
          {
            "name": "h",
            "type": "uint256",
            "internalType": "uint256"
          },
          {
            "name": "sBx",
            "type": "uint256",
            "internalType": "uint256"
          },
          {
            "name": "sBy",
            "type": "uint256",
            "internalType": "uint256"
          },
          {
            "name": "hVx",
            "type": "uint256",
            "internalType": "uint256"
          },
          {
            "name": "hVy",
            "type": "uint256",
            "internalType": "uint256"
          }
        ]
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "tokensReceived",
    "inputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "from",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "to",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "amount",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "userData",
        "type": "bytes",
        "internalType": "bytes"
      },
      {
        "name": "",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "updateDomainSeparator",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "updateLedgerDomainSeparator",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "event",
    "name": "ChannelBalanceDecreased",
    "inputs": [
      {
        "name": "channelId",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      },
      {
        "name": "channel",
        "type": "bytes32",
        "indexed": false,
        "internalType": "bytes32"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "ChannelBalanceIncreased",
    "inputs": [
      {
        "name": "channelId",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      },
      {
        "name": "channel",
        "type": "bytes32",
        "indexed": false,
        "internalType": "bytes32"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "ChannelClosed",
    "inputs": [
      {
        "name": "channelId",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      },
      {
        "name": "channel",
        "type": "bytes32",
        "indexed": false,
        "internalType": "bytes32"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "ChannelOpened",
    "inputs": [
      {
        "name": "channelId",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      },
      {
        "name": "source",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "destination",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "channel",
        "type": "bytes32",
        "indexed": false,
        "internalType": "bytes32"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "DomainSeparatorUpdated",
    "inputs": [
      {
        "name": "domainSeparator",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "LedgerDomainSeparatorUpdated",
    "inputs": [
      {
        "name": "ledgerDomainSeparator",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "OutgoingChannelClosureInitiated",
    "inputs": [
      {
        "name": "channelId",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      },
      {
        "name": "channel",
        "type": "bytes32",
        "indexed": false,
        "internalType": "bytes32"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "TicketRedeemed",
    "inputs": [
      {
        "name": "channelId",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      },
      {
        "name": "channel",
        "type": "bytes32",
        "indexed": false,
        "internalType": "bytes32"
      }
    ],
    "anonymous": false
  },
  {
    "type": "error",
    "name": "AddressEmptyCode",
    "inputs": [
      {
        "name": "target",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "AlreadyInitialized",
    "inputs": []
  },
  {
    "type": "error",
    "name": "BalanceExceedsGlobalPerChannelAllowance",
    "inputs": []
  },
  {
    "type": "error",
    "name": "ContractNotResponsible",
    "inputs": []
  },
  {
    "type": "error",
    "name": "ECDSAInvalidSignature",
    "inputs": []
  },
  {
    "type": "error",
    "name": "ECDSAInvalidSignatureLength",
    "inputs": [
      {
        "name": "length",
        "type": "uint256",
        "internalType": "uint256"
      }
    ]
  },
  {
    "type": "error",
    "name": "ECDSAInvalidSignatureS",
    "inputs": [
      {
        "name": "s",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ]
  },
  {
    "type": "error",
    "name": "FailedCall",
    "inputs": []
  },
  {
    "type": "error",
    "name": "InsufficientChannelBalance",
    "inputs": []
  },
  {
    "type": "error",
    "name": "InvalidBalance",
    "inputs": []
  },
  {
    "type": "error",
    "name": "InvalidCurvePoint",
    "inputs": []
  },
  {
    "type": "error",
    "name": "InvalidFieldElement",
    "inputs": []
  },
  {
    "type": "error",
    "name": "InvalidNoticePeriod",
    "inputs": []
  },
  {
    "type": "error",
    "name": "InvalidPointWitness",
    "inputs": []
  },
  {
    "type": "error",
    "name": "InvalidSafeAddress",
    "inputs": []
  },
  {
    "type": "error",
    "name": "InvalidTicketIndex",
    "inputs": []
  },
  {
    "type": "error",
    "name": "InvalidTicketSignature",
    "inputs": []
  },
  {
    "type": "error",
    "name": "InvalidTokenRecipient",
    "inputs": []
  },
  {
    "type": "error",
    "name": "InvalidTokensReceivedUsage",
    "inputs": []
  },
  {
    "type": "error",
    "name": "InvalidVRFProof",
    "inputs": []
  },
  {
    "type": "error",
    "name": "MultiSigUninitialized",
    "inputs": []
  },
  {
    "type": "error",
    "name": "NoticePeriodNotDue",
    "inputs": []
  },
  {
    "type": "error",
    "name": "SourceEqualsDestination",
    "inputs": []
  },
  {
    "type": "error",
    "name": "TicketIsNotAWin",
    "inputs": []
  },
  {
    "type": "error",
    "name": "TokenTransferFailed",
    "inputs": []
  },
  {
    "type": "error",
    "name": "WrongChannelState",
    "inputs": [
      {
        "name": "reason",
        "type": "string",
        "internalType": "string"
      }
    ]
  },
  {
    "type": "error",
    "name": "WrongToken",
    "inputs": []
  },
  {
    "type": "error",
    "name": "ZeroAddress",
    "inputs": [
      {
        "name": "reason",
        "type": "string",
        "internalType": "string"
      }
    ]
  },
  {
    "type": "error",
    "name": "ZeroInterval",
    "inputs": []
  }
]

Structs

AddressEmptyCode

Custom error with signature AddressEmptyCode(address) and selector 0x9996b315.

AlreadyInitialized

Custom error with signature AlreadyInitialized() and selector 0x0dc149f0.

BalanceExceedsGlobalPerChannelAllowance

Custom error with signature BalanceExceedsGlobalPerChannelAllowance() and selector 0xa4f3bbe4.

ChannelBalanceDecreased

Event with signature ChannelBalanceDecreased(bytes32,bytes32) and selector 0x0c4672f14b63bb6354fac475ee498055a2784455673af224717b9770fd68d8d1.

ChannelBalanceIncreased

Event with signature ChannelBalanceIncreased(bytes32,bytes32) and selector 0xec21547ca1e22edc3f9b4f4e0da94638b5b94dcb18d52dcc072abe5801a8013c.

ChannelClosed

Event with signature ChannelClosed(bytes32,bytes32) and selector 0x197bba684a6afb7ba24f1d265605414b1d0051a6e295d6ff7b6e78e870d7a7f0.

ChannelOpened

Event with signature ChannelOpened(bytes32,address,address,bytes32) and selector 0x004c0177ad15bb302b4419e9ac96065d6f436e0f61972f97bbe272834c40f934.

ContractNotResponsible

Custom error with signature ContractNotResponsible() and selector 0xacd5a823.

DomainSeparatorUpdated

Event with signature DomainSeparatorUpdated(bytes32) and selector 0x771f5240ae5fd8a7640d3fb82fa70aab2fb1dbf35f2ef464f8509946717664c5.

ECDSAInvalidSignature

Custom error with signature ECDSAInvalidSignature() and selector 0xf645eedf.

ECDSAInvalidSignatureLength

Custom error with signature ECDSAInvalidSignatureLength(uint256) and selector 0xfce698f7.

ECDSAInvalidSignatureS

Custom error with signature ECDSAInvalidSignatureS(bytes32) and selector 0xd78bce0c.

ERC777_HOOK_FUND_CHANNEL_MULTI_SIZECall

Function with signature ERC777_HOOK_FUND_CHANNEL_MULTI_SIZE() and selector 0x78d8016d.

ERC777_HOOK_FUND_CHANNEL_MULTI_SIZEReturn

Container type for the return parameters of the ERC777_HOOK_FUND_CHANNEL_MULTI_SIZE() function.

ERC777_HOOK_FUND_CHANNEL_SIZECall

Function with signature ERC777_HOOK_FUND_CHANNEL_SIZE() and selector 0x44dae6f8.

ERC777_HOOK_FUND_CHANNEL_SIZEReturn

Container type for the return parameters of the ERC777_HOOK_FUND_CHANNEL_SIZE() function.

FailedCall

Custom error with signature FailedCall() and selector 0xd6bda275.

HoprChannelsInstance

A HoprChannels instance.

InsufficientChannelBalance

Custom error with signature InsufficientChannelBalance() and selector 0xb147636c.

InvalidBalance

Custom error with signature InvalidBalance() and selector 0xc52e3eff.

InvalidCurvePoint

Custom error with signature InvalidCurvePoint() and selector 0x72454a82.

InvalidFieldElement

Custom error with signature InvalidFieldElement() and selector 0x3ae4ed6b.

InvalidNoticePeriod

Custom error with signature InvalidNoticePeriod() and selector 0xf9ee9107.

InvalidPointWitness

Custom error with signature InvalidPointWitness() and selector 0xedfdcd98.

InvalidSafeAddress

Custom error with signature InvalidSafeAddress() and selector 0x8e9d7c5e.

InvalidTicketIndex

Custom error with signature InvalidTicketIndex() and selector 0x31de4847.

InvalidTicketSignature

Custom error with signature InvalidTicketSignature() and selector 0xcddd5356.

InvalidTokenRecipient

Custom error with signature InvalidTokenRecipient() and selector 0xb9c49108.

InvalidTokensReceivedUsage

Custom error with signature InvalidTokensReceivedUsage() and selector 0x69ee6f28.

InvalidVRFProof

Custom error with signature InvalidVRFProof() and selector 0x95fdbdb8.

LEDGER_VERSIONCall

Function with signature LEDGER_VERSION() and selector 0xddad1902.

LEDGER_VERSIONReturn

Container type for the return parameters of the LEDGER_VERSION() function.

LedgerDomainSeparatorUpdated

Event with signature LedgerDomainSeparatorUpdated(bytes32) and selector 0xa43fad83920fd09445855e854e73c9c532e17402c9ceb09993a2392843a5bdb9.

MAX_USED_BALANCECall

Function with signature MAX_USED_BALANCE() and selector 0x5d2f07c5.

MAX_USED_BALANCEReturn

Container type for the return parameters of the MAX_USED_BALANCE() function.

MIN_USED_BALANCECall

Function with signature MIN_USED_BALANCE() and selector 0x29392e32.

MIN_USED_BALANCEReturn

Container type for the return parameters of the MIN_USED_BALANCE() function.

MultiSigUninitialized

Custom error with signature MultiSigUninitialized() and selector 0x454a20c8.

NOTICE_PERIOD_CHANNEL_CLOSURECall

Function with signature NOTICE_PERIOD_CHANNEL_CLOSURE() and selector 0xd94cc1b4.

NOTICE_PERIOD_CHANNEL_CLOSUREReturn

Container type for the return parameters of the NOTICE_PERIOD_CHANNEL_CLOSURE() function.

NoticePeriodNotDue

Custom error with signature NoticePeriodNotDue() and selector 0x7164032a.

OutgoingChannelClosureInitiated

Event with signature OutgoingChannelClosureInitiated(bytes32,bytes32) and selector 0x0061e8037197b4d91cf8add2a0736613459081f236d1c8ab58549ebbc330b4e3.

RedeemableTicket

SNAPSHOT_INTERVALCall

Function with signature SNAPSHOT_INTERVAL() and selector 0x6d2beef1.

SNAPSHOT_INTERVALReturn

Container type for the return parameters of the SNAPSHOT_INTERVAL() function.

SourceEqualsDestination

Custom error with signature SourceEqualsDestination() and selector 0x97a3aed2.

TOKENCall

Function with signature TOKEN() and selector 0x82bfefc8.

TOKENReturn

Container type for the return parameters of the TOKEN() function.

TOKENS_RECIPIENT_INTERFACE_HASHCall

Function with signature TOKENS_RECIPIENT_INTERFACE_HASH() and selector 0x72581cc0.

TOKENS_RECIPIENT_INTERFACE_HASHReturn

Container type for the return parameters of the TOKENS_RECIPIENT_INTERFACE_HASH() function.

TicketData

TicketIsNotAWin

Custom error with signature TicketIsNotAWin() and selector 0xee835c89.

TicketRedeemed

Event with signature TicketRedeemed(bytes32,bytes32) and selector 0xed22f34d154d253a7f6fd477439be59080a1554aa0d0642686d64223ab544c8a.

TokenTransferFailed

Custom error with signature TokenTransferFailed() and selector 0x045c4b02.

VERSIONCall

Function with signature VERSION() and selector 0xffa1ad74.

VERSIONReturn

Container type for the return parameters of the VERSION() function.

WrongChannelState

Custom error with signature WrongChannelState(string) and selector 0x499463c1.

WrongToken

Custom error with signature WrongToken() and selector 0xa0f3feea.

ZeroAddress

Custom error with signature ZeroAddress(string) and selector 0xeac0d389.

ZeroInterval

Custom error with signature ZeroInterval() and selector 0x346ff607.

_currentBlockTimestampCall

Function with signature _currentBlockTimestamp() and selector 0xb920deed.

_currentBlockTimestampReturn

Container type for the return parameters of the _currentBlockTimestamp() function.

_getChannelIdCall

Function with signature _getChannelId(address,address) and selector 0xbe9babdc.

_getChannelIdReturn

Container type for the return parameters of the _getChannelId(address,address) function.

_getTicketHashCall

Function with signature _getTicketHash(((bytes32,uint96,uint48,uint24,uint56),(bytes32,bytes32),uint256)) and selector 0xab66ccab.

_getTicketHashReturn

Container type for the return parameters of the _getTicketHash(((bytes32,uint96,uint48,uint24,uint56),(bytes32,bytes32),uint256)) function.

_isWinningTicketCall

Function with signature _isWinningTicket(bytes32,((bytes32,uint96,uint48,uint24,uint56),(bytes32,bytes32),uint256),(uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256)) and selector 0x78d149a8.

_isWinningTicketReturn

Container type for the return parameters of the _isWinningTicket(bytes32,((bytes32,uint96,uint48,uint24,uint56),(bytes32,bytes32),uint256),(uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256)) function.

channelStateCall

Function with signature channelState(bytes32) and selector 0x762bd76b.

channelStateReturn

Container type for the return parameters of the channelState(bytes32) function.

channelsCall

Function with signature channels(bytes32) and selector 0x7a7ebd7b.

channelsReturn

Container type for the return parameters of the channels(bytes32) function.

closeIncomingChannelCall

Function with signature closeIncomingChannel(address) and selector 0x1a7ffe7a.

closeIncomingChannelReturn

Container type for the return parameters of the closeIncomingChannel(address) function.

closeIncomingChannelSafeCall

Function with signature closeIncomingChannelSafe(address,address) and selector 0x54a2edf5.

closeIncomingChannelSafeReturn

Container type for the return parameters of the closeIncomingChannelSafe(address,address) function.

constructorCall

Constructor`.

domainSeparatorCall

Function with signature domainSeparator() and selector 0xf698da25.

domainSeparatorReturn

Container type for the return parameters of the domainSeparator() function.

finalizeOutgoingChannelClosureCall

Function with signature finalizeOutgoingChannelClosure(address) and selector 0x23cb3ac0.

finalizeOutgoingChannelClosureReturn

Container type for the return parameters of the finalizeOutgoingChannelClosure(address) function.

finalizeOutgoingChannelClosureSafeCall

Function with signature finalizeOutgoingChannelClosureSafe(address,address) and selector 0x651514bf.

finalizeOutgoingChannelClosureSafeReturn

Container type for the return parameters of the finalizeOutgoingChannelClosureSafe(address,address) function.

fundChannelCall

Function with signature fundChannel(address,uint96) and selector 0xfc55309a.

fundChannelReturn

Container type for the return parameters of the fundChannel(address,uint96) function.

fundChannelSafeCall

Function with signature fundChannelSafe(address,address,uint96) and selector 0x0abec58f.

fundChannelSafeReturn

Container type for the return parameters of the fundChannelSafe(address,address,uint96) function.

initiateOutgoingChannelClosureCall

Function with signature initiateOutgoingChannelClosure(address) and selector 0x7c8e28da.

initiateOutgoingChannelClosureReturn

Container type for the return parameters of the initiateOutgoingChannelClosure(address) function.

initiateOutgoingChannelClosureSafeCall

Function with signature initiateOutgoingChannelClosureSafe(address,address) and selector 0xbda65f45.

initiateOutgoingChannelClosureSafeReturn

Container type for the return parameters of the initiateOutgoingChannelClosureSafe(address,address) function.

latestRootCall

Function with signature latestRoot() and selector 0xd7b0fef1.

latestRootReturn

Container type for the return parameters of the latestRoot() function.

latestSnapshotRootCall

Function with signature latestSnapshotRoot() and selector 0x0df18f94.

latestSnapshotRootReturn

Container type for the return parameters of the latestSnapshotRoot() function.

ledgerDomainSeparatorCall

Function with signature ledgerDomainSeparator() and selector 0xc966c4fe.

ledgerDomainSeparatorReturn

Container type for the return parameters of the ledgerDomainSeparator() function.

multicallCall

Function with signature multicall(bytes[]) and selector 0xac9650d8.

multicallReturn

Container type for the return parameters of the multicall(bytes[]) function.

redeemTicketCall

Function with signature redeemTicket(((bytes32,uint96,uint48,uint24,uint56),(bytes32,bytes32),uint256),(uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256)) and selector 0x65e3fa72.

redeemTicketReturn

Container type for the return parameters of the redeemTicket(((bytes32,uint96,uint48,uint24,uint56),(bytes32,bytes32),uint256),(uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256)) function.

redeemTicketSafeCall

Function with signature redeemTicketSafe(address,((bytes32,uint96,uint48,uint24,uint56),(bytes32,bytes32),uint256),(uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256)) and selector 0x2d50b18b.

redeemTicketSafeReturn

Container type for the return parameters of the redeemTicketSafe(address,((bytes32,uint96,uint48,uint24,uint56),(bytes32,bytes32),uint256),(uint256,uint256,uint256,uint256,uint256,uint256,uint256,uint256)) function.

tokensReceivedCall

Function with signature tokensReceived(address,address,address,uint256,bytes,bytes) and selector 0x0023de29.

tokensReceivedReturn

Container type for the return parameters of the tokensReceived(address,address,address,uint256,bytes,bytes) function.

updateDomainSeparatorCall

Function with signature updateDomainSeparator() and selector 0x89ccfe89.

updateDomainSeparatorReturn

Container type for the return parameters of the updateDomainSeparator() function.

updateLedgerDomainSeparatorCall

Function with signature updateLedgerDomainSeparator() and selector 0xdc96fd50.

updateLedgerDomainSeparatorReturn

Container type for the return parameters of the updateLedgerDomainSeparator() function.

Enums

HoprChannelsCalls

Container for all the HoprChannels function calls.

HoprChannelsErrors

Container for all the HoprChannels custom errors.

HoprChannelsEvents

Container for all the HoprChannels events.

Statics

BYTECODE

The creation / init bytecode of the contract.

DEPLOYED_BYTECODE

The runtime bytecode of the contract, as deployed on the network.

Functions

deploy

Deploys this contract using the given provider and constructor arguments, if any.

deploy_builder

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

new

Creates a new wrapper around an on-chain HoprChannels contract instance.