hopr_bindings::hopr_channels::HoprChannels

Enum HoprChannelsCalls

pub enum HoprChannelsCalls {

    ERC777_HOOK_FUND_CHANNEL_MULTI_SIZE(ERC777_HOOK_FUND_CHANNEL_MULTI_SIZECall),
    ERC777_HOOK_FUND_CHANNEL_SIZE(ERC777_HOOK_FUND_CHANNEL_SIZECall),
    LEDGER_VERSION(LEDGER_VERSIONCall),
    MAX_USED_BALANCE(MAX_USED_BALANCECall),
    MIN_USED_BALANCE(MIN_USED_BALANCECall),
    NOTICE_PERIOD_CHANNEL_CLOSURE(NOTICE_PERIOD_CHANNEL_CLOSURECall),
    SNAPSHOT_INTERVAL(SNAPSHOT_INTERVALCall),
    TOKEN(TOKENCall),
    TOKENS_RECIPIENT_INTERFACE_HASH(TOKENS_RECIPIENT_INTERFACE_HASHCall),
    VERSION(VERSIONCall),
    _currentBlockTimestamp(_currentBlockTimestampCall),
    _getChannelId(_getChannelIdCall),
    _getTicketHash(_getTicketHashCall),
    _isWinningTicket(_isWinningTicketCall),
    channelState(channelStateCall),
    channels(channelsCall),
    closeIncomingChannel(closeIncomingChannelCall),
    closeIncomingChannelSafe(closeIncomingChannelSafeCall),
    domainSeparator(domainSeparatorCall),
    finalizeOutgoingChannelClosure(finalizeOutgoingChannelClosureCall),
    finalizeOutgoingChannelClosureSafe(finalizeOutgoingChannelClosureSafeCall),
    fundChannel(fundChannelCall),
    fundChannelSafe(fundChannelSafeCall),
    initiateOutgoingChannelClosure(initiateOutgoingChannelClosureCall),
    initiateOutgoingChannelClosureSafe(initiateOutgoingChannelClosureSafeCall),
    latestRoot(latestRootCall),
    latestSnapshotRoot(latestSnapshotRootCall),
    ledgerDomainSeparator(ledgerDomainSeparatorCall),
    multicall(multicallCall),
    redeemTicket(redeemTicketCall),
    redeemTicketSafe(redeemTicketSafeCall),
    tokensReceived(tokensReceivedCall),
    updateDomainSeparator(updateDomainSeparatorCall),
    updateLedgerDomainSeparator(updateLedgerDomainSeparatorCall),

}

Container for all the HoprChannels function calls.

Variants

ERC777_HOOK_FUND_CHANNEL_MULTI_SIZE(ERC777_HOOK_FUND_CHANNEL_MULTI_SIZECall)

ERC777_HOOK_FUND_CHANNEL_SIZE(ERC777_HOOK_FUND_CHANNEL_SIZECall)

LEDGER_VERSION(LEDGER_VERSIONCall)

MAX_USED_BALANCE(MAX_USED_BALANCECall)

MIN_USED_BALANCE(MIN_USED_BALANCECall)

NOTICE_PERIOD_CHANNEL_CLOSURE(NOTICE_PERIOD_CHANNEL_CLOSURECall)

SNAPSHOT_INTERVAL(SNAPSHOT_INTERVALCall)

TOKEN(TOKENCall)

TOKENS_RECIPIENT_INTERFACE_HASH(TOKENS_RECIPIENT_INTERFACE_HASHCall)

VERSION(VERSIONCall)

_currentBlockTimestamp(_currentBlockTimestampCall)

_getChannelId(_getChannelIdCall)

_getTicketHash(_getTicketHashCall)

_isWinningTicket(_isWinningTicketCall)

channelState(channelStateCall)

channels(channelsCall)

closeIncomingChannel(closeIncomingChannelCall)

closeIncomingChannelSafe(closeIncomingChannelSafeCall)

domainSeparator(domainSeparatorCall)

finalizeOutgoingChannelClosure(finalizeOutgoingChannelClosureCall)

finalizeOutgoingChannelClosureSafe(finalizeOutgoingChannelClosureSafeCall)

fundChannel(fundChannelCall)

fundChannelSafe(fundChannelSafeCall)

initiateOutgoingChannelClosure(initiateOutgoingChannelClosureCall)

initiateOutgoingChannelClosureSafe(initiateOutgoingChannelClosureSafeCall)

latestRoot(latestRootCall)

latestSnapshotRoot(latestSnapshotRootCall)

ledgerDomainSeparator(ledgerDomainSeparatorCall)

multicall(multicallCall)

redeemTicket(redeemTicketCall)

redeemTicketSafe(redeemTicketSafeCall)

tokensReceived(tokensReceivedCall)

updateDomainSeparator(updateDomainSeparatorCall)

updateLedgerDomainSeparator(updateLedgerDomainSeparatorCall)

Implementations

impl HoprChannelsCalls

pub const SELECTORS: &'static [[u8; 4]]

All the selectors of this enum.

Note that the selectors might not be in the same order as the variants. No guarantees are made about the order of the selectors.

Prefer using SolInterface methods instead.

pub const VARIANT_NAMES: &'static [&'static str]

The names of the variants in the same order as SELECTORS.

pub const SIGNATURES: &'static [&'static str]

The signatures in the same order as SELECTORS.

pub fn signature_by_selector(selector: [u8; 4]) -> Option<&'static str>

Returns the signature for the given selector, if known.

pub fn name_by_selector(selector: [u8; 4]) -> Option<&'static str>

Returns the enum variant name for the given selector, if known.