hopr_bindings::hopr_channels_events

Module HoprChannelsEvents

Generated by the following Solidity interface…

interface HoprChannelsEvents {
    event ChannelBalanceDecreased(bytes32 indexed channelId, bytes32 channel);
    event ChannelBalanceIncreased(bytes32 indexed channelId, bytes32 channel);
    event ChannelClosed(bytes32 indexed channelId, bytes32 channel);
    event ChannelOpened(bytes32 indexed channelId, address indexed source, address indexed destination, bytes32 channel);
    event DomainSeparatorUpdated(bytes32 indexed domainSeparator);
    event OutgoingChannelClosureInitiated(bytes32 indexed channelId, bytes32 channel);
    event TicketRedeemed(bytes32 indexed channelId, bytes32 channel);
}

…which was generated by the following JSON ABI:

[
  {
    "type": "event",
    "name": "ChannelBalanceDecreased",
    "inputs": [
      {
        "name": "channelId",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      },
      {
        "name": "channel",
        "type": "bytes32",
        "indexed": false,
        "internalType": "bytes32"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "ChannelBalanceIncreased",
    "inputs": [
      {
        "name": "channelId",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      },
      {
        "name": "channel",
        "type": "bytes32",
        "indexed": false,
        "internalType": "bytes32"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "ChannelClosed",
    "inputs": [
      {
        "name": "channelId",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      },
      {
        "name": "channel",
        "type": "bytes32",
        "indexed": false,
        "internalType": "bytes32"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "ChannelOpened",
    "inputs": [
      {
        "name": "channelId",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      },
      {
        "name": "source",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "destination",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "channel",
        "type": "bytes32",
        "indexed": false,
        "internalType": "bytes32"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "DomainSeparatorUpdated",
    "inputs": [
      {
        "name": "domainSeparator",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "OutgoingChannelClosureInitiated",
    "inputs": [
      {
        "name": "channelId",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      },
      {
        "name": "channel",
        "type": "bytes32",
        "indexed": false,
        "internalType": "bytes32"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "TicketRedeemed",
    "inputs": [
      {
        "name": "channelId",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      },
      {
        "name": "channel",
        "type": "bytes32",
        "indexed": false,
        "internalType": "bytes32"
      }
    ],
    "anonymous": false
  }
]

Structs

ChannelBalanceDecreased

Event with signature ChannelBalanceDecreased(bytes32,bytes32) and selector 0x0c4672f14b63bb6354fac475ee498055a2784455673af224717b9770fd68d8d1.

ChannelBalanceIncreased

Event with signature ChannelBalanceIncreased(bytes32,bytes32) and selector 0xec21547ca1e22edc3f9b4f4e0da94638b5b94dcb18d52dcc072abe5801a8013c.

ChannelClosed

Event with signature ChannelClosed(bytes32,bytes32) and selector 0x197bba684a6afb7ba24f1d265605414b1d0051a6e295d6ff7b6e78e870d7a7f0.

ChannelOpened

Event with signature ChannelOpened(bytes32,address,address,bytes32) and selector 0x004c0177ad15bb302b4419e9ac96065d6f436e0f61972f97bbe272834c40f934.

DomainSeparatorUpdated

Event with signature DomainSeparatorUpdated(bytes32) and selector 0x771f5240ae5fd8a7640d3fb82fa70aab2fb1dbf35f2ef464f8509946717664c5.

HoprChannelsEventsInstance

A HoprChannelsEvents instance.

OutgoingChannelClosureInitiated

Event with signature OutgoingChannelClosureInitiated(bytes32,bytes32) and selector 0x0061e8037197b4d91cf8add2a0736613459081f236d1c8ab58549ebbc330b4e3.

TicketRedeemed

Event with signature TicketRedeemed(bytes32,bytes32) and selector 0xed22f34d154d253a7f6fd477439be59080a1554aa0d0642686d64223ab544c8a.

Enums

HoprChannelsEventsEvents

Container for all the HoprChannelsEvents events.

Statics

BYTECODE

The creation / init bytecode of the contract.

DEPLOYED_BYTECODE

The runtime bytecode of the contract, as deployed on the network.

Functions

deploy

Deploys this contract using the given provider and constructor arguments, if any.

deploy_builder

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

new

Creates a new wrapper around an on-chain HoprChannelsEvents contract instance.