hopr_bindings

Module hopr_ledger

Modules

hopr_ledger

This module was auto-generated with ethers-rs Abigen. More information at: https://github.com/gakonst/ethers-rs

Structs

HoprLedger

LedgerDomainSeparatorCall

Container type for all input parameters for the ledgerDomainSeparator function with signature ledgerDomainSeparator() and selector 0xc966c4fe

LedgerDomainSeparatorReturn

Container type for all return fields from the ledgerDomainSeparator function with signature ledgerDomainSeparator() and selector 0xc966c4fe

LedgerDomainSeparatorUpdatedFilter

LedgerVersionCall

Container type for all input parameters for the LEDGER_VERSION function with signature LEDGER_VERSION() and selector 0xddad1902

LedgerVersionReturn

Container type for all return fields from the LEDGER_VERSION function with signature LEDGER_VERSION() and selector 0xddad1902

UpdateLedgerDomainSeparatorCall

Container type for all input parameters for the updateLedgerDomainSeparator function with signature updateLedgerDomainSeparator() and selector 0xdc96fd50

Enums

HoprLedgerCalls

Container type for all of the contract’s call

Statics

HOPRLEDGER_ABI

The parsed JSON ABI of the contract.