hopr_bindings::hopr_ledger

Module HoprLedger

Generated by the following Solidity interface…

interface HoprLedger {
    struct RootStruct {
        bytes28 rootHash;
        uint32 timestamp;
    }

    error ZeroInterval();

    event LedgerDomainSeparatorUpdated(bytes32 indexed ledgerDomainSeparator);

    function LEDGER_VERSION() external view returns (string memory);
    function SNAPSHOT_INTERVAL() external view returns (uint256);
    function latestRoot() external view returns (RootStruct memory);
    function latestSnapshotRoot() external view returns (RootStruct memory);
    function ledgerDomainSeparator() external view returns (bytes32);
    function updateLedgerDomainSeparator() external;
}

…which was generated by the following JSON ABI:

[
  {
    "type": "function",
    "name": "LEDGER_VERSION",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "string",
        "internalType": "string"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "SNAPSHOT_INTERVAL",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "latestRoot",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "tuple",
        "internalType": "struct HoprLedger.RootStruct",
        "components": [
          {
            "name": "rootHash",
            "type": "bytes28",
            "internalType": "bytes28"
          },
          {
            "name": "timestamp",
            "type": "uint32",
            "internalType": "uint32"
          }
        ]
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "latestSnapshotRoot",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "tuple",
        "internalType": "struct HoprLedger.RootStruct",
        "components": [
          {
            "name": "rootHash",
            "type": "bytes28",
            "internalType": "bytes28"
          },
          {
            "name": "timestamp",
            "type": "uint32",
            "internalType": "uint32"
          }
        ]
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "ledgerDomainSeparator",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "updateLedgerDomainSeparator",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "event",
    "name": "LedgerDomainSeparatorUpdated",
    "inputs": [
      {
        "name": "ledgerDomainSeparator",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      }
    ],
    "anonymous": false
  },
  {
    "type": "error",
    "name": "ZeroInterval",
    "inputs": []
  }
]

Structs

HoprLedgerInstance

A HoprLedger instance.

LEDGER_VERSIONCall

Function with signature LEDGER_VERSION() and selector 0xddad1902.

LEDGER_VERSIONReturn

Container type for the return parameters of the LEDGER_VERSION() function.

LedgerDomainSeparatorUpdated

Event with signature LedgerDomainSeparatorUpdated(bytes32) and selector 0xa43fad83920fd09445855e854e73c9c532e17402c9ceb09993a2392843a5bdb9.

RootStruct

SNAPSHOT_INTERVALCall

Function with signature SNAPSHOT_INTERVAL() and selector 0x6d2beef1.

SNAPSHOT_INTERVALReturn

Container type for the return parameters of the SNAPSHOT_INTERVAL() function.

ZeroInterval

Custom error with signature ZeroInterval() and selector 0x346ff607.

latestRootCall

Function with signature latestRoot() and selector 0xd7b0fef1.

latestRootReturn

Container type for the return parameters of the latestRoot() function.

latestSnapshotRootCall

Function with signature latestSnapshotRoot() and selector 0x0df18f94.

latestSnapshotRootReturn

Container type for the return parameters of the latestSnapshotRoot() function.

ledgerDomainSeparatorCall

Function with signature ledgerDomainSeparator() and selector 0xc966c4fe.

ledgerDomainSeparatorReturn

Container type for the return parameters of the ledgerDomainSeparator() function.

updateLedgerDomainSeparatorCall

Function with signature updateLedgerDomainSeparator() and selector 0xdc96fd50.

updateLedgerDomainSeparatorReturn

Container type for the return parameters of the updateLedgerDomainSeparator() function.

Enums

HoprLedgerCalls

Container for all the HoprLedger function calls.

HoprLedgerErrors

Container for all the HoprLedger custom errors.

HoprLedgerEvents

Container for all the HoprLedger events.

Statics

BYTECODE

The creation / init bytecode of the contract.

DEPLOYED_BYTECODE

The runtime bytecode of the contract, as deployed on the network.

Functions

deploy

Deploys this contract using the given provider and constructor arguments, if any.

deploy_builder

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

new

Creates a new wrapper around an on-chain HoprLedger contract instance.