hopr_bindings::hopr_node_management_module

Module HoprNodeManagementModule

Generated by the following Solidity interface…

library Enum {
    type Operation is uint8;
}

interface HoprNodeManagementModule {
    type GranularPermission is uint8;
    type Target is uint256;

    error AddressEmptyCode(address target);
    error AddressIsZero();
    error ArrayTooLong();
    error ArraysDifferentLength();
    error CalldataOutOfBounds();
    error CannotChangeOwner();
    error DefaultPermissionRejected();
    error DelegateCallNotAllowed();
    error ERC1967InvalidImplementation(address implementation);
    error ERC1967NonPayable();
    error FailedCall();
    error FailedToSendEthToNode();
    error FunctionSignatureTooShort();
    error GranularPermissionRejected();
    error InvalidInitialization();
    error LengthIsZero();
    error NoMembership();
    error NodePermissionRejected();
    error NonExistentKey();
    error NotInitializing();
    error OwnableInvalidOwner(address owner);
    error OwnableUnauthorizedAccount(address account);
    error ParameterNotAllowed();
    error PermissionNotConfigured();
    error PermissionNotFound();
    error SafeMultisendSameAddress();
    error SendNotAllowed();
    error TargetAddressNotAllowed();
    error TargetIsNotScoped();
    error TargetIsScoped();
    error TooManyCapabilities();
    error UUPSUnauthorizedCallContext();
    error UUPSUnsupportedProxiableUUID(bytes32 slot);
    error UnacceptableMultiSendOffset();
    error WithMembership();

    event ExecutionFailure();
    event ExecutionSuccess();
    event Initialized(uint64 version);
    event NodeAdded(address indexed node);
    event NodeRemoved(address indexed node);
    event OwnershipTransferred(address indexed previousOwner, address indexed newOwner);
    event RevokedTarget(address indexed targetAddress);
    event ScopedGranularChannelCapability(address indexed targetAddress, bytes32 indexed channelId, bytes4 selector, GranularPermission permission);
    event ScopedGranularSendCapability(address indexed nodeAddress, address indexed recipientAddress, GranularPermission permission);
    event ScopedGranularTokenCapability(address indexed nodeAddress, address indexed targetAddress, address indexed recipientAddress, bytes4 selector, GranularPermission permission);
    event ScopedTargetChannels(address indexed targetAddress, Target target);
    event ScopedTargetSend(address indexed targetAddress, Target target);
    event ScopedTargetToken(address indexed targetAddress, Target target);
    event SetMultisendAddress(address indexed multisendAddress);
    event Upgraded(address indexed implementation);

    constructor();

    function UPGRADE_INTERFACE_VERSION() external view returns (string memory);
    function VERSION() external view returns (string memory);
    function addChannelsAndTokenTarget(Target defaultTarget) external;
    function addNode(address nodeAddress) external payable;
    function addNodes(address[] memory nodeAddresses) external payable;
    function decodeFunctionSigsAndPermissions(bytes32 encoded, uint256 length) external pure returns (bytes4[] memory functionSigs, GranularPermission[] memory permissions);
    function encodeFunctionSigsAndPermissions(bytes4[] memory functionSigs, GranularPermission[] memory permissions) external pure returns (bytes32 encoded, uint256 length);
    function execTransactionFromModule(address to, uint256 value, bytes memory data, Enum.Operation operation) external returns (bool success);
    function execTransactionFromModuleReturnData(address to, uint256 value, bytes memory data, Enum.Operation operation) external returns (bool, bytes memory);
    function getGranularPermissions(bytes32 capabilityKey, bytes32 pairId) external view returns (GranularPermission);
    function getTargets() external view returns (Target[] memory);
    function includeNode(Target nodeDefaultTarget) external;
    function includeNodes(address[] memory nodeAddresses) external payable;
    function initialize(bytes memory initParams) external;
    function isHoprNodeManagementModule() external view returns (bool);
    function isNode(address nodeAddress) external view returns (bool);
    function multisend() external view returns (address);
    function owner() external view returns (address);
    function proxiableUUID() external view returns (bytes32);
    function removeNode(address nodeAddress) external;
    function renounceOwnership() external;
    function revokeTarget(address targetAddress) external;
    function scopeChannelsCapabilities(address targetAddress, bytes32 channelId, bytes32 encodedSigsPermissions) external;
    function scopeSendCapability(address nodeAddress, address beneficiary, GranularPermission permission) external;
    function scopeTargetChannels(Target defaultTarget) external;
    function scopeTargetSend(Target defaultTarget) external;
    function scopeTargetToken(Target defaultTarget) external;
    function scopeTokenCapabilities(address nodeAddress, address targetAddress, address beneficiary, bytes32 encodedSigsPermissions) external;
    function setMultisend(address _multisend) external;
    function transferOwnership(address newOwner) external;
    function tryGetTarget(address targetAddress) external view returns (bool, Target);
    function upgradeToAndCall(address newImplementation, bytes memory data) external payable;
}

…which was generated by the following JSON ABI:

[
  {
    "type": "constructor",
    "inputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "UPGRADE_INTERFACE_VERSION",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "string",
        "internalType": "string"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "VERSION",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "string",
        "internalType": "string"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "addChannelsAndTokenTarget",
    "inputs": [
      {
        "name": "defaultTarget",
        "type": "uint256",
        "internalType": "Target"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "addNode",
    "inputs": [
      {
        "name": "nodeAddress",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "payable"
  },
  {
    "type": "function",
    "name": "addNodes",
    "inputs": [
      {
        "name": "nodeAddresses",
        "type": "address[]",
        "internalType": "address[]"
      }
    ],
    "outputs": [],
    "stateMutability": "payable"
  },
  {
    "type": "function",
    "name": "decodeFunctionSigsAndPermissions",
    "inputs": [
      {
        "name": "encoded",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "length",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "functionSigs",
        "type": "bytes4[]",
        "internalType": "bytes4[]"
      },
      {
        "name": "permissions",
        "type": "uint8[]",
        "internalType": "enum GranularPermission[]"
      }
    ],
    "stateMutability": "pure"
  },
  {
    "type": "function",
    "name": "encodeFunctionSigsAndPermissions",
    "inputs": [
      {
        "name": "functionSigs",
        "type": "bytes4[]",
        "internalType": "bytes4[]"
      },
      {
        "name": "permissions",
        "type": "uint8[]",
        "internalType": "enum GranularPermission[]"
      }
    ],
    "outputs": [
      {
        "name": "encoded",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "length",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "pure"
  },
  {
    "type": "function",
    "name": "execTransactionFromModule",
    "inputs": [
      {
        "name": "to",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "value",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "data",
        "type": "bytes",
        "internalType": "bytes"
      },
      {
        "name": "operation",
        "type": "uint8",
        "internalType": "enum Enum.Operation"
      }
    ],
    "outputs": [
      {
        "name": "success",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "execTransactionFromModuleReturnData",
    "inputs": [
      {
        "name": "to",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "value",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "data",
        "type": "bytes",
        "internalType": "bytes"
      },
      {
        "name": "operation",
        "type": "uint8",
        "internalType": "enum Enum.Operation"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      },
      {
        "name": "",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "getGranularPermissions",
    "inputs": [
      {
        "name": "capabilityKey",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "pairId",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint8",
        "internalType": "enum GranularPermission"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getTargets",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256[]",
        "internalType": "Target[]"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "includeNode",
    "inputs": [
      {
        "name": "nodeDefaultTarget",
        "type": "uint256",
        "internalType": "Target"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "includeNodes",
    "inputs": [
      {
        "name": "nodeAddresses",
        "type": "address[]",
        "internalType": "address[]"
      }
    ],
    "outputs": [],
    "stateMutability": "payable"
  },
  {
    "type": "function",
    "name": "initialize",
    "inputs": [
      {
        "name": "initParams",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "isHoprNodeManagementModule",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "isNode",
    "inputs": [
      {
        "name": "nodeAddress",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "multisend",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "owner",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "proxiableUUID",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "removeNode",
    "inputs": [
      {
        "name": "nodeAddress",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "renounceOwnership",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "revokeTarget",
    "inputs": [
      {
        "name": "targetAddress",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "scopeChannelsCapabilities",
    "inputs": [
      {
        "name": "targetAddress",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "channelId",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "encodedSigsPermissions",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "scopeSendCapability",
    "inputs": [
      {
        "name": "nodeAddress",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "beneficiary",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "permission",
        "type": "uint8",
        "internalType": "enum GranularPermission"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "scopeTargetChannels",
    "inputs": [
      {
        "name": "defaultTarget",
        "type": "uint256",
        "internalType": "Target"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "scopeTargetSend",
    "inputs": [
      {
        "name": "defaultTarget",
        "type": "uint256",
        "internalType": "Target"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "scopeTargetToken",
    "inputs": [
      {
        "name": "defaultTarget",
        "type": "uint256",
        "internalType": "Target"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "scopeTokenCapabilities",
    "inputs": [
      {
        "name": "nodeAddress",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "targetAddress",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "beneficiary",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "encodedSigsPermissions",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "setMultisend",
    "inputs": [
      {
        "name": "_multisend",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "transferOwnership",
    "inputs": [
      {
        "name": "newOwner",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "tryGetTarget",
    "inputs": [
      {
        "name": "targetAddress",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      },
      {
        "name": "",
        "type": "uint256",
        "internalType": "Target"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "upgradeToAndCall",
    "inputs": [
      {
        "name": "newImplementation",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "data",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "outputs": [],
    "stateMutability": "payable"
  },
  {
    "type": "event",
    "name": "ExecutionFailure",
    "inputs": [],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "ExecutionSuccess",
    "inputs": [],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Initialized",
    "inputs": [
      {
        "name": "version",
        "type": "uint64",
        "indexed": false,
        "internalType": "uint64"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "NodeAdded",
    "inputs": [
      {
        "name": "node",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "NodeRemoved",
    "inputs": [
      {
        "name": "node",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "OwnershipTransferred",
    "inputs": [
      {
        "name": "previousOwner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "newOwner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "RevokedTarget",
    "inputs": [
      {
        "name": "targetAddress",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "ScopedGranularChannelCapability",
    "inputs": [
      {
        "name": "targetAddress",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "channelId",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      },
      {
        "name": "selector",
        "type": "bytes4",
        "indexed": false,
        "internalType": "bytes4"
      },
      {
        "name": "permission",
        "type": "uint8",
        "indexed": false,
        "internalType": "enum GranularPermission"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "ScopedGranularSendCapability",
    "inputs": [
      {
        "name": "nodeAddress",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "recipientAddress",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "permission",
        "type": "uint8",
        "indexed": false,
        "internalType": "enum GranularPermission"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "ScopedGranularTokenCapability",
    "inputs": [
      {
        "name": "nodeAddress",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "targetAddress",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "recipientAddress",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "selector",
        "type": "bytes4",
        "indexed": false,
        "internalType": "bytes4"
      },
      {
        "name": "permission",
        "type": "uint8",
        "indexed": false,
        "internalType": "enum GranularPermission"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "ScopedTargetChannels",
    "inputs": [
      {
        "name": "targetAddress",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "target",
        "type": "uint256",
        "indexed": false,
        "internalType": "Target"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "ScopedTargetSend",
    "inputs": [
      {
        "name": "targetAddress",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "target",
        "type": "uint256",
        "indexed": false,
        "internalType": "Target"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "ScopedTargetToken",
    "inputs": [
      {
        "name": "targetAddress",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "target",
        "type": "uint256",
        "indexed": false,
        "internalType": "Target"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "SetMultisendAddress",
    "inputs": [
      {
        "name": "multisendAddress",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Upgraded",
    "inputs": [
      {
        "name": "implementation",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "error",
    "name": "AddressEmptyCode",
    "inputs": [
      {
        "name": "target",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "AddressIsZero",
    "inputs": []
  },
  {
    "type": "error",
    "name": "ArrayTooLong",
    "inputs": []
  },
  {
    "type": "error",
    "name": "ArraysDifferentLength",
    "inputs": []
  },
  {
    "type": "error",
    "name": "CalldataOutOfBounds",
    "inputs": []
  },
  {
    "type": "error",
    "name": "CannotChangeOwner",
    "inputs": []
  },
  {
    "type": "error",
    "name": "DefaultPermissionRejected",
    "inputs": []
  },
  {
    "type": "error",
    "name": "DelegateCallNotAllowed",
    "inputs": []
  },
  {
    "type": "error",
    "name": "ERC1967InvalidImplementation",
    "inputs": [
      {
        "name": "implementation",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "ERC1967NonPayable",
    "inputs": []
  },
  {
    "type": "error",
    "name": "FailedCall",
    "inputs": []
  },
  {
    "type": "error",
    "name": "FailedToSendEthToNode",
    "inputs": []
  },
  {
    "type": "error",
    "name": "FunctionSignatureTooShort",
    "inputs": []
  },
  {
    "type": "error",
    "name": "GranularPermissionRejected",
    "inputs": []
  },
  {
    "type": "error",
    "name": "InvalidInitialization",
    "inputs": []
  },
  {
    "type": "error",
    "name": "LengthIsZero",
    "inputs": []
  },
  {
    "type": "error",
    "name": "NoMembership",
    "inputs": []
  },
  {
    "type": "error",
    "name": "NodePermissionRejected",
    "inputs": []
  },
  {
    "type": "error",
    "name": "NonExistentKey",
    "inputs": []
  },
  {
    "type": "error",
    "name": "NotInitializing",
    "inputs": []
  },
  {
    "type": "error",
    "name": "OwnableInvalidOwner",
    "inputs": [
      {
        "name": "owner",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "OwnableUnauthorizedAccount",
    "inputs": [
      {
        "name": "account",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "ParameterNotAllowed",
    "inputs": []
  },
  {
    "type": "error",
    "name": "PermissionNotConfigured",
    "inputs": []
  },
  {
    "type": "error",
    "name": "PermissionNotFound",
    "inputs": []
  },
  {
    "type": "error",
    "name": "SafeMultisendSameAddress",
    "inputs": []
  },
  {
    "type": "error",
    "name": "SendNotAllowed",
    "inputs": []
  },
  {
    "type": "error",
    "name": "TargetAddressNotAllowed",
    "inputs": []
  },
  {
    "type": "error",
    "name": "TargetIsNotScoped",
    "inputs": []
  },
  {
    "type": "error",
    "name": "TargetIsScoped",
    "inputs": []
  },
  {
    "type": "error",
    "name": "TooManyCapabilities",
    "inputs": []
  },
  {
    "type": "error",
    "name": "UUPSUnauthorizedCallContext",
    "inputs": []
  },
  {
    "type": "error",
    "name": "UUPSUnsupportedProxiableUUID",
    "inputs": [
      {
        "name": "slot",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ]
  },
  {
    "type": "error",
    "name": "UnacceptableMultiSendOffset",
    "inputs": []
  },
  {
    "type": "error",
    "name": "WithMembership",
    "inputs": []
  }
]

Structs

AddressEmptyCode

Custom error with signature AddressEmptyCode(address) and selector 0x9996b315.

AddressIsZero

Custom error with signature AddressIsZero() and selector 0x867915ab.

ArrayTooLong

Custom error with signature ArrayTooLong() and selector 0xbd26cc38.

ArraysDifferentLength

Custom error with signature ArraysDifferentLength() and selector 0x74f4d537.

CalldataOutOfBounds

Custom error with signature CalldataOutOfBounds() and selector 0x742638b4.

CannotChangeOwner

Custom error with signature CannotChangeOwner() and selector 0xfd670ebe.

DefaultPermissionRejected

Custom error with signature DefaultPermissionRejected() and selector 0x58723037.

DelegateCallNotAllowed

Custom error with signature DelegateCallNotAllowed() and selector 0x0d89438e.

ERC1967InvalidImplementation

Custom error with signature ERC1967InvalidImplementation(address) and selector 0x4c9c8ce3.

ERC1967NonPayable

Custom error with signature ERC1967NonPayable() and selector 0xb398979f.

ExecutionFailure

Event with signature ExecutionFailure() and selector 0xc24d93608a03d263ff191d7677141f5e94c496e593108f3aae0cb5b70494c4d3.

ExecutionSuccess

Event with signature ExecutionSuccess() and selector 0x4e2e86d21375ebcbf6e93df5ebdd5a915bf830245904c3b54f48adf0170aae4b.

FailedCall

Custom error with signature FailedCall() and selector 0xd6bda275.

FailedToSendEthToNode

Custom error with signature FailedToSendEthToNode() and selector 0x9a73ee99.

FunctionSignatureTooShort

Custom error with signature FunctionSignatureTooShort() and selector 0x4684c122.

GranularPermission

GranularPermissionRejected

Custom error with signature GranularPermissionRejected() and selector 0x864dd1e7.

HoprNodeManagementModuleInstance

A HoprNodeManagementModule instance.

Initialized

Event with signature Initialized(uint64) and selector 0xc7f505b2f371ae2175ee4913f4499e1f2633a7b5936321eed1cdaeb6115181d2.

InvalidInitialization

Custom error with signature InvalidInitialization() and selector 0xf92ee8a9.

LengthIsZero

Custom error with signature LengthIsZero() and selector 0x0c71ec17.

NoMembership

Custom error with signature NoMembership() and selector 0xfd8e9f28.

NodeAdded

Event with signature NodeAdded(address) and selector 0xb25d03aaf308d7291709be1ea28b800463cf3a9a4c4a5555d7333a964c1dfebd.

NodePermissionRejected

Custom error with signature NodePermissionRejected() and selector 0x6eb0315f.

NodeRemoved

Event with signature NodeRemoved(address) and selector 0xcfc24166db4bb677e857cacabd1541fb2b30645021b27c5130419589b84db52b.

NonExistentKey

Custom error with signature NonExistentKey() and selector 0x2d0519ad.

NotInitializing

Custom error with signature NotInitializing() and selector 0xd7e6bcf8.

OwnableInvalidOwner

Custom error with signature OwnableInvalidOwner(address) and selector 0x1e4fbdf7.

OwnableUnauthorizedAccount

Custom error with signature OwnableUnauthorizedAccount(address) and selector 0x118cdaa7.

OwnershipTransferred

Event with signature OwnershipTransferred(address,address) and selector 0x8be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e0.

ParameterNotAllowed

Custom error with signature ParameterNotAllowed() and selector 0x31e98246.

PermissionNotConfigured

Custom error with signature PermissionNotConfigured() and selector 0x46ad4588.

PermissionNotFound

Custom error with signature PermissionNotFound() and selector 0xd8455a13.

RevokedTarget

Event with signature RevokedTarget(address) and selector 0x0dfce1ea4ba1eeba891ffb2a066790fbc293a9e517fe61d49d156a30165f93f3.

SafeMultisendSameAddress

Custom error with signature SafeMultisendSameAddress() and selector 0x598a0e21.

ScopedGranularChannelCapability

Event with signature ScopedGranularChannelCapability(address,bytes32,bytes4,uint8) and selector 0xf2ffd4f09d58d06824188033d3318d06eb957bfb1a8ffed9af78e1f19168b904.

ScopedGranularSendCapability

Event with signature ScopedGranularSendCapability(address,address,uint8) and selector 0x7487530ddff120799505e52b1b19b6933f85a9eeae9220c80a7ad7c429b612ae.

ScopedGranularTokenCapability

Event with signature ScopedGranularTokenCapability(address,address,address,bytes4,uint8) and selector 0xa3df710420b01cc30ff300309abbc7fadd4630d4ab385b0f5a126fb4babe762b.

ScopedTargetChannels

Event with signature ScopedTargetChannels(address,uint256) and selector 0x5ffb06b0b0e8ad6a8f3c5831d499dfa612d9c9d4dc107bbd66f18f61a6492e71.

ScopedTargetSend

Event with signature ScopedTargetSend(address,uint256) and selector 0x1ee2791f2caf0e92a9dc32a37a9ea53ab6ac7a6fb8f2d090e53a067d3a43f6ac.

ScopedTargetToken

Event with signature ScopedTargetToken(address,uint256) and selector 0xaaf26bb12aa89ee96bbe19667a6a055727b75d3f6ed7b8b611ef6519180209d6.

SendNotAllowed

Custom error with signature SendNotAllowed() and selector 0x09e9cd49.

SetMultisendAddress

Event with signature SetMultisendAddress(address) and selector 0x5fe6aabf4e790843df43ae0e22b58620066fb389295bedc06a92df6c3b28777d.

Target

TargetAddressNotAllowed

Custom error with signature TargetAddressNotAllowed() and selector 0xef3440ac.

TargetIsNotScoped

Custom error with signature TargetIsNotScoped() and selector 0x4a890321.

TargetIsScoped

Custom error with signature TargetIsScoped() and selector 0xe8c07d2a.

TooManyCapabilities

Custom error with signature TooManyCapabilities() and selector 0xb44af9af.

UPGRADE_INTERFACE_VERSIONCall

Function with signature UPGRADE_INTERFACE_VERSION() and selector 0xad3cb1cc.

UPGRADE_INTERFACE_VERSIONReturn

Container type for the return parameters of the UPGRADE_INTERFACE_VERSION() function.

UUPSUnauthorizedCallContext

Custom error with signature UUPSUnauthorizedCallContext() and selector 0xe07c8dba.

UUPSUnsupportedProxiableUUID

Custom error with signature UUPSUnsupportedProxiableUUID(bytes32) and selector 0xaa1d49a4.

UnacceptableMultiSendOffset

Custom error with signature UnacceptableMultiSendOffset() and selector 0x7ed11137.

Upgraded

Event with signature Upgraded(address) and selector 0xbc7cd75a20ee27fd9adebab32041f755214dbc6bffa90cc0225b39da2e5c2d3b.

VERSIONCall

Function with signature VERSION() and selector 0xffa1ad74.

VERSIONReturn

Container type for the return parameters of the VERSION() function.

WithMembership

Custom error with signature WithMembership() and selector 0xe3a05a94.

addChannelsAndTokenTargetCall

Function with signature addChannelsAndTokenTarget(uint256) and selector 0xa2450f89.

addChannelsAndTokenTargetReturn

Container type for the return parameters of the addChannelsAndTokenTarget(uint256) function.

addNodeCall

Function with signature addNode(address) and selector 0x9d95f1cc.

addNodeReturn

Container type for the return parameters of the addNode(address) function.

addNodesCall

Function with signature addNodes(address[]) and selector 0xdf9620eb.

addNodesReturn

Container type for the return parameters of the addNodes(address[]) function.

constructorCall

Constructor`.

decodeFunctionSigsAndPermissionsCall

Function with signature decodeFunctionSigsAndPermissions(bytes32,uint256) and selector 0x60976c4b.

decodeFunctionSigsAndPermissionsReturn

Container type for the return parameters of the decodeFunctionSigsAndPermissions(bytes32,uint256) function.

encodeFunctionSigsAndPermissionsCall

Function with signature encodeFunctionSigsAndPermissions(bytes4[],uint8[]) and selector 0x56f55117.

encodeFunctionSigsAndPermissionsReturn

Container type for the return parameters of the encodeFunctionSigsAndPermissions(bytes4[],uint8[]) function.

execTransactionFromModuleCall

Function with signature execTransactionFromModule(address,uint256,bytes,uint8) and selector 0x468721a7.

execTransactionFromModuleReturn

Container type for the return parameters of the execTransactionFromModule(address,uint256,bytes,uint8) function.

execTransactionFromModuleReturnDataCall

Function with signature execTransactionFromModuleReturnData(address,uint256,bytes,uint8) and selector 0x5229073f.

execTransactionFromModuleReturnDataReturn

Container type for the return parameters of the execTransactionFromModuleReturnData(address,uint256,bytes,uint8) function.

getGranularPermissionsCall

Function with signature getGranularPermissions(bytes32,bytes32) and selector 0xdc446a4a.

getGranularPermissionsReturn

Container type for the return parameters of the getGranularPermissions(bytes32,bytes32) function.

getTargetsCall

Function with signature getTargets() and selector 0x63fe3b56.

getTargetsReturn

Container type for the return parameters of the getTargets() function.

includeNodeCall

Function with signature includeNode(uint256) and selector 0xb5736962.

includeNodeReturn

Container type for the return parameters of the includeNode(uint256) function.

includeNodesCall

Function with signature includeNodes(address[]) and selector 0x110dcee7.

includeNodesReturn

Container type for the return parameters of the includeNodes(address[]) function.

initializeCall

Function with signature initialize(bytes) and selector 0x439fab91.

initializeReturn

Container type for the return parameters of the initialize(bytes) function.

isHoprNodeManagementModuleCall

Function with signature isHoprNodeManagementModule() and selector 0x4a1ba408.

isHoprNodeManagementModuleReturn

Container type for the return parameters of the isHoprNodeManagementModule() function.

isNodeCall

Function with signature isNode(address) and selector 0x01750152.

isNodeReturn

Container type for the return parameters of the isNode(address) function.

multisendCall

Function with signature multisend() and selector 0x294402cc.

multisendReturn

Container type for the return parameters of the multisend() function.

ownerCall

Function with signature owner() and selector 0x8da5cb5b.

ownerReturn

Container type for the return parameters of the owner() function.

proxiableUUIDCall

Function with signature proxiableUUID() and selector 0x52d1902d.

proxiableUUIDReturn

Container type for the return parameters of the proxiableUUID() function.

removeNodeCall

Function with signature removeNode(address) and selector 0xb2b99ec9.

removeNodeReturn

Container type for the return parameters of the removeNode(address) function.

renounceOwnershipCall

Function with signature renounceOwnership() and selector 0x715018a6.

renounceOwnershipReturn

Container type for the return parameters of the renounceOwnership() function.

revokeTargetCall

Function with signature revokeTarget(address) and selector 0x3401cde8.

revokeTargetReturn

Container type for the return parameters of the revokeTarget(address) function.

scopeChannelsCapabilitiesCall

Function with signature scopeChannelsCapabilities(address,bytes32,bytes32) and selector 0xfa19501d.

scopeChannelsCapabilitiesReturn

Container type for the return parameters of the scopeChannelsCapabilities(address,bytes32,bytes32) function.

scopeSendCapabilityCall

Function with signature scopeSendCapability(address,address,uint8) and selector 0xc68c3a83.

scopeSendCapabilityReturn

Container type for the return parameters of the scopeSendCapability(address,address,uint8) function.

scopeTargetChannelsCall

Function with signature scopeTargetChannels(uint256) and selector 0x739c4b08.

scopeTargetChannelsReturn

Container type for the return parameters of the scopeTargetChannels(uint256) function.

scopeTargetSendCall

Function with signature scopeTargetSend(uint256) and selector 0xdc06109d.

scopeTargetSendReturn

Container type for the return parameters of the scopeTargetSend(uint256) function.

scopeTargetTokenCall

Function with signature scopeTargetToken(uint256) and selector 0xa76c9a2f.

scopeTargetTokenReturn

Container type for the return parameters of the scopeTargetToken(uint256) function.

scopeTokenCapabilitiesCall

Function with signature scopeTokenCapabilities(address,address,address,bytes32) and selector 0xc68605c8.

scopeTokenCapabilitiesReturn

Container type for the return parameters of the scopeTokenCapabilities(address,address,address,bytes32) function.

setMultisendCall

Function with signature setMultisend(address) and selector 0x8b95eccd.

setMultisendReturn

Container type for the return parameters of the setMultisend(address) function.

transferOwnershipCall

Function with signature transferOwnership(address) and selector 0xf2fde38b.

transferOwnershipReturn

Container type for the return parameters of the transferOwnership(address) function.

tryGetTargetCall

Function with signature tryGetTarget(address) and selector 0xdf4e6f8a.

tryGetTargetReturn

Container type for the return parameters of the tryGetTarget(address) function.

upgradeToAndCallCall

Function with signature upgradeToAndCall(address,bytes) and selector 0x4f1ef286.

upgradeToAndCallReturn

Container type for the return parameters of the upgradeToAndCall(address,bytes) function.

Enums

HoprNodeManagementModuleCalls

Container for all the HoprNodeManagementModule function calls.

HoprNodeManagementModuleErrors

Container for all the HoprNodeManagementModule custom errors.

HoprNodeManagementModuleEvents

Container for all the HoprNodeManagementModule events.

Statics

BYTECODE

The creation / init bytecode of the contract.

DEPLOYED_BYTECODE

The runtime bytecode of the contract, as deployed on the network.

Functions

deploy

Deploys this contract using the given provider and constructor arguments, if any.

deploy_builder

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

new

Creates a new wrapper around an on-chain HoprNodeManagementModule contract instance.