hopr_bindings::hopr_node_safe_migration

Module HoprNodeSafeMigration

Generated by the following Solidity interface…

interface HoprNodeSafeMigration {
    error ModuleNotEnabledInSafe();
    error UndeployedContract(string contractName);

    event ChangedMasterCopy(address singleton);
    event ChangedModuleImplementation(address moduleProxy);
    event DeployedNewV4Module(address newModuleProxy);
    event SafeAndModuleMigrationCompleted(address safeProxy, address oldModuleProxy, address newModuleProxy);

    constructor(address moduleSingleton, address nodeStakeFactory);

    function FACTORY_ADDRESS() external view returns (address);
    function MIGRATION_SINGLETON() external view returns (address);
    function MODULE_SINGLETON() external view returns (address);
    function SAFE_FALLBACK_HANDLER() external view returns (address);
    function SAFE_L2_SINGLETON() external view returns (address);
    function SAFE_SINGLETON() external view returns (address);
    function deployNewV4Module(bytes32 defaultTarget, uint256 nonce, address[] memory nodes) external returns (address newModuleProxy);
    function migrateL2Singleton() external;
    function migrateL2WithFallbackHandler() external;
    function migrateModuleSingleton(address moduleProxy, bytes memory data) external;
    function migrateSafeV141ToL2AndMigrateToUpgradeableModule(address oldModuleProxy, bytes32 defaultTarget, uint256 nonce, address[] memory nodes) external;
    function migrateSingleton() external;
    function migrateWithFallbackHandler() external;
}

…which was generated by the following JSON ABI:

[
  {
    "type": "constructor",
    "inputs": [
      {
        "name": "moduleSingleton",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "nodeStakeFactory",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "FACTORY_ADDRESS",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "MIGRATION_SINGLETON",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "MODULE_SINGLETON",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "SAFE_FALLBACK_HANDLER",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "SAFE_L2_SINGLETON",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "SAFE_SINGLETON",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "deployNewV4Module",
    "inputs": [
      {
        "name": "defaultTarget",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "nonce",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "nodes",
        "type": "address[]",
        "internalType": "address[]"
      }
    ],
    "outputs": [
      {
        "name": "newModuleProxy",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "migrateL2Singleton",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "migrateL2WithFallbackHandler",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "migrateModuleSingleton",
    "inputs": [
      {
        "name": "moduleProxy",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "data",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "migrateSafeV141ToL2AndMigrateToUpgradeableModule",
    "inputs": [
      {
        "name": "oldModuleProxy",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "defaultTarget",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "nonce",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "nodes",
        "type": "address[]",
        "internalType": "address[]"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "migrateSingleton",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "migrateWithFallbackHandler",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "event",
    "name": "ChangedMasterCopy",
    "inputs": [
      {
        "name": "singleton",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "ChangedModuleImplementation",
    "inputs": [
      {
        "name": "moduleProxy",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "DeployedNewV4Module",
    "inputs": [
      {
        "name": "newModuleProxy",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "SafeAndModuleMigrationCompleted",
    "inputs": [
      {
        "name": "safeProxy",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      },
      {
        "name": "oldModuleProxy",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      },
      {
        "name": "newModuleProxy",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "error",
    "name": "ModuleNotEnabledInSafe",
    "inputs": []
  },
  {
    "type": "error",
    "name": "UndeployedContract",
    "inputs": [
      {
        "name": "contractName",
        "type": "string",
        "internalType": "string"
      }
    ]
  }
]

Structs

ChangedMasterCopy

Event with signature ChangedMasterCopy(address) and selector 0x75e41bc35ff1bf14d81d1d2f649c0084a0f974f9289c803ec9898eeec4c8d0b8.

ChangedModuleImplementation

Event with signature ChangedModuleImplementation(address) and selector 0xc86c9607623339ec5f0adff99fe1777064788722ba91a0c8897fc96e354af9d5.

DeployedNewV4Module

Event with signature DeployedNewV4Module(address) and selector 0x70f67686ee0df86aea03fffe1cdb388f83717e8b8a1243f162cfc9bcbb0bcb3a.

FACTORY_ADDRESSCall

Function with signature FACTORY_ADDRESS() and selector 0x8823151b.

FACTORY_ADDRESSReturn

Container type for the return parameters of the FACTORY_ADDRESS() function.

HoprNodeSafeMigrationInstance

A HoprNodeSafeMigration instance.

MIGRATION_SINGLETONCall

Function with signature MIGRATION_SINGLETON() and selector 0x72f7a956.

MIGRATION_SINGLETONReturn

Container type for the return parameters of the MIGRATION_SINGLETON() function.

MODULE_SINGLETONCall

Function with signature MODULE_SINGLETON() and selector 0xa3c15cc9.

MODULE_SINGLETONReturn

Container type for the return parameters of the MODULE_SINGLETON() function.

ModuleNotEnabledInSafe

Custom error with signature ModuleNotEnabledInSafe() and selector 0x7bb58776.

SAFE_FALLBACK_HANDLERCall

Function with signature SAFE_FALLBACK_HANDLER() and selector 0x0d7101f7.

SAFE_FALLBACK_HANDLERReturn

Container type for the return parameters of the SAFE_FALLBACK_HANDLER() function.

SAFE_L2_SINGLETONCall

Function with signature SAFE_L2_SINGLETON() and selector 0x9bf47d6e.

SAFE_L2_SINGLETONReturn

Container type for the return parameters of the SAFE_L2_SINGLETON() function.

SAFE_SINGLETONCall

Function with signature SAFE_SINGLETON() and selector 0xcaa12add.

SAFE_SINGLETONReturn

Container type for the return parameters of the SAFE_SINGLETON() function.

SafeAndModuleMigrationCompleted

Event with signature SafeAndModuleMigrationCompleted(address,address,address) and selector 0x33122e5cced82e071229ec54c5cb7db55fe210b78d8a006cba16378993e83e3e.

UndeployedContract

Custom error with signature UndeployedContract(string) and selector 0x1e476f2b.

constructorCall

Constructor`.

deployNewV4ModuleCall

Function with signature deployNewV4Module(bytes32,uint256,address[]) and selector 0xe3d2809e.

deployNewV4ModuleReturn

Container type for the return parameters of the deployNewV4Module(bytes32,uint256,address[]) function.

migrateL2SingletonCall

Function with signature migrateL2Singleton() and selector 0x07f464a4.

migrateL2SingletonReturn

Container type for the return parameters of the migrateL2Singleton() function.

migrateL2WithFallbackHandlerCall

Function with signature migrateL2WithFallbackHandler() and selector 0x68cb3d94.

migrateL2WithFallbackHandlerReturn

Container type for the return parameters of the migrateL2WithFallbackHandler() function.

migrateModuleSingletonCall

Function with signature migrateModuleSingleton(address,bytes) and selector 0x4dd5be6d.

migrateModuleSingletonReturn

Container type for the return parameters of the migrateModuleSingleton(address,bytes) function.

migrateSafeV141ToL2AndMigrateToUpgradeableModuleCall

Function with signature migrateSafeV141ToL2AndMigrateToUpgradeableModule(address,bytes32,uint256,address[]) and selector 0x6210cce3.

migrateSafeV141ToL2AndMigrateToUpgradeableModuleReturn

Container type for the return parameters of the migrateSafeV141ToL2AndMigrateToUpgradeableModule(address,bytes32,uint256,address[]) function.

migrateSingletonCall

Function with signature migrateSingleton() and selector 0xf6682ab0.

migrateSingletonReturn

Container type for the return parameters of the migrateSingleton() function.

migrateWithFallbackHandlerCall

Function with signature migrateWithFallbackHandler() and selector 0xed007fc6.

migrateWithFallbackHandlerReturn

Container type for the return parameters of the migrateWithFallbackHandler() function.

Enums

HoprNodeSafeMigrationCalls

Container for all the HoprNodeSafeMigration function calls.

HoprNodeSafeMigrationErrors

Container for all the HoprNodeSafeMigration custom errors.

HoprNodeSafeMigrationEvents

Container for all the HoprNodeSafeMigration events.

Statics

BYTECODE

The creation / init bytecode of the contract.

DEPLOYED_BYTECODE

The runtime bytecode of the contract, as deployed on the network.

Functions

deploy

Deploys this contract using the given provider and constructor arguments, if any.

deploy_builder

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

new

Creates a new wrapper around an on-chain HoprNodeSafeMigration contract instance.