hopr_bindings::hopr_node_stake_factory

Module HoprNodeStakeFactory

Generated by the following Solidity interface…

interface HoprNodeStakeFactory {
    struct HoprNetwork {
        address tokenAddress;
        uint256 defaultTokenAllowance;
        bytes32 defaultAnnouncementTarget;
    }
    struct SafeLibAddress {
        address safeAddress;
        address safeProxyFactoryAddress;
        address fallbackHandlerAddress;
        address multiSendAddress;
    }
    struct SafeModuleDeployment {
        address safeProxyInstance;
        address moduleProxyInstance;
    }

    error Create2EmptyBytecode();
    error ExistingSafeModule();
    error FailedDeployment();
    error InsufficientBalance(uint256 balance, uint256 needed);
    error InvalidFunctionSelector();
    error InvalidOwner();
    error NotTokenRecipient();
    error OwnableInvalidOwner(address owner);
    error OwnableUnauthorizedAccount(address account);
    error SafeERC20FailedOperation(address token);
    error TooFewOwners();
    error UnauthorizedToken();

    event HoprNodeStakeHoprNetworkUpdated(HoprNetwork hoprNetwork);
    event HoprNodeStakeModuleUpdated(address moduleImplementation);
    event HoprNodeStakeSafeLibUpdated(SafeLibAddress safeLibAddresses);
    event NewHoprNodeStakeModule(address instance);
    event NewHoprNodeStakeModuleForSafe(address module, address safe);
    event NewHoprNodeStakeSafe(address instance);
    event OwnershipTransferStarted(address indexed previousOwner, address indexed newOwner);
    event OwnershipTransferred(address indexed previousOwner, address indexed newOwner);

    constructor(address _moduleSingletonAddress, address _announcementAddress, address initialOwner);

    function DEPLOYSAFEANDMODULEANDINCLUDENODES_IDENTIFIER() external view returns (bytes32);
    function DEPLOYSAFEMODULE_FUNCTION_IDENTIFIER() external view returns (bytes32);
    function TOKENS_RECIPIENT_INTERFACE_HASH() external view returns (bytes32);
    function acceptOwnership() external;
    function clone(uint256 nonce, bytes32 defaultTarget, address[] memory admins) external returns (address, address payable);
    function defaultHoprNetwork() external view returns (address tokenAddress, uint256 defaultTokenAllowance, bytes32 defaultAnnouncementTarget);
    function deployModule(address safeProxyAddr, bytes32 defaultTarget, uint256 nonce) external returns (address moduleProxy);
    function getAllDeployments() external view returns (SafeModuleDeployment[] memory);
    function getDeploymentAtIndex(uint256 index) external view returns (SafeModuleDeployment memory);
    function getDeploymentsCount() external view returns (uint256);
    function isSafeDeployedByFactory(address safeProxyAddress) external view returns (bool);
    function moduleSingletonAddress() external view returns (address);
    function owner() external view returns (address);
    function pendingOwner() external view returns (address);
    function predictModuleAddress(bytes32 salt, address safe, bytes32 defaultTarget) external view returns (address predicted);
    function predictModuleAddress(address caller, uint256 nonce, address safe, bytes32 defaultTarget) external view returns (address predicted);
    function predictSafeAddress(address[] memory admins, uint256 nonce) external view returns (address predicted);
    function reclaimErc20(address token) external;
    function renounceOwnership() external;
    function safeLibAddresses() external view returns (address safeAddress, address safeProxyFactoryAddress, address fallbackHandlerAddress, address multiSendAddress);
    function safeVersion() external pure returns (string memory);
    function tokensReceived(address, address from, address to, uint256 amount, bytes memory userData, bytes memory) external;
    function transferOwnership(address newOwner) external;
    function tryGetDeployment(address safeProxyAddress) external view returns (bool, uint256, SafeModuleDeployment memory);
    function updateErc1820Implementer(address _newImplementer) external;
    function updateHoprNetwork(HoprNetwork memory _newHoprNetwork) external;
    function updateModuleSingletonAddress(address _newModuleSingletonAddress) external;
    function updateSafeLibAddress(SafeLibAddress memory _newSafeLibAddresses) external;
}

…which was generated by the following JSON ABI:

[
  {
    "type": "constructor",
    "inputs": [
      {
        "name": "_moduleSingletonAddress",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "_announcementAddress",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "initialOwner",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "DEPLOYSAFEANDMODULEANDINCLUDENODES_IDENTIFIER",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "DEPLOYSAFEMODULE_FUNCTION_IDENTIFIER",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "TOKENS_RECIPIENT_INTERFACE_HASH",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "acceptOwnership",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "clone",
    "inputs": [
      {
        "name": "nonce",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "defaultTarget",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "admins",
        "type": "address[]",
        "internalType": "address[]"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "",
        "type": "address",
        "internalType": "address payable"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "defaultHoprNetwork",
    "inputs": [],
    "outputs": [
      {
        "name": "tokenAddress",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "defaultTokenAllowance",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "defaultAnnouncementTarget",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "deployModule",
    "inputs": [
      {
        "name": "safeProxyAddr",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "defaultTarget",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "nonce",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "moduleProxy",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "getAllDeployments",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "tuple[]",
        "internalType": "struct SafeModuleDeployment[]",
        "components": [
          {
            "name": "safeProxyInstance",
            "type": "address",
            "internalType": "address"
          },
          {
            "name": "moduleProxyInstance",
            "type": "address",
            "internalType": "address"
          }
        ]
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getDeploymentAtIndex",
    "inputs": [
      {
        "name": "index",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "tuple",
        "internalType": "struct SafeModuleDeployment",
        "components": [
          {
            "name": "safeProxyInstance",
            "type": "address",
            "internalType": "address"
          },
          {
            "name": "moduleProxyInstance",
            "type": "address",
            "internalType": "address"
          }
        ]
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getDeploymentsCount",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "isSafeDeployedByFactory",
    "inputs": [
      {
        "name": "safeProxyAddress",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "moduleSingletonAddress",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "owner",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "pendingOwner",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "predictModuleAddress",
    "inputs": [
      {
        "name": "salt",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "safe",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "defaultTarget",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "outputs": [
      {
        "name": "predicted",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "predictModuleAddress",
    "inputs": [
      {
        "name": "caller",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "nonce",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "safe",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "defaultTarget",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "outputs": [
      {
        "name": "predicted",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "predictSafeAddress",
    "inputs": [
      {
        "name": "admins",
        "type": "address[]",
        "internalType": "address[]"
      },
      {
        "name": "nonce",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "predicted",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "reclaimErc20",
    "inputs": [
      {
        "name": "token",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "renounceOwnership",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "safeLibAddresses",
    "inputs": [],
    "outputs": [
      {
        "name": "safeAddress",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "safeProxyFactoryAddress",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "fallbackHandlerAddress",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "multiSendAddress",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "safeVersion",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "string",
        "internalType": "string"
      }
    ],
    "stateMutability": "pure"
  },
  {
    "type": "function",
    "name": "tokensReceived",
    "inputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "from",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "to",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "amount",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "userData",
        "type": "bytes",
        "internalType": "bytes"
      },
      {
        "name": "",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "transferOwnership",
    "inputs": [
      {
        "name": "newOwner",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "tryGetDeployment",
    "inputs": [
      {
        "name": "safeProxyAddress",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      },
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "",
        "type": "tuple",
        "internalType": "struct SafeModuleDeployment",
        "components": [
          {
            "name": "safeProxyInstance",
            "type": "address",
            "internalType": "address"
          },
          {
            "name": "moduleProxyInstance",
            "type": "address",
            "internalType": "address"
          }
        ]
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "updateErc1820Implementer",
    "inputs": [
      {
        "name": "_newImplementer",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "updateHoprNetwork",
    "inputs": [
      {
        "name": "_newHoprNetwork",
        "type": "tuple",
        "internalType": "struct HoprNodeStakeFactory.HoprNetwork",
        "components": [
          {
            "name": "tokenAddress",
            "type": "address",
            "internalType": "address"
          },
          {
            "name": "defaultTokenAllowance",
            "type": "uint256",
            "internalType": "uint256"
          },
          {
            "name": "defaultAnnouncementTarget",
            "type": "bytes32",
            "internalType": "bytes32"
          }
        ]
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "updateModuleSingletonAddress",
    "inputs": [
      {
        "name": "_newModuleSingletonAddress",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "updateSafeLibAddress",
    "inputs": [
      {
        "name": "_newSafeLibAddresses",
        "type": "tuple",
        "internalType": "struct HoprNodeStakeFactory.SafeLibAddress",
        "components": [
          {
            "name": "safeAddress",
            "type": "address",
            "internalType": "address"
          },
          {
            "name": "safeProxyFactoryAddress",
            "type": "address",
            "internalType": "address"
          },
          {
            "name": "fallbackHandlerAddress",
            "type": "address",
            "internalType": "address"
          },
          {
            "name": "multiSendAddress",
            "type": "address",
            "internalType": "address"
          }
        ]
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "event",
    "name": "HoprNodeStakeHoprNetworkUpdated",
    "inputs": [
      {
        "name": "hoprNetwork",
        "type": "tuple",
        "indexed": false,
        "internalType": "struct HoprNodeStakeFactory.HoprNetwork",
        "components": [
          {
            "name": "tokenAddress",
            "type": "address",
            "internalType": "address"
          },
          {
            "name": "defaultTokenAllowance",
            "type": "uint256",
            "internalType": "uint256"
          },
          {
            "name": "defaultAnnouncementTarget",
            "type": "bytes32",
            "internalType": "bytes32"
          }
        ]
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "HoprNodeStakeModuleUpdated",
    "inputs": [
      {
        "name": "moduleImplementation",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "HoprNodeStakeSafeLibUpdated",
    "inputs": [
      {
        "name": "safeLibAddresses",
        "type": "tuple",
        "indexed": false,
        "internalType": "struct HoprNodeStakeFactory.SafeLibAddress",
        "components": [
          {
            "name": "safeAddress",
            "type": "address",
            "internalType": "address"
          },
          {
            "name": "safeProxyFactoryAddress",
            "type": "address",
            "internalType": "address"
          },
          {
            "name": "fallbackHandlerAddress",
            "type": "address",
            "internalType": "address"
          },
          {
            "name": "multiSendAddress",
            "type": "address",
            "internalType": "address"
          }
        ]
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "NewHoprNodeStakeModule",
    "inputs": [
      {
        "name": "instance",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "NewHoprNodeStakeModuleForSafe",
    "inputs": [
      {
        "name": "module",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      },
      {
        "name": "safe",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "NewHoprNodeStakeSafe",
    "inputs": [
      {
        "name": "instance",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "OwnershipTransferStarted",
    "inputs": [
      {
        "name": "previousOwner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "newOwner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "OwnershipTransferred",
    "inputs": [
      {
        "name": "previousOwner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "newOwner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "error",
    "name": "Create2EmptyBytecode",
    "inputs": []
  },
  {
    "type": "error",
    "name": "ExistingSafeModule",
    "inputs": []
  },
  {
    "type": "error",
    "name": "FailedDeployment",
    "inputs": []
  },
  {
    "type": "error",
    "name": "InsufficientBalance",
    "inputs": [
      {
        "name": "balance",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "needed",
        "type": "uint256",
        "internalType": "uint256"
      }
    ]
  },
  {
    "type": "error",
    "name": "InvalidFunctionSelector",
    "inputs": []
  },
  {
    "type": "error",
    "name": "InvalidOwner",
    "inputs": []
  },
  {
    "type": "error",
    "name": "NotTokenRecipient",
    "inputs": []
  },
  {
    "type": "error",
    "name": "OwnableInvalidOwner",
    "inputs": [
      {
        "name": "owner",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "OwnableUnauthorizedAccount",
    "inputs": [
      {
        "name": "account",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "SafeERC20FailedOperation",
    "inputs": [
      {
        "name": "token",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "TooFewOwners",
    "inputs": []
  },
  {
    "type": "error",
    "name": "UnauthorizedToken",
    "inputs": []
  }
]

Structs

Create2EmptyBytecode

Custom error with signature Create2EmptyBytecode() and selector 0x4ca249dc.

DEPLOYSAFEANDMODULEANDINCLUDENODES_IDENTIFIERCall

Function with signature DEPLOYSAFEANDMODULEANDINCLUDENODES_IDENTIFIER() and selector 0xf8ffe3c0.

DEPLOYSAFEANDMODULEANDINCLUDENODES_IDENTIFIERReturn

Container type for the return parameters of the DEPLOYSAFEANDMODULEANDINCLUDENODES_IDENTIFIER() function.

DEPLOYSAFEMODULE_FUNCTION_IDENTIFIERCall

Function with signature DEPLOYSAFEMODULE_FUNCTION_IDENTIFIER() and selector 0x63f17ff9.

DEPLOYSAFEMODULE_FUNCTION_IDENTIFIERReturn

Container type for the return parameters of the DEPLOYSAFEMODULE_FUNCTION_IDENTIFIER() function.

ExistingSafeModule

Custom error with signature ExistingSafeModule() and selector 0xad7f101c.

FailedDeployment

Custom error with signature FailedDeployment() and selector 0xb06ebf3d.

HoprNetwork

HoprNodeStakeFactoryInstance

A HoprNodeStakeFactory instance.

HoprNodeStakeHoprNetworkUpdated

Event with signature HoprNodeStakeHoprNetworkUpdated((address,uint256,bytes32)) and selector 0x54cfcf4bdb6f20632ba544cd684130fc9db4dc2fe6360926d6b7866297f0315e.

HoprNodeStakeModuleUpdated

Event with signature HoprNodeStakeModuleUpdated(address) and selector 0xdd60725a88508fe2dcb769bd824280a2a01356198c795515d175abc4bc40b74e.

HoprNodeStakeSafeLibUpdated

Event with signature HoprNodeStakeSafeLibUpdated((address,address,address,address)) and selector 0x87eb507b2b42eb659807d7a1f63c89b8e22c674689603eac07a3713f06fec9c3.

InsufficientBalance

Custom error with signature InsufficientBalance(uint256,uint256) and selector 0xcf479181.

InvalidFunctionSelector

Custom error with signature InvalidFunctionSelector() and selector 0x42868c9b.

InvalidOwner

Custom error with signature InvalidOwner() and selector 0x49e27cff.

NewHoprNodeStakeModule

Event with signature NewHoprNodeStakeModule(address) and selector 0xf966ce717349a940e403b8faae71fdaf573d8b9847deea16fca3d870cf6c292d.

NewHoprNodeStakeModuleForSafe

Event with signature NewHoprNodeStakeModuleForSafe(address,address) and selector 0x83a5c31bfee49d254d95e3946d8e8434e5975382fa8b849182db9e251aaaaa26.

NewHoprNodeStakeSafe

Event with signature NewHoprNodeStakeSafe(address) and selector 0x8231d169f416b666ae7fa43faa24a18899738075a53f32c97617d173b189e386.

NotTokenRecipient

Custom error with signature NotTokenRecipient() and selector 0x720773f2.

OwnableInvalidOwner

Custom error with signature OwnableInvalidOwner(address) and selector 0x1e4fbdf7.

OwnableUnauthorizedAccount

Custom error with signature OwnableUnauthorizedAccount(address) and selector 0x118cdaa7.

OwnershipTransferStarted

Event with signature OwnershipTransferStarted(address,address) and selector 0x38d16b8cac22d99fc7c124b9cd0de2d3fa1faef420bfe791d8c362d765e22700.

OwnershipTransferred

Event with signature OwnershipTransferred(address,address) and selector 0x8be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e0.

SafeERC20FailedOperation

Custom error with signature SafeERC20FailedOperation(address) and selector 0x5274afe7.

SafeLibAddress

SafeModuleDeployment

TOKENS_RECIPIENT_INTERFACE_HASHCall

Function with signature TOKENS_RECIPIENT_INTERFACE_HASH() and selector 0x72581cc0.

TOKENS_RECIPIENT_INTERFACE_HASHReturn

Container type for the return parameters of the TOKENS_RECIPIENT_INTERFACE_HASH() function.

TooFewOwners

Custom error with signature TooFewOwners() and selector 0x0d7fd1aa.

UnauthorizedToken

Custom error with signature UnauthorizedToken() and selector 0x7e5f9349.

acceptOwnershipCall

Function with signature acceptOwnership() and selector 0x79ba5097.

acceptOwnershipReturn

Container type for the return parameters of the acceptOwnership() function.

cloneCall

Function with signature clone(uint256,bytes32,address[]) and selector 0x9a94addf.

cloneReturn

Container type for the return parameters of the clone(uint256,bytes32,address[]) function.

constructorCall

Constructor`.

defaultHoprNetworkCall

Function with signature defaultHoprNetwork() and selector 0xa7a6e4b8.

defaultHoprNetworkReturn

Container type for the return parameters of the defaultHoprNetwork() function.

deployModuleCall

Function with signature deployModule(address,bytes32,uint256) and selector 0x696ab635.

deployModuleReturn

Container type for the return parameters of the deployModule(address,bytes32,uint256) function.

getAllDeploymentsCall

Function with signature getAllDeployments() and selector 0x8eec00b8.

getAllDeploymentsReturn

Container type for the return parameters of the getAllDeployments() function.

getDeploymentAtIndexCall

Function with signature getDeploymentAtIndex(uint256) and selector 0xa2bc64e4.

getDeploymentAtIndexReturn

Container type for the return parameters of the getDeploymentAtIndex(uint256) function.

getDeploymentsCountCall

Function with signature getDeploymentsCount() and selector 0x950ff8a8.

getDeploymentsCountReturn

Container type for the return parameters of the getDeploymentsCount() function.

isSafeDeployedByFactoryCall

Function with signature isSafeDeployedByFactory(address) and selector 0x0f6c7b95.

isSafeDeployedByFactoryReturn

Container type for the return parameters of the isSafeDeployedByFactory(address) function.

moduleSingletonAddressCall

Function with signature moduleSingletonAddress() and selector 0x9aef1ed9.

moduleSingletonAddressReturn

Container type for the return parameters of the moduleSingletonAddress() function.

ownerCall

Function with signature owner() and selector 0x8da5cb5b.

ownerReturn

Container type for the return parameters of the owner() function.

pendingOwnerCall

Function with signature pendingOwner() and selector 0xe30c3978.

pendingOwnerReturn

Container type for the return parameters of the pendingOwner() function.

predictModuleAddress_0Call

Function with signature predictModuleAddress(bytes32,address,bytes32) and selector 0x8afc3bf5.

predictModuleAddress_0Return

Container type for the return parameters of the predictModuleAddress(bytes32,address,bytes32) function.

predictModuleAddress_1Call

Function with signature predictModuleAddress(address,uint256,address,bytes32) and selector 0xd48a66bd.

predictModuleAddress_1Return

Container type for the return parameters of the predictModuleAddress(address,uint256,address,bytes32) function.

predictSafeAddressCall

Function with signature predictSafeAddress(address[],uint256) and selector 0xc248a4e7.

predictSafeAddressReturn

Container type for the return parameters of the predictSafeAddress(address[],uint256) function.

reclaimErc20Call

Function with signature reclaimErc20(address) and selector 0x79d101fa.

reclaimErc20Return

Container type for the return parameters of the reclaimErc20(address) function.

renounceOwnershipCall

Function with signature renounceOwnership() and selector 0x715018a6.

renounceOwnershipReturn

Container type for the return parameters of the renounceOwnership() function.

safeLibAddressesCall

Function with signature safeLibAddresses() and selector 0x8947890a.

safeLibAddressesReturn

Container type for the return parameters of the safeLibAddresses() function.

safeVersionCall

Function with signature safeVersion() and selector 0x7ef2721f.

safeVersionReturn

Container type for the return parameters of the safeVersion() function.

tokensReceivedCall

Function with signature tokensReceived(address,address,address,uint256,bytes,bytes) and selector 0x0023de29.

tokensReceivedReturn

Container type for the return parameters of the tokensReceived(address,address,address,uint256,bytes,bytes) function.

transferOwnershipCall

Function with signature transferOwnership(address) and selector 0xf2fde38b.

transferOwnershipReturn

Container type for the return parameters of the transferOwnership(address) function.

tryGetDeploymentCall

Function with signature tryGetDeployment(address) and selector 0x22d917eb.

tryGetDeploymentReturn

Container type for the return parameters of the tryGetDeployment(address) function.

updateErc1820ImplementerCall

Function with signature updateErc1820Implementer(address) and selector 0xabed205b.

updateErc1820ImplementerReturn

Container type for the return parameters of the updateErc1820Implementer(address) function.

updateHoprNetworkCall

Function with signature updateHoprNetwork((address,uint256,bytes32)) and selector 0x32838f6b.

updateHoprNetworkReturn

Container type for the return parameters of the updateHoprNetwork((address,uint256,bytes32)) function.

updateModuleSingletonAddressCall

Function with signature updateModuleSingletonAddress(address) and selector 0x0233296b.

updateModuleSingletonAddressReturn

Container type for the return parameters of the updateModuleSingletonAddress(address) function.

updateSafeLibAddressCall

Function with signature updateSafeLibAddress((address,address,address,address)) and selector 0xfa2aeab4.

updateSafeLibAddressReturn

Container type for the return parameters of the updateSafeLibAddress((address,address,address,address)) function.

Enums

HoprNodeStakeFactoryCalls

Container for all the HoprNodeStakeFactory function calls.

HoprNodeStakeFactoryErrors

Container for all the HoprNodeStakeFactory custom errors.

HoprNodeStakeFactoryEvents

Container for all the HoprNodeStakeFactory events.

Statics

BYTECODE

The creation / init bytecode of the contract.

DEPLOYED_BYTECODE

The runtime bytecode of the contract, as deployed on the network.

Functions

deploy

Deploys this contract using the given provider and constructor arguments, if any.

deploy_builder

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

new

Creates a new wrapper around an on-chain HoprNodeStakeFactory contract instance.