hopr_bindings::hopr_node_stake_factory::HoprNodeStakeFactory

Struct HoprNodeStakeFactoryInstance

pub struct HoprNodeStakeFactoryInstance<P, N = Ethereum> { /* private fields */ }

A HoprNodeStakeFactory instance.

Contains type-safe methods for interacting with an on-chain instance of the HoprNodeStakeFactory contract located at a given address, using a given provider P.

If the contract bytecode is available (see the sol! documentation on how to provide it), the deploy and deploy_builder methods can be used to deploy a new instance of the contract.

See the module-level documentation for all the available methods.

Implementations

impl<P: Provider<N>, N: Network> HoprNodeStakeFactoryInstance<P, N>

Instantiation and getters/setters.

pub const fn new(address: Address, __provider: P) -> Self

Creates a new wrapper around an on-chain HoprNodeStakeFactory contract instance.

See the wrapper’s documentation for more details.

pub async fn deploy( __provider: P, _moduleSingletonAddress: Address, _announcementAddress: Address, initialOwner: Address, ) -> Result<HoprNodeStakeFactoryInstance<P, N>>

Deploys this contract using the given provider and constructor arguments, if any.

Returns a new instance of the contract, if the deployment was successful.

For more fine-grained control over the deployment process, use deploy_builder instead.

pub fn deploy_builder( __provider: P, _moduleSingletonAddress: Address, _announcementAddress: Address, initialOwner: Address, ) -> RawCallBuilder<P, N>

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

This is a simple wrapper around creating a RawCallBuilder with the data set to the bytecode concatenated with the constructor’s ABI-encoded arguments.

pub const fn address(&self) -> &Address

Returns a reference to the address.

pub fn set_address(&mut self, address: Address)

Sets the address.

pub fn at(self, address: Address) -> Self

Sets the address and returns self.

pub const fn provider(&self) -> &P

Returns a reference to the provider.

impl<P: Clone, N> HoprNodeStakeFactoryInstance<&P, N>

pub fn with_cloned_provider(self) -> HoprNodeStakeFactoryInstance<P, N>

Clones the provider and returns a new instance with the cloned provider.

impl<P: Provider<N>, N: Network> HoprNodeStakeFactoryInstance<P, N>

Function calls.

pub fn call_builder<C: SolCall>(&self, call: &C) -> SolCallBuilder<&P, C, N>

Creates a new call builder using this contract instance’s provider and address.

Note that the call can be any function call, not just those defined in this contract. Prefer using the other methods for building type-safe contract calls.

pub fn DEPLOYSAFEANDMODULEANDINCLUDENODES_IDENTIFIER( &self, ) -> SolCallBuilder<&P, DEPLOYSAFEANDMODULEANDINCLUDENODES_IDENTIFIERCall, N>

Creates a new call builder for the [DEPLOYSAFEANDMODULEANDINCLUDENODES_IDENTIFIER] function.

pub fn DEPLOYSAFEMODULE_FUNCTION_IDENTIFIER( &self, ) -> SolCallBuilder<&P, DEPLOYSAFEMODULE_FUNCTION_IDENTIFIERCall, N>

Creates a new call builder for the [DEPLOYSAFEMODULE_FUNCTION_IDENTIFIER] function.

pub fn TOKENS_RECIPIENT_INTERFACE_HASH( &self, ) -> SolCallBuilder<&P, TOKENS_RECIPIENT_INTERFACE_HASHCall, N>

Creates a new call builder for the [TOKENS_RECIPIENT_INTERFACE_HASH] function.

pub fn acceptOwnership(&self) -> SolCallBuilder<&P, acceptOwnershipCall, N>

Creates a new call builder for the [acceptOwnership] function.

pub fn clone( &self, nonce: U256, defaultTarget: FixedBytes<32>, admins: Vec<Address>, ) -> SolCallBuilder<&P, cloneCall, N>

Creates a new call builder for the [clone] function.

pub fn defaultHoprNetwork( &self, ) -> SolCallBuilder<&P, defaultHoprNetworkCall, N>

Creates a new call builder for the [defaultHoprNetwork] function.

pub fn deployModule( &self, safeProxyAddr: Address, defaultTarget: FixedBytes<32>, nonce: U256, ) -> SolCallBuilder<&P, deployModuleCall, N>

Creates a new call builder for the [deployModule] function.

pub fn getAllDeployments(&self) -> SolCallBuilder<&P, getAllDeploymentsCall, N>

Creates a new call builder for the [getAllDeployments] function.

pub fn getDeploymentAtIndex( &self, index: U256, ) -> SolCallBuilder<&P, getDeploymentAtIndexCall, N>

Creates a new call builder for the [getDeploymentAtIndex] function.

pub fn getDeploymentsCount( &self, ) -> SolCallBuilder<&P, getDeploymentsCountCall, N>

Creates a new call builder for the [getDeploymentsCount] function.

pub fn isSafeDeployedByFactory( &self, safeProxyAddress: Address, ) -> SolCallBuilder<&P, isSafeDeployedByFactoryCall, N>

Creates a new call builder for the [isSafeDeployedByFactory] function.

pub fn moduleSingletonAddress( &self, ) -> SolCallBuilder<&P, moduleSingletonAddressCall, N>

Creates a new call builder for the [moduleSingletonAddress] function.

pub fn owner(&self) -> SolCallBuilder<&P, ownerCall, N>

Creates a new call builder for the [owner] function.

pub fn pendingOwner(&self) -> SolCallBuilder<&P, pendingOwnerCall, N>

Creates a new call builder for the [pendingOwner] function.

pub fn predictModuleAddress_0( &self, salt: FixedBytes<32>, safe: Address, defaultTarget: FixedBytes<32>, ) -> SolCallBuilder<&P, predictModuleAddress_0Call, N>

Creates a new call builder for the [predictModuleAddress_0] function.

pub fn predictModuleAddress_1( &self, caller: Address, nonce: U256, safe: Address, defaultTarget: FixedBytes<32>, ) -> SolCallBuilder<&P, predictModuleAddress_1Call, N>

Creates a new call builder for the [predictModuleAddress_1] function.

pub fn predictSafeAddress( &self, admins: Vec<Address>, nonce: U256, ) -> SolCallBuilder<&P, predictSafeAddressCall, N>

Creates a new call builder for the [predictSafeAddress] function.

pub fn reclaimErc20( &self, token: Address, ) -> SolCallBuilder<&P, reclaimErc20Call, N>

Creates a new call builder for the [reclaimErc20] function.

pub fn renounceOwnership(&self) -> SolCallBuilder<&P, renounceOwnershipCall, N>

Creates a new call builder for the [renounceOwnership] function.

pub fn safeLibAddresses(&self) -> SolCallBuilder<&P, safeLibAddressesCall, N>

Creates a new call builder for the [safeLibAddresses] function.

pub fn safeVersion(&self) -> SolCallBuilder<&P, safeVersionCall, N>

Creates a new call builder for the [safeVersion] function.

pub fn tokensReceived( &self, _0: Address, from: Address, to: Address, amount: U256, userData: Bytes, _5: Bytes, ) -> SolCallBuilder<&P, tokensReceivedCall, N>

Creates a new call builder for the [tokensReceived] function.

pub fn transferOwnership( &self, newOwner: Address, ) -> SolCallBuilder<&P, transferOwnershipCall, N>

Creates a new call builder for the [transferOwnership] function.

pub fn tryGetDeployment( &self, safeProxyAddress: Address, ) -> SolCallBuilder<&P, tryGetDeploymentCall, N>

Creates a new call builder for the [tryGetDeployment] function.

pub fn updateErc1820Implementer( &self, _newImplementer: Address, ) -> SolCallBuilder<&P, updateErc1820ImplementerCall, N>

Creates a new call builder for the [updateErc1820Implementer] function.

pub fn updateHoprNetwork( &self, _newHoprNetwork: <HoprNetwork as SolType>::RustType, ) -> SolCallBuilder<&P, updateHoprNetworkCall, N>

Creates a new call builder for the [updateHoprNetwork] function.

pub fn updateModuleSingletonAddress( &self, _newModuleSingletonAddress: Address, ) -> SolCallBuilder<&P, updateModuleSingletonAddressCall, N>

Creates a new call builder for the [updateModuleSingletonAddress] function.

pub fn updateSafeLibAddress( &self, _newSafeLibAddresses: <SafeLibAddress as SolType>::RustType, ) -> SolCallBuilder<&P, updateSafeLibAddressCall, N>

Creates a new call builder for the [updateSafeLibAddress] function.

impl<P: Provider<N>, N: Network> HoprNodeStakeFactoryInstance<P, N>

Event filters.

pub fn event_filter<E: SolEvent>(&self) -> Event<&P, E, N>

Creates a new event filter using this contract instance’s provider and address.

Note that the type can be any event, not just those defined in this contract. Prefer using the other methods for building type-safe event filters.

pub fn HoprNodeStakeHoprNetworkUpdated_filter( &self, ) -> Event<&P, HoprNodeStakeHoprNetworkUpdated, N>

Creates a new event filter for the HoprNodeStakeHoprNetworkUpdated event.

pub fn HoprNodeStakeModuleUpdated_filter( &self, ) -> Event<&P, HoprNodeStakeModuleUpdated, N>

Creates a new event filter for the HoprNodeStakeModuleUpdated event.

pub fn HoprNodeStakeSafeLibUpdated_filter( &self, ) -> Event<&P, HoprNodeStakeSafeLibUpdated, N>

Creates a new event filter for the HoprNodeStakeSafeLibUpdated event.

pub fn NewHoprNodeStakeModule_filter( &self, ) -> Event<&P, NewHoprNodeStakeModule, N>

Creates a new event filter for the NewHoprNodeStakeModule event.

pub fn NewHoprNodeStakeModuleForSafe_filter( &self, ) -> Event<&P, NewHoprNodeStakeModuleForSafe, N>

Creates a new event filter for the NewHoprNodeStakeModuleForSafe event.

pub fn NewHoprNodeStakeSafe_filter(&self) -> Event<&P, NewHoprNodeStakeSafe, N>

Creates a new event filter for the NewHoprNodeStakeSafe event.

pub fn OwnershipTransferStarted_filter( &self, ) -> Event<&P, OwnershipTransferStarted, N>

Creates a new event filter for the OwnershipTransferStarted event.

pub fn OwnershipTransferred_filter(&self) -> Event<&P, OwnershipTransferred, N>

Creates a new event filter for the OwnershipTransferred event.