hopr_bindings::hopr_node_stake_factory_events

Module HoprNodeStakeFactoryEvents

Generated by the following Solidity interface…

library HoprNodeStakeFactory {
    struct HoprNetwork {
        address tokenAddress;
        uint256 defaultTokenAllowance;
        bytes32 defaultAnnouncementTarget;
    }
    struct SafeLibAddress {
        address safeAddress;
        address safeProxyFactoryAddress;
        address fallbackHandlerAddress;
        address multiSendAddress;
    }
}

interface HoprNodeStakeFactoryEvents {
    event HoprNodeStakeHoprNetworkUpdated(HoprNodeStakeFactory.HoprNetwork hoprNetwork);
    event HoprNodeStakeModuleUpdated(address moduleImplementation);
    event HoprNodeStakeSafeLibUpdated(HoprNodeStakeFactory.SafeLibAddress safeLibAddresses);
    event NewHoprNodeStakeModule(address instance);
    event NewHoprNodeStakeModuleForSafe(address module, address safe);
    event NewHoprNodeStakeSafe(address instance);
}

…which was generated by the following JSON ABI:

[
  {
    "type": "event",
    "name": "HoprNodeStakeHoprNetworkUpdated",
    "inputs": [
      {
        "name": "hoprNetwork",
        "type": "tuple",
        "indexed": false,
        "internalType": "struct HoprNodeStakeFactory.HoprNetwork",
        "components": [
          {
            "name": "tokenAddress",
            "type": "address",
            "internalType": "address"
          },
          {
            "name": "defaultTokenAllowance",
            "type": "uint256",
            "internalType": "uint256"
          },
          {
            "name": "defaultAnnouncementTarget",
            "type": "bytes32",
            "internalType": "bytes32"
          }
        ]
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "HoprNodeStakeModuleUpdated",
    "inputs": [
      {
        "name": "moduleImplementation",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "HoprNodeStakeSafeLibUpdated",
    "inputs": [
      {
        "name": "safeLibAddresses",
        "type": "tuple",
        "indexed": false,
        "internalType": "struct HoprNodeStakeFactory.SafeLibAddress",
        "components": [
          {
            "name": "safeAddress",
            "type": "address",
            "internalType": "address"
          },
          {
            "name": "safeProxyFactoryAddress",
            "type": "address",
            "internalType": "address"
          },
          {
            "name": "fallbackHandlerAddress",
            "type": "address",
            "internalType": "address"
          },
          {
            "name": "multiSendAddress",
            "type": "address",
            "internalType": "address"
          }
        ]
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "NewHoprNodeStakeModule",
    "inputs": [
      {
        "name": "instance",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "NewHoprNodeStakeModuleForSafe",
    "inputs": [
      {
        "name": "module",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      },
      {
        "name": "safe",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "NewHoprNodeStakeSafe",
    "inputs": [
      {
        "name": "instance",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  }
]

Structs

HoprNodeStakeFactoryEventsInstance

A HoprNodeStakeFactoryEvents instance.

HoprNodeStakeHoprNetworkUpdated

Event with signature HoprNodeStakeHoprNetworkUpdated((address,uint256,bytes32)) and selector 0x54cfcf4bdb6f20632ba544cd684130fc9db4dc2fe6360926d6b7866297f0315e.

HoprNodeStakeModuleUpdated

Event with signature HoprNodeStakeModuleUpdated(address) and selector 0xdd60725a88508fe2dcb769bd824280a2a01356198c795515d175abc4bc40b74e.

HoprNodeStakeSafeLibUpdated

Event with signature HoprNodeStakeSafeLibUpdated((address,address,address,address)) and selector 0x87eb507b2b42eb659807d7a1f63c89b8e22c674689603eac07a3713f06fec9c3.

NewHoprNodeStakeModule

Event with signature NewHoprNodeStakeModule(address) and selector 0xf966ce717349a940e403b8faae71fdaf573d8b9847deea16fca3d870cf6c292d.

NewHoprNodeStakeModuleForSafe

Event with signature NewHoprNodeStakeModuleForSafe(address,address) and selector 0x83a5c31bfee49d254d95e3946d8e8434e5975382fa8b849182db9e251aaaaa26.

NewHoprNodeStakeSafe

Event with signature NewHoprNodeStakeSafe(address) and selector 0x8231d169f416b666ae7fa43faa24a18899738075a53f32c97617d173b189e386.

Enums

HoprNodeStakeFactoryEventsEvents

Container for all the HoprNodeStakeFactoryEvents events.

Statics

BYTECODE

The creation / init bytecode of the contract.

DEPLOYED_BYTECODE

The runtime bytecode of the contract, as deployed on the network.

Functions

deploy

Deploys this contract using the given provider and constructor arguments, if any.

deploy_builder

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

new

Creates a new wrapper around an on-chain HoprNodeStakeFactoryEvents contract instance.