hopr_bindings

Module hopr_token

Modules

hopr_token

This module was auto-generated with ethers-rs Abigen. More information at: https://github.com/gakonst/ethers-rs

Structs

AccountSnapshotsCall

Container type for all input parameters for the accountSnapshots function with signature accountSnapshots(address,uint256) and selector 0x2497aee6

AccountSnapshotsReturn

Container type for all return fields from the accountSnapshots function with signature accountSnapshots(address,uint256) and selector 0x2497aee6

AllowanceCall

Container type for all input parameters for the allowance function with signature allowance(address,address) and selector 0xdd62ed3e

AllowanceReturn

Container type for all return fields from the allowance function with signature allowance(address,address) and selector 0xdd62ed3e

ApprovalFilter

ApproveCall

Container type for all input parameters for the approve function with signature approve(address,uint256) and selector 0x095ea7b3

ApproveReturn

Container type for all return fields from the approve function with signature approve(address,uint256) and selector 0x095ea7b3

AuthorizeOperatorCall

Container type for all input parameters for the authorizeOperator function with signature authorizeOperator(address) and selector 0x959b8c3f

AuthorizedOperatorFilter

BalanceOfAtCall

Container type for all input parameters for the balanceOfAt function with signature balanceOfAt(address,uint128) and selector 0xf772a092

BalanceOfAtReturn

Container type for all return fields from the balanceOfAt function with signature balanceOfAt(address,uint128) and selector 0xf772a092

BalanceOfCall

Container type for all input parameters for the balanceOf function with signature balanceOf(address) and selector 0x70a08231

BalanceOfReturn

Container type for all return fields from the balanceOf function with signature balanceOf(address) and selector 0x70a08231

BurnCall

Container type for all input parameters for the burn function with signature burn(uint256,bytes) and selector 0xfe9d9303

BurnedFilter

DecimalsCall

Container type for all input parameters for the decimals function with signature decimals() and selector 0x313ce567

DecimalsReturn

Container type for all return fields from the decimals function with signature decimals() and selector 0x313ce567

DefaultAdminRoleCall

Container type for all input parameters for the DEFAULT_ADMIN_ROLE function with signature DEFAULT_ADMIN_ROLE() and selector 0xa217fddf

DefaultAdminRoleReturn

Container type for all return fields from the DEFAULT_ADMIN_ROLE function with signature DEFAULT_ADMIN_ROLE() and selector 0xa217fddf

DefaultOperatorsCall

Container type for all input parameters for the defaultOperators function with signature defaultOperators() and selector 0x06e48538

DefaultOperatorsReturn

Container type for all return fields from the defaultOperators function with signature defaultOperators() and selector 0x06e48538

GetRoleAdminCall

Container type for all input parameters for the getRoleAdmin function with signature getRoleAdmin(bytes32) and selector 0x248a9ca3

GetRoleAdminReturn

Container type for all return fields from the getRoleAdmin function with signature getRoleAdmin(bytes32) and selector 0x248a9ca3

GetRoleMemberCall

Container type for all input parameters for the getRoleMember function with signature getRoleMember(bytes32,uint256) and selector 0x9010d07c

GetRoleMemberCountCall

Container type for all input parameters for the getRoleMemberCount function with signature getRoleMemberCount(bytes32) and selector 0xca15c873

GetRoleMemberCountReturn

Container type for all return fields from the getRoleMemberCount function with signature getRoleMemberCount(bytes32) and selector 0xca15c873

GetRoleMemberReturn

Container type for all return fields from the getRoleMember function with signature getRoleMember(bytes32,uint256) and selector 0x9010d07c

GrantRoleCall

Container type for all input parameters for the grantRole function with signature grantRole(bytes32,address) and selector 0x2f2ff15d

GranularityCall

Container type for all input parameters for the granularity function with signature granularity() and selector 0x556f0dc7

GranularityReturn

Container type for all return fields from the granularity function with signature granularity() and selector 0x556f0dc7

HasRoleCall

Container type for all input parameters for the hasRole function with signature hasRole(bytes32,address) and selector 0x91d14854

HasRoleReturn

Container type for all return fields from the hasRole function with signature hasRole(bytes32,address) and selector 0x91d14854

HoprToken

IsOperatorForCall

Container type for all input parameters for the isOperatorFor function with signature isOperatorFor(address,address) and selector 0xd95b6371

IsOperatorForReturn

Container type for all return fields from the isOperatorFor function with signature isOperatorFor(address,address) and selector 0xd95b6371

MintCall

Container type for all input parameters for the mint function with signature mint(address,uint256,bytes,bytes) and selector 0xdcdc7dd0

MintedFilter

MinterRoleCall

Container type for all input parameters for the MINTER_ROLE function with signature MINTER_ROLE() and selector 0xd5391393

MinterRoleReturn

Container type for all return fields from the MINTER_ROLE function with signature MINTER_ROLE() and selector 0xd5391393

NameCall

Container type for all input parameters for the name function with signature name() and selector 0x06fdde03

NameReturn

Container type for all return fields from the name function with signature name() and selector 0x06fdde03

OperatorBurnCall

Container type for all input parameters for the operatorBurn function with signature operatorBurn(address,uint256,bytes,bytes) and selector 0xfc673c4f

OperatorSendCall

Container type for all input parameters for the operatorSend function with signature operatorSend(address,address,uint256,bytes,bytes) and selector 0x62ad1b83

RenounceRoleCall

Container type for all input parameters for the renounceRole function with signature renounceRole(bytes32,address) and selector 0x36568abe

RevokeOperatorCall

Container type for all input parameters for the revokeOperator function with signature revokeOperator(address) and selector 0xfad8b32a

RevokeRoleCall

Container type for all input parameters for the revokeRole function with signature revokeRole(bytes32,address) and selector 0xd547741f

RevokedOperatorFilter

RoleAdminChangedFilter

RoleGrantedFilter

RoleRevokedFilter

SendCall

Container type for all input parameters for the send function with signature send(address,uint256,bytes) and selector 0x9bd9bbc6

SentFilter

SupportsInterfaceCall

Container type for all input parameters for the supportsInterface function with signature supportsInterface(bytes4) and selector 0x01ffc9a7

SupportsInterfaceReturn

Container type for all return fields from the supportsInterface function with signature supportsInterface(bytes4) and selector 0x01ffc9a7

SymbolCall

Container type for all input parameters for the symbol function with signature symbol() and selector 0x95d89b41

SymbolReturn

Container type for all return fields from the symbol function with signature symbol() and selector 0x95d89b41

TotalSupplyAtCall

Container type for all input parameters for the totalSupplyAt function with signature totalSupplyAt(uint128) and selector 0x947975d9

TotalSupplyAtReturn

Container type for all return fields from the totalSupplyAt function with signature totalSupplyAt(uint128) and selector 0x947975d9

TotalSupplyCall

Container type for all input parameters for the totalSupply function with signature totalSupply() and selector 0x18160ddd

TotalSupplyReturn

Container type for all return fields from the totalSupply function with signature totalSupply() and selector 0x18160ddd

TotalSupplySnapshotsCall

Container type for all input parameters for the totalSupplySnapshots function with signature totalSupplySnapshots(uint256) and selector 0xb7d78b1a

TotalSupplySnapshotsReturn

Container type for all return fields from the totalSupplySnapshots function with signature totalSupplySnapshots(uint256) and selector 0xb7d78b1a

TransferCall

Container type for all input parameters for the transfer function with signature transfer(address,uint256) and selector 0xa9059cbb

TransferFilter

TransferFromCall

Container type for all input parameters for the transferFrom function with signature transferFrom(address,address,uint256) and selector 0x23b872dd

TransferFromReturn

Container type for all return fields from the transferFrom function with signature transferFrom(address,address,uint256) and selector 0x23b872dd

TransferReturn

Container type for all return fields from the transfer function with signature transfer(address,uint256) and selector 0xa9059cbb

Enums

HoprTokenCalls

Container type for all of the contract’s call

HoprTokenEvents

Container type for all of the contract’s events

Statics

HOPRTOKEN_ABI

The parsed JSON ABI of the contract.

HOPRTOKEN_BYTECODE

The bytecode of the contract.

HOPRTOKEN_DEPLOYED_BYTECODE

The deployed bytecode of the contract.