hopr_bindings::hoprannouncements

Module HoprAnnouncements

Generated by the following Solidity interface…

interface HoprAnnouncements {
    error AlreadyInitialized();
    error ContractNotResponsible();
    error InvalidSafeAddress();
    error MultiSigUninitialized();

    event AddressAnnouncement(address node, string baseMultiaddr);
    event KeyBinding(bytes32 ed25519_sig_0, bytes32 ed25519_sig_1, bytes32 ed25519_pub_key, address chain_key);
    event RevokeAnnouncement(address node);

    constructor(address safeRegistry);

    function announce(string memory baseMultiaddr) external;
    function announceSafe(address selfAddress, string memory baseMultiaddr) external;
    function bindKeys(bytes32 ed25519_sig_0, bytes32 ed25519_sig_1, bytes32 ed25519_pub_key) external;
    function bindKeysAnnounce(bytes32 ed25519_sig_0, bytes32 ed25519_sig_1, bytes32 ed25519_pub_key, string memory baseMultiaddr) external;
    function bindKeysAnnounceSafe(address selfAddress, bytes32 ed25519_sig_0, bytes32 ed25519_sig_1, bytes32 ed25519_pub_key, string memory baseMultiaddr) external;
    function bindKeysSafe(address selfAddress, bytes32 ed25519_sig_0, bytes32 ed25519_sig_1, bytes32 ed25519_pub_key) external;
    function multicall(bytes[] memory data) external returns (bytes[] memory results);
    function revoke() external;
    function revokeSafe(address selfAddress) external;
}

…which was generated by the following JSON ABI:

[
  {
    "type": "constructor",
    "inputs": [
      {
        "name": "safeRegistry",
        "type": "address",
        "internalType": "contract HoprNodeSafeRegistry"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "announce",
    "inputs": [
      {
        "name": "baseMultiaddr",
        "type": "string",
        "internalType": "string"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "announceSafe",
    "inputs": [
      {
        "name": "selfAddress",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "baseMultiaddr",
        "type": "string",
        "internalType": "string"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "bindKeys",
    "inputs": [
      {
        "name": "ed25519_sig_0",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "ed25519_sig_1",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "ed25519_pub_key",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "bindKeysAnnounce",
    "inputs": [
      {
        "name": "ed25519_sig_0",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "ed25519_sig_1",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "ed25519_pub_key",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "baseMultiaddr",
        "type": "string",
        "internalType": "string"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "bindKeysAnnounceSafe",
    "inputs": [
      {
        "name": "selfAddress",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "ed25519_sig_0",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "ed25519_sig_1",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "ed25519_pub_key",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "baseMultiaddr",
        "type": "string",
        "internalType": "string"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "bindKeysSafe",
    "inputs": [
      {
        "name": "selfAddress",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "ed25519_sig_0",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "ed25519_sig_1",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "ed25519_pub_key",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "multicall",
    "inputs": [
      {
        "name": "data",
        "type": "bytes[]",
        "internalType": "bytes[]"
      }
    ],
    "outputs": [
      {
        "name": "results",
        "type": "bytes[]",
        "internalType": "bytes[]"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "revoke",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "revokeSafe",
    "inputs": [
      {
        "name": "selfAddress",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "event",
    "name": "AddressAnnouncement",
    "inputs": [
      {
        "name": "node",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      },
      {
        "name": "baseMultiaddr",
        "type": "string",
        "indexed": false,
        "internalType": "string"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "KeyBinding",
    "inputs": [
      {
        "name": "ed25519_sig_0",
        "type": "bytes32",
        "indexed": false,
        "internalType": "bytes32"
      },
      {
        "name": "ed25519_sig_1",
        "type": "bytes32",
        "indexed": false,
        "internalType": "bytes32"
      },
      {
        "name": "ed25519_pub_key",
        "type": "bytes32",
        "indexed": false,
        "internalType": "bytes32"
      },
      {
        "name": "chain_key",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "RevokeAnnouncement",
    "inputs": [
      {
        "name": "node",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "error",
    "name": "AlreadyInitialized",
    "inputs": []
  },
  {
    "type": "error",
    "name": "ContractNotResponsible",
    "inputs": []
  },
  {
    "type": "error",
    "name": "InvalidSafeAddress",
    "inputs": []
  },
  {
    "type": "error",
    "name": "MultiSigUninitialized",
    "inputs": []
  }
]

Structs

AddressAnnouncement

Event with signature AddressAnnouncement(address,string) and selector 0xc4df5ba16814838ab2618829d68f8623bb897302f24dbdba2279dbe45adb3d14.

AlreadyInitialized

Custom error with signature AlreadyInitialized() and selector 0x0dc149f0.

ContractNotResponsible

Custom error with signature ContractNotResponsible() and selector 0xacd5a823.

HoprAnnouncementsInstance

A HoprAnnouncements instance.

InvalidSafeAddress

Custom error with signature InvalidSafeAddress() and selector 0x8e9d7c5e.

KeyBinding

Event with signature KeyBinding(bytes32,bytes32,bytes32,address) and selector 0xbf49ccd41a4118c7e1e72143bad844e5f6ed12bf60caa642d88d19ca10244c36.

MultiSigUninitialized

Custom error with signature MultiSigUninitialized() and selector 0x454a20c8.

RevokeAnnouncement

Event with signature RevokeAnnouncement(address) and selector 0xa4de30a528becadf82649d1395c0e30dd18ae35b5a96ce71e9295bb14bc9f3bc.

announceCall

Function with signature announce(string) and selector 0xea0a5237.

announceReturn

Container type for the return parameters of the announce(string) function.

announceSafeCall

Function with signature announceSafe(address,string) and selector 0xfad0e5a2.

announceSafeReturn

Container type for the return parameters of the announceSafe(address,string) function.

bindKeysAnnounceCall

Function with signature bindKeysAnnounce(bytes32,bytes32,bytes32,string) and selector 0xf884a9cb.

bindKeysAnnounceReturn

Container type for the return parameters of the bindKeysAnnounce(bytes32,bytes32,bytes32,string) function.

bindKeysAnnounceSafeCall

Function with signature bindKeysAnnounceSafe(address,bytes32,bytes32,bytes32,string) and selector 0xddb9dfc1.

bindKeysAnnounceSafeReturn

Container type for the return parameters of the bindKeysAnnounceSafe(address,bytes32,bytes32,bytes32,string) function.

bindKeysCall

Function with signature bindKeys(bytes32,bytes32,bytes32) and selector 0xdbb98d91.

bindKeysReturn

Container type for the return parameters of the bindKeys(bytes32,bytes32,bytes32) function.

bindKeysSafeCall

Function with signature bindKeysSafe(address,bytes32,bytes32,bytes32) and selector 0xcae2b434.

bindKeysSafeReturn

Container type for the return parameters of the bindKeysSafe(address,bytes32,bytes32,bytes32) function.

constructorCall

Constructor`.

multicallCall

Function with signature multicall(bytes[]) and selector 0xac9650d8.

multicallReturn

Container type for the return parameters of the multicall(bytes[]) function.

revokeCall

Function with signature revoke() and selector 0xb6549f75.

revokeReturn

Container type for the return parameters of the revoke() function.

revokeSafeCall

Function with signature revokeSafe(address) and selector 0x308c712e.

revokeSafeReturn

Container type for the return parameters of the revokeSafe(address) function.

Enums

HoprAnnouncementsCalls

Container for all the HoprAnnouncements function calls.

HoprAnnouncementsErrors

Container for all the HoprAnnouncements custom errors.

HoprAnnouncementsEvents

Container for all the HoprAnnouncements events.

Statics

BYTECODE

The creation / init bytecode of the contract.

DEPLOYED_BYTECODE

The runtime bytecode of the contract, as deployed on the network.

Functions

deploy

Deploys this contract using the given provider and constructor arguments, if any.

deploy_builder

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

new

Creates a new wrapper around an on-chain HoprAnnouncements contract instance.