hopr_bindings::hoprannouncements::HoprAnnouncements

Struct HoprAnnouncementsInstance

pub struct HoprAnnouncementsInstance<P, N = Ethereum> { /* private fields */ }

A HoprAnnouncements instance.

Contains type-safe methods for interacting with an on-chain instance of the HoprAnnouncements contract located at a given address, using a given provider P.

If the contract bytecode is available (see the sol! documentation on how to provide it), the deploy and deploy_builder methods can be used to deploy a new instance of the contract.

See the module-level documentation for all the available methods.

Implementations

impl<P: Provider<N>, N: Network> HoprAnnouncementsInstance<P, N>

Instantiation and getters/setters.

pub const fn new(address: Address, provider: P) -> Self

Creates a new wrapper around an on-chain HoprAnnouncements contract instance.

See the wrapper’s documentation for more details.

pub async fn deploy( provider: P, safeRegistry: Address, ) -> Result<HoprAnnouncementsInstance<P, N>>

Deploys this contract using the given provider and constructor arguments, if any.

Returns a new instance of the contract, if the deployment was successful.

For more fine-grained control over the deployment process, use deploy_builder instead.

pub fn deploy_builder( provider: P, safeRegistry: Address, ) -> RawCallBuilder<P, N>

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

This is a simple wrapper around creating a RawCallBuilder with the data set to the bytecode concatenated with the constructor’s ABI-encoded arguments.

pub const fn address(&self) -> &Address

Returns a reference to the address.

pub fn set_address(&mut self, address: Address)

Sets the address.

pub fn at(self, address: Address) -> Self

Sets the address and returns self.

pub const fn provider(&self) -> &P

Returns a reference to the provider.

impl<P: Clone, N> HoprAnnouncementsInstance<&P, N>

pub fn with_cloned_provider(self) -> HoprAnnouncementsInstance<P, N>

Clones the provider and returns a new instance with the cloned provider.

impl<P: Provider<N>, N: Network> HoprAnnouncementsInstance<P, N>

Function calls.

pub fn call_builder<C: SolCall>(&self, call: &C) -> SolCallBuilder<&P, C, N>

Creates a new call builder using this contract instance’s provider and address.

Note that the call can be any function call, not just those defined in this contract. Prefer using the other methods for building type-safe contract calls.

pub fn announce( &self, baseMultiaddr: String, ) -> SolCallBuilder<&P, announceCall, N>

Creates a new call builder for the [announce] function.

pub fn announceSafe( &self, selfAddress: Address, baseMultiaddr: String, ) -> SolCallBuilder<&P, announceSafeCall, N>

Creates a new call builder for the [announceSafe] function.

pub fn bindKeys( &self, ed25519_sig_0: FixedBytes<32>, ed25519_sig_1: FixedBytes<32>, ed25519_pub_key: FixedBytes<32>, ) -> SolCallBuilder<&P, bindKeysCall, N>

Creates a new call builder for the [bindKeys] function.

pub fn bindKeysAnnounce( &self, ed25519_sig_0: FixedBytes<32>, ed25519_sig_1: FixedBytes<32>, ed25519_pub_key: FixedBytes<32>, baseMultiaddr: String, ) -> SolCallBuilder<&P, bindKeysAnnounceCall, N>

Creates a new call builder for the [bindKeysAnnounce] function.

pub fn bindKeysAnnounceSafe( &self, selfAddress: Address, ed25519_sig_0: FixedBytes<32>, ed25519_sig_1: FixedBytes<32>, ed25519_pub_key: FixedBytes<32>, baseMultiaddr: String, ) -> SolCallBuilder<&P, bindKeysAnnounceSafeCall, N>

Creates a new call builder for the [bindKeysAnnounceSafe] function.

pub fn bindKeysSafe( &self, selfAddress: Address, ed25519_sig_0: FixedBytes<32>, ed25519_sig_1: FixedBytes<32>, ed25519_pub_key: FixedBytes<32>, ) -> SolCallBuilder<&P, bindKeysSafeCall, N>

Creates a new call builder for the [bindKeysSafe] function.

pub fn multicall( &self, data: Vec<Bytes>, ) -> SolCallBuilder<&P, multicallCall, N>

Creates a new call builder for the [multicall] function.

pub fn revoke(&self) -> SolCallBuilder<&P, revokeCall, N>

Creates a new call builder for the [revoke] function.

pub fn revokeSafe( &self, selfAddress: Address, ) -> SolCallBuilder<&P, revokeSafeCall, N>

Creates a new call builder for the [revokeSafe] function.

impl<P: Provider<N>, N: Network> HoprAnnouncementsInstance<P, N>

Event filters.

pub fn event_filter<E: SolEvent>(&self) -> Event<&P, E, N>

Creates a new event filter using this contract instance’s provider and address.

Note that the type can be any event, not just those defined in this contract. Prefer using the other methods for building type-safe event filters.

pub fn AddressAnnouncement_filter(&self) -> Event<&P, AddressAnnouncement, N>

Creates a new event filter for the AddressAnnouncement event.

pub fn KeyBinding_filter(&self) -> Event<&P, KeyBinding, N>

Creates a new event filter for the KeyBinding event.

pub fn RevokeAnnouncement_filter(&self) -> Event<&P, RevokeAnnouncement, N>

Creates a new event filter for the RevokeAnnouncement event.