hopr_bindings::hoprboost

Module HoprBoost

Generated by the following Solidity interface…

interface HoprBoost {
    event Approval(address indexed owner, address indexed approved, uint256 indexed tokenId);
    event ApprovalForAll(address indexed owner, address indexed operator, bool approved);
    event BoostMinted(uint256 indexed boostTypeIndex, uint256 indexed boostNumerator, uint256 indexed redeemDeadline);
    event RoleAdminChanged(bytes32 indexed role, bytes32 indexed previousAdminRole, bytes32 indexed newAdminRole);
    event RoleGranted(bytes32 indexed role, address indexed account, address indexed sender);
    event RoleRevoked(bytes32 indexed role, address indexed account, address indexed sender);
    event Transfer(address indexed from, address indexed to, uint256 indexed tokenId);

    constructor(address newAdmin, string baseTokenURI);

    function DEFAULT_ADMIN_ROLE() external view returns (bytes32);
    function MINTER_ROLE() external view returns (bytes32);
    function approve(address to, uint256 tokenId) external;
    function balanceOf(address owner) external view returns (uint256);
    function batchMint(address[] memory to, string memory boostType, string memory boostRank, uint256 boostNumerator, uint256 redeemDeadline) external;
    function boostOf(uint256 tokenId) external view returns (uint256, uint256);
    function getApproved(uint256 tokenId) external view returns (address);
    function getRoleAdmin(bytes32 role) external view returns (bytes32);
    function getRoleMember(bytes32 role, uint256 index) external view returns (address);
    function getRoleMemberCount(bytes32 role) external view returns (uint256);
    function grantRole(bytes32 role, address account) external;
    function hasRole(bytes32 role, address account) external view returns (bool);
    function isApprovedForAll(address owner, address operator) external view returns (bool);
    function mint(address to, string memory boostType, string memory boostRank, uint256 boostNumerator, uint256 redeemDeadline) external;
    function name() external view returns (string memory);
    function ownerOf(uint256 tokenId) external view returns (address);
    function reclaimErc20Tokens(address tokenAddress) external;
    function reclaimErc721Tokens(address tokenAddress, uint256 tokenId) external;
    function renounceRole(bytes32 role, address account) external;
    function revokeRole(bytes32 role, address account) external;
    function safeTransferFrom(address from, address to, uint256 tokenId) external;
    function safeTransferFrom(address from, address to, uint256 tokenId, bytes memory _data) external;
    function setApprovalForAll(address operator, bool approved) external;
    function supportsInterface(bytes4 interfaceId) external view returns (bool);
    function symbol() external view returns (string memory);
    function tokenByIndex(uint256 index) external view returns (uint256);
    function tokenOfOwnerByIndex(address owner, uint256 index) external view returns (uint256);
    function tokenURI(uint256 tokenId) external view returns (string memory);
    function totalSupply() external view returns (uint256);
    function transferFrom(address from, address to, uint256 tokenId) external;
    function typeAt(uint256 typeIndex) external view returns (string memory);
    function typeIndexOf(uint256 tokenId) external view returns (uint256);
    function typeOf(uint256 tokenId) external view returns (string memory);
    function updateBaseURI(string memory baseTokenURI) external;
}

…which was generated by the following JSON ABI:

[
  {
    "type": "constructor",
    "inputs": [
      {
        "name": "newAdmin",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "baseTokenURI",
        "type": "string",
        "internalType": "string"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "DEFAULT_ADMIN_ROLE",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "MINTER_ROLE",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "approve",
    "inputs": [
      {
        "name": "to",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "tokenId",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "balanceOf",
    "inputs": [
      {
        "name": "owner",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "batchMint",
    "inputs": [
      {
        "name": "to",
        "type": "address[]",
        "internalType": "address[]"
      },
      {
        "name": "boostType",
        "type": "string",
        "internalType": "string"
      },
      {
        "name": "boostRank",
        "type": "string",
        "internalType": "string"
      },
      {
        "name": "boostNumerator",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "redeemDeadline",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "boostOf",
    "inputs": [
      {
        "name": "tokenId",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getApproved",
    "inputs": [
      {
        "name": "tokenId",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getRoleAdmin",
    "inputs": [
      {
        "name": "role",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getRoleMember",
    "inputs": [
      {
        "name": "role",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "index",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getRoleMemberCount",
    "inputs": [
      {
        "name": "role",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "grantRole",
    "inputs": [
      {
        "name": "role",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "account",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "hasRole",
    "inputs": [
      {
        "name": "role",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "account",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "isApprovedForAll",
    "inputs": [
      {
        "name": "owner",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "operator",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "mint",
    "inputs": [
      {
        "name": "to",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "boostType",
        "type": "string",
        "internalType": "string"
      },
      {
        "name": "boostRank",
        "type": "string",
        "internalType": "string"
      },
      {
        "name": "boostNumerator",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "redeemDeadline",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "name",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "string",
        "internalType": "string"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "ownerOf",
    "inputs": [
      {
        "name": "tokenId",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "reclaimErc20Tokens",
    "inputs": [
      {
        "name": "tokenAddress",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "reclaimErc721Tokens",
    "inputs": [
      {
        "name": "tokenAddress",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "tokenId",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "renounceRole",
    "inputs": [
      {
        "name": "role",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "account",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "revokeRole",
    "inputs": [
      {
        "name": "role",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "account",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "safeTransferFrom",
    "inputs": [
      {
        "name": "from",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "to",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "tokenId",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "safeTransferFrom",
    "inputs": [
      {
        "name": "from",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "to",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "tokenId",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "_data",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "setApprovalForAll",
    "inputs": [
      {
        "name": "operator",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "approved",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "supportsInterface",
    "inputs": [
      {
        "name": "interfaceId",
        "type": "bytes4",
        "internalType": "bytes4"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "symbol",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "string",
        "internalType": "string"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "tokenByIndex",
    "inputs": [
      {
        "name": "index",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "tokenOfOwnerByIndex",
    "inputs": [
      {
        "name": "owner",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "index",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "tokenURI",
    "inputs": [
      {
        "name": "tokenId",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "string",
        "internalType": "string"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "totalSupply",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "transferFrom",
    "inputs": [
      {
        "name": "from",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "to",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "tokenId",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "typeAt",
    "inputs": [
      {
        "name": "typeIndex",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "string",
        "internalType": "string"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "typeIndexOf",
    "inputs": [
      {
        "name": "tokenId",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "typeOf",
    "inputs": [
      {
        "name": "tokenId",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "string",
        "internalType": "string"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "updateBaseURI",
    "inputs": [
      {
        "name": "baseTokenURI",
        "type": "string",
        "internalType": "string"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "event",
    "name": "Approval",
    "inputs": [
      {
        "name": "owner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "approved",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "tokenId",
        "type": "uint256",
        "indexed": true,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "ApprovalForAll",
    "inputs": [
      {
        "name": "owner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "operator",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "approved",
        "type": "bool",
        "indexed": false,
        "internalType": "bool"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "BoostMinted",
    "inputs": [
      {
        "name": "boostTypeIndex",
        "type": "uint256",
        "indexed": true,
        "internalType": "uint256"
      },
      {
        "name": "boostNumerator",
        "type": "uint256",
        "indexed": true,
        "internalType": "uint256"
      },
      {
        "name": "redeemDeadline",
        "type": "uint256",
        "indexed": true,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "RoleAdminChanged",
    "inputs": [
      {
        "name": "role",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      },
      {
        "name": "previousAdminRole",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      },
      {
        "name": "newAdminRole",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "RoleGranted",
    "inputs": [
      {
        "name": "role",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      },
      {
        "name": "account",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "sender",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "RoleRevoked",
    "inputs": [
      {
        "name": "role",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      },
      {
        "name": "account",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "sender",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Transfer",
    "inputs": [
      {
        "name": "from",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "to",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "tokenId",
        "type": "uint256",
        "indexed": true,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  }
]

Structs

Approval

Event with signature Approval(address,address,uint256) and selector 0x8c5be1e5ebec7d5bd14f71427d1e84f3dd0314c0f7b2291e5b200ac8c7c3b925.

ApprovalForAll

Event with signature ApprovalForAll(address,address,bool) and selector 0x17307eab39ab6107e8899845ad3d59bd9653f200f220920489ca2b5937696c31.

BoostMinted

Event with signature BoostMinted(uint256,uint256,uint256) and selector 0x735e5eb5afdd72cc64657ded7e5de390af3708afb5b9a80b0e5c911a0499d98f.

DEFAULT_ADMIN_ROLECall

Function with signature DEFAULT_ADMIN_ROLE() and selector 0xa217fddf.

DEFAULT_ADMIN_ROLEReturn

Container type for the return parameters of the DEFAULT_ADMIN_ROLE() function.

HoprBoostInstance

A HoprBoost instance.

MINTER_ROLECall

Function with signature MINTER_ROLE() and selector 0xd5391393.

MINTER_ROLEReturn

Container type for the return parameters of the MINTER_ROLE() function.

RoleAdminChanged

Event with signature RoleAdminChanged(bytes32,bytes32,bytes32) and selector 0xbd79b86ffe0ab8e8776151514217cd7cacd52c909f66475c3af44e129f0b00ff.

RoleGranted

Event with signature RoleGranted(bytes32,address,address) and selector 0x2f8788117e7eff1d82e926ec794901d17c78024a50270940304540a733656f0d.

RoleRevoked

Event with signature RoleRevoked(bytes32,address,address) and selector 0xf6391f5c32d9c69d2a47ea670b442974b53935d1edc7fd64eb21e047a839171b.

Transfer

Event with signature Transfer(address,address,uint256) and selector 0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef.

approveCall

Function with signature approve(address,uint256) and selector 0x095ea7b3.

approveReturn

Container type for the return parameters of the approve(address,uint256) function.

balanceOfCall

Function with signature balanceOf(address) and selector 0x70a08231.

balanceOfReturn

Container type for the return parameters of the balanceOf(address) function.

batchMintCall

Function with signature batchMint(address[],string,string,uint256,uint256) and selector 0x9365c533.

batchMintReturn

Container type for the return parameters of the batchMint(address[],string,string,uint256,uint256) function.

boostOfCall

Function with signature boostOf(uint256) and selector 0x896cddf4.

boostOfReturn

Container type for the return parameters of the boostOf(uint256) function.

constructorCall

Constructor`.

getApprovedCall

Function with signature getApproved(uint256) and selector 0x081812fc.

getApprovedReturn

Container type for the return parameters of the getApproved(uint256) function.

getRoleAdminCall

Function with signature getRoleAdmin(bytes32) and selector 0x248a9ca3.

getRoleAdminReturn

Container type for the return parameters of the getRoleAdmin(bytes32) function.

getRoleMemberCall

Function with signature getRoleMember(bytes32,uint256) and selector 0x9010d07c.

getRoleMemberCountCall

Function with signature getRoleMemberCount(bytes32) and selector 0xca15c873.

getRoleMemberCountReturn

Container type for the return parameters of the getRoleMemberCount(bytes32) function.

getRoleMemberReturn

Container type for the return parameters of the getRoleMember(bytes32,uint256) function.

grantRoleCall

Function with signature grantRole(bytes32,address) and selector 0x2f2ff15d.

grantRoleReturn

Container type for the return parameters of the grantRole(bytes32,address) function.

hasRoleCall

Function with signature hasRole(bytes32,address) and selector 0x91d14854.

hasRoleReturn

Container type for the return parameters of the hasRole(bytes32,address) function.

isApprovedForAllCall

Function with signature isApprovedForAll(address,address) and selector 0xe985e9c5.

isApprovedForAllReturn

Container type for the return parameters of the isApprovedForAll(address,address) function.

mintCall

Function with signature mint(address,string,string,uint256,uint256) and selector 0x29867dac.

mintReturn

Container type for the return parameters of the mint(address,string,string,uint256,uint256) function.

nameCall

Function with signature name() and selector 0x06fdde03.

nameReturn

Container type for the return parameters of the name() function.

ownerOfCall

Function with signature ownerOf(uint256) and selector 0x6352211e.

ownerOfReturn

Container type for the return parameters of the ownerOf(uint256) function.

reclaimErc20TokensCall

Function with signature reclaimErc20Tokens(address) and selector 0x6067bc15.

reclaimErc20TokensReturn

Container type for the return parameters of the reclaimErc20Tokens(address) function.

reclaimErc721TokensCall

Function with signature reclaimErc721Tokens(address,uint256) and selector 0x48c64e41.

reclaimErc721TokensReturn

Container type for the return parameters of the reclaimErc721Tokens(address,uint256) function.

renounceRoleCall

Function with signature renounceRole(bytes32,address) and selector 0x36568abe.

renounceRoleReturn

Container type for the return parameters of the renounceRole(bytes32,address) function.

revokeRoleCall

Function with signature revokeRole(bytes32,address) and selector 0xd547741f.

revokeRoleReturn

Container type for the return parameters of the revokeRole(bytes32,address) function.

safeTransferFrom_0Call

Function with signature safeTransferFrom(address,address,uint256) and selector 0x42842e0e.

safeTransferFrom_0Return

Container type for the return parameters of the safeTransferFrom(address,address,uint256) function.

safeTransferFrom_1Call

Function with signature safeTransferFrom(address,address,uint256,bytes) and selector 0xb88d4fde.

safeTransferFrom_1Return

Container type for the return parameters of the safeTransferFrom(address,address,uint256,bytes) function.

setApprovalForAllCall

Function with signature setApprovalForAll(address,bool) and selector 0xa22cb465.

setApprovalForAllReturn

Container type for the return parameters of the setApprovalForAll(address,bool) function.

supportsInterfaceCall

Function with signature supportsInterface(bytes4) and selector 0x01ffc9a7.

supportsInterfaceReturn

Container type for the return parameters of the supportsInterface(bytes4) function.

symbolCall

Function with signature symbol() and selector 0x95d89b41.

symbolReturn

Container type for the return parameters of the symbol() function.

tokenByIndexCall

Function with signature tokenByIndex(uint256) and selector 0x4f6ccce7.

tokenByIndexReturn

Container type for the return parameters of the tokenByIndex(uint256) function.

tokenOfOwnerByIndexCall

Function with signature tokenOfOwnerByIndex(address,uint256) and selector 0x2f745c59.

tokenOfOwnerByIndexReturn

Container type for the return parameters of the tokenOfOwnerByIndex(address,uint256) function.

tokenURICall

Function with signature tokenURI(uint256) and selector 0xc87b56dd.

tokenURIReturn

Container type for the return parameters of the tokenURI(uint256) function.

totalSupplyCall

Function with signature totalSupply() and selector 0x18160ddd.

totalSupplyReturn

Container type for the return parameters of the totalSupply() function.

transferFromCall

Function with signature transferFrom(address,address,uint256) and selector 0x23b872dd.

transferFromReturn

Container type for the return parameters of the transferFrom(address,address,uint256) function.

typeAtCall

Function with signature typeAt(uint256) and selector 0xb5185a88.

typeAtReturn

Container type for the return parameters of the typeAt(uint256) function.

typeIndexOfCall

Function with signature typeIndexOf(uint256) and selector 0x562317c5.

typeIndexOfReturn

Container type for the return parameters of the typeIndexOf(uint256) function.

typeOfCall

Function with signature typeOf(uint256) and selector 0xc588ff8b.

typeOfReturn

Container type for the return parameters of the typeOf(uint256) function.

updateBaseURICall

Function with signature updateBaseURI(string) and selector 0x931688cb.

updateBaseURIReturn

Container type for the return parameters of the updateBaseURI(string) function.

Enums

HoprBoostCalls

Container for all the HoprBoost function calls.

HoprBoostEvents

Container for all the HoprBoost events.

Statics

BYTECODE

The creation / init bytecode of the contract.

DEPLOYED_BYTECODE

The runtime bytecode of the contract, as deployed on the network.

Functions

deploy

Deploys this contract using the given provider and constructor arguments, if any.

deploy_builder

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

new

Creates a new wrapper around an on-chain HoprBoost contract instance.