hopr_bindings::hoprboost::HoprBoost

Struct HoprBoostInstance

pub struct HoprBoostInstance<P, N = Ethereum> { /* private fields */ }

A HoprBoost instance.

Contains type-safe methods for interacting with an on-chain instance of the HoprBoost contract located at a given address, using a given provider P.

If the contract bytecode is available (see the sol! documentation on how to provide it), the deploy and deploy_builder methods can be used to deploy a new instance of the contract.

See the module-level documentation for all the available methods.

Implementations

impl<P: Provider<N>, N: Network> HoprBoostInstance<P, N>

Instantiation and getters/setters.

pub const fn new(address: Address, provider: P) -> Self

Creates a new wrapper around an on-chain HoprBoost contract instance.

See the wrapper’s documentation for more details.

pub async fn deploy( provider: P, newAdmin: Address, baseTokenURI: String, ) -> Result<HoprBoostInstance<P, N>>

Deploys this contract using the given provider and constructor arguments, if any.

Returns a new instance of the contract, if the deployment was successful.

For more fine-grained control over the deployment process, use deploy_builder instead.

pub fn deploy_builder( provider: P, newAdmin: Address, baseTokenURI: String, ) -> RawCallBuilder<P, N>

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

This is a simple wrapper around creating a RawCallBuilder with the data set to the bytecode concatenated with the constructor’s ABI-encoded arguments.

pub const fn address(&self) -> &Address

Returns a reference to the address.

pub fn set_address(&mut self, address: Address)

Sets the address.

pub fn at(self, address: Address) -> Self

Sets the address and returns self.

pub const fn provider(&self) -> &P

Returns a reference to the provider.

impl<P: Clone, N> HoprBoostInstance<&P, N>

pub fn with_cloned_provider(self) -> HoprBoostInstance<P, N>

Clones the provider and returns a new instance with the cloned provider.

impl<P: Provider<N>, N: Network> HoprBoostInstance<P, N>

Function calls.

pub fn call_builder<C: SolCall>(&self, call: &C) -> SolCallBuilder<&P, C, N>

Creates a new call builder using this contract instance’s provider and address.

Note that the call can be any function call, not just those defined in this contract. Prefer using the other methods for building type-safe contract calls.

pub fn DEFAULT_ADMIN_ROLE( &self, ) -> SolCallBuilder<&P, DEFAULT_ADMIN_ROLECall, N>

Creates a new call builder for the [DEFAULT_ADMIN_ROLE] function.

pub fn MINTER_ROLE(&self) -> SolCallBuilder<&P, MINTER_ROLECall, N>

Creates a new call builder for the [MINTER_ROLE] function.

pub fn approve( &self, to: Address, tokenId: U256, ) -> SolCallBuilder<&P, approveCall, N>

Creates a new call builder for the [approve] function.

pub fn balanceOf(&self, owner: Address) -> SolCallBuilder<&P, balanceOfCall, N>

Creates a new call builder for the [balanceOf] function.

pub fn batchMint( &self, to: Vec<Address>, boostType: String, boostRank: String, boostNumerator: U256, redeemDeadline: U256, ) -> SolCallBuilder<&P, batchMintCall, N>

Creates a new call builder for the [batchMint] function.

pub fn boostOf(&self, tokenId: U256) -> SolCallBuilder<&P, boostOfCall, N>

Creates a new call builder for the [boostOf] function.

pub fn getApproved( &self, tokenId: U256, ) -> SolCallBuilder<&P, getApprovedCall, N>

Creates a new call builder for the [getApproved] function.

pub fn getRoleAdmin( &self, role: FixedBytes<32>, ) -> SolCallBuilder<&P, getRoleAdminCall, N>

Creates a new call builder for the [getRoleAdmin] function.

pub fn getRoleMember( &self, role: FixedBytes<32>, index: U256, ) -> SolCallBuilder<&P, getRoleMemberCall, N>

Creates a new call builder for the [getRoleMember] function.

pub fn getRoleMemberCount( &self, role: FixedBytes<32>, ) -> SolCallBuilder<&P, getRoleMemberCountCall, N>

Creates a new call builder for the [getRoleMemberCount] function.

pub fn grantRole( &self, role: FixedBytes<32>, account: Address, ) -> SolCallBuilder<&P, grantRoleCall, N>

Creates a new call builder for the [grantRole] function.

pub fn hasRole( &self, role: FixedBytes<32>, account: Address, ) -> SolCallBuilder<&P, hasRoleCall, N>

Creates a new call builder for the [hasRole] function.

pub fn isApprovedForAll( &self, owner: Address, operator: Address, ) -> SolCallBuilder<&P, isApprovedForAllCall, N>

Creates a new call builder for the [isApprovedForAll] function.

pub fn mint( &self, to: Address, boostType: String, boostRank: String, boostNumerator: U256, redeemDeadline: U256, ) -> SolCallBuilder<&P, mintCall, N>

Creates a new call builder for the [mint] function.

pub fn name(&self) -> SolCallBuilder<&P, nameCall, N>

Creates a new call builder for the [name] function.

pub fn ownerOf(&self, tokenId: U256) -> SolCallBuilder<&P, ownerOfCall, N>

Creates a new call builder for the [ownerOf] function.

pub fn reclaimErc20Tokens( &self, tokenAddress: Address, ) -> SolCallBuilder<&P, reclaimErc20TokensCall, N>

Creates a new call builder for the [reclaimErc20Tokens] function.

pub fn reclaimErc721Tokens( &self, tokenAddress: Address, tokenId: U256, ) -> SolCallBuilder<&P, reclaimErc721TokensCall, N>

Creates a new call builder for the [reclaimErc721Tokens] function.

pub fn renounceRole( &self, role: FixedBytes<32>, account: Address, ) -> SolCallBuilder<&P, renounceRoleCall, N>

Creates a new call builder for the [renounceRole] function.

pub fn revokeRole( &self, role: FixedBytes<32>, account: Address, ) -> SolCallBuilder<&P, revokeRoleCall, N>

Creates a new call builder for the [revokeRole] function.

pub fn safeTransferFrom_0( &self, from: Address, to: Address, tokenId: U256, ) -> SolCallBuilder<&P, safeTransferFrom_0Call, N>

Creates a new call builder for the [safeTransferFrom_0] function.

pub fn safeTransferFrom_1( &self, from: Address, to: Address, tokenId: U256, _data: Bytes, ) -> SolCallBuilder<&P, safeTransferFrom_1Call, N>

Creates a new call builder for the [safeTransferFrom_1] function.

pub fn setApprovalForAll( &self, operator: Address, approved: bool, ) -> SolCallBuilder<&P, setApprovalForAllCall, N>

Creates a new call builder for the [setApprovalForAll] function.

pub fn supportsInterface( &self, interfaceId: FixedBytes<4>, ) -> SolCallBuilder<&P, supportsInterfaceCall, N>

Creates a new call builder for the [supportsInterface] function.

pub fn symbol(&self) -> SolCallBuilder<&P, symbolCall, N>

Creates a new call builder for the [symbol] function.

pub fn tokenByIndex( &self, index: U256, ) -> SolCallBuilder<&P, tokenByIndexCall, N>

Creates a new call builder for the [tokenByIndex] function.

pub fn tokenOfOwnerByIndex( &self, owner: Address, index: U256, ) -> SolCallBuilder<&P, tokenOfOwnerByIndexCall, N>

Creates a new call builder for the [tokenOfOwnerByIndex] function.

pub fn tokenURI(&self, tokenId: U256) -> SolCallBuilder<&P, tokenURICall, N>

Creates a new call builder for the [tokenURI] function.

pub fn totalSupply(&self) -> SolCallBuilder<&P, totalSupplyCall, N>

Creates a new call builder for the [totalSupply] function.

pub fn transferFrom( &self, from: Address, to: Address, tokenId: U256, ) -> SolCallBuilder<&P, transferFromCall, N>

Creates a new call builder for the [transferFrom] function.

pub fn typeAt(&self, typeIndex: U256) -> SolCallBuilder<&P, typeAtCall, N>

Creates a new call builder for the [typeAt] function.

pub fn typeIndexOf( &self, tokenId: U256, ) -> SolCallBuilder<&P, typeIndexOfCall, N>

Creates a new call builder for the [typeIndexOf] function.

pub fn typeOf(&self, tokenId: U256) -> SolCallBuilder<&P, typeOfCall, N>

Creates a new call builder for the [typeOf] function.

pub fn updateBaseURI( &self, baseTokenURI: String, ) -> SolCallBuilder<&P, updateBaseURICall, N>

Creates a new call builder for the [updateBaseURI] function.

impl<P: Provider<N>, N: Network> HoprBoostInstance<P, N>

Event filters.

pub fn event_filter<E: SolEvent>(&self) -> Event<&P, E, N>

Creates a new event filter using this contract instance’s provider and address.

Note that the type can be any event, not just those defined in this contract. Prefer using the other methods for building type-safe event filters.

pub fn Approval_filter(&self) -> Event<&P, Approval, N>

Creates a new event filter for the Approval event.

pub fn ApprovalForAll_filter(&self) -> Event<&P, ApprovalForAll, N>

Creates a new event filter for the ApprovalForAll event.

pub fn BoostMinted_filter(&self) -> Event<&P, BoostMinted, N>

Creates a new event filter for the BoostMinted event.

pub fn RoleAdminChanged_filter(&self) -> Event<&P, RoleAdminChanged, N>

Creates a new event filter for the RoleAdminChanged event.

pub fn RoleGranted_filter(&self) -> Event<&P, RoleGranted, N>

Creates a new event filter for the RoleGranted event.

pub fn RoleRevoked_filter(&self) -> Event<&P, RoleRevoked, N>

Creates a new event filter for the RoleRevoked event.

pub fn Transfer_filter(&self) -> Event<&P, Transfer, N>

Creates a new event filter for the Transfer event.