hopr_bindings::hoprcapabilitypermissions::HoprCapabilityPermissions

Enum HoprCapabilityPermissionsCalls

pub enum HoprCapabilityPermissionsCalls {
    APPROVE_SELECTOR(APPROVE_SELECTORCall),
    CLOSE_INCOMING_CHANNEL_SELECTOR(CLOSE_INCOMING_CHANNEL_SELECTORCall),
    FINALIZE_OUTGOING_CHANNEL_CLOSURE_SELECTOR(FINALIZE_OUTGOING_CHANNEL_CLOSURE_SELECTORCall),
    FUND_CHANNEL_SELECTOR(FUND_CHANNEL_SELECTORCall),
    INITIATE_OUTGOING_CHANNEL_CLOSURE_SELECTOR(INITIATE_OUTGOING_CHANNEL_CLOSURE_SELECTORCall),
    REDEEM_TICKET_SELECTOR(REDEEM_TICKET_SELECTORCall),
    SEND_SELECTOR(SEND_SELECTORCall),
}

Container for all the HoprCapabilityPermissions function calls.

Variants

APPROVE_SELECTOR(APPROVE_SELECTORCall)

CLOSE_INCOMING_CHANNEL_SELECTOR(CLOSE_INCOMING_CHANNEL_SELECTORCall)

FINALIZE_OUTGOING_CHANNEL_CLOSURE_SELECTOR(FINALIZE_OUTGOING_CHANNEL_CLOSURE_SELECTORCall)

FUND_CHANNEL_SELECTOR(FUND_CHANNEL_SELECTORCall)

INITIATE_OUTGOING_CHANNEL_CLOSURE_SELECTOR(INITIATE_OUTGOING_CHANNEL_CLOSURE_SELECTORCall)

REDEEM_TICKET_SELECTOR(REDEEM_TICKET_SELECTORCall)

SEND_SELECTOR(SEND_SELECTORCall)

Implementations

impl HoprCapabilityPermissionsCalls

pub const SELECTORS: &'static [[u8; 4]]

All the selectors of this enum.

Note that the selectors might not be in the same order as the variants. No guarantees are made about the order of the selectors.

Prefer using SolInterface methods instead.