hopr_bindings::hoprcapabilitypermissions::HoprCapabilityPermissions

Enum HoprCapabilityPermissionsErrors

pub enum HoprCapabilityPermissionsErrors {

    AddressIsZero(AddressIsZero),
    ArrayTooLong(ArrayTooLong),
    ArraysDifferentLength(ArraysDifferentLength),
    CalldataOutOfBounds(CalldataOutOfBounds),
    DefaultPermissionRejected(DefaultPermissionRejected),
    DelegateCallNotAllowed(DelegateCallNotAllowed),
    FunctionSignatureTooShort(FunctionSignatureTooShort),
    GranularPermissionRejected(GranularPermissionRejected),
    NoMembership(NoMembership),
    NodePermissionRejected(NodePermissionRejected),
    ParameterNotAllowed(ParameterNotAllowed),
    PermissionNotConfigured(PermissionNotConfigured),
    SendNotAllowed(SendNotAllowed),
    TargetAddressNotAllowed(TargetAddressNotAllowed),
    TargetIsNotScoped(TargetIsNotScoped),
    TargetIsScoped(TargetIsScoped),
    UnacceptableMultiSendOffset(UnacceptableMultiSendOffset),

}

Container for all the HoprCapabilityPermissions custom errors.

Variants

AddressIsZero(AddressIsZero)

ArrayTooLong(ArrayTooLong)

ArraysDifferentLength(ArraysDifferentLength)

CalldataOutOfBounds(CalldataOutOfBounds)

DefaultPermissionRejected(DefaultPermissionRejected)

DelegateCallNotAllowed(DelegateCallNotAllowed)

FunctionSignatureTooShort(FunctionSignatureTooShort)

GranularPermissionRejected(GranularPermissionRejected)

NoMembership(NoMembership)

NodePermissionRejected(NodePermissionRejected)

ParameterNotAllowed(ParameterNotAllowed)

PermissionNotConfigured(PermissionNotConfigured)

SendNotAllowed(SendNotAllowed)

TargetAddressNotAllowed(TargetAddressNotAllowed)

TargetIsNotScoped(TargetIsNotScoped)

TargetIsScoped(TargetIsScoped)

UnacceptableMultiSendOffset(UnacceptableMultiSendOffset)

Implementations

impl HoprCapabilityPermissionsErrors

pub const SELECTORS: &'static [[u8; 4]]

All the selectors of this enum.

Note that the selectors might not be in the same order as the variants. No guarantees are made about the order of the selectors.

Prefer using SolInterface methods instead.