hopr_bindings::hoprchannelsevents

Module HoprChannelsEvents

Generated by the following Solidity interface…

library HoprChannels {
    type Balance is uint96;
    type TicketIndex is uint48;
    type Timestamp is uint32;
}

interface HoprChannelsEvents {
    event ChannelBalanceDecreased(bytes32 indexed channelId, HoprChannels.Balance newBalance);
    event ChannelBalanceIncreased(bytes32 indexed channelId, HoprChannels.Balance newBalance);
    event ChannelClosed(bytes32 indexed channelId);
    event ChannelOpened(address indexed source, address indexed destination);
    event DomainSeparatorUpdated(bytes32 indexed domainSeparator);
    event OutgoingChannelClosureInitiated(bytes32 indexed channelId, HoprChannels.Timestamp closureTime);
    event TicketRedeemed(bytes32 indexed channelId, HoprChannels.TicketIndex newTicketIndex);
}

…which was generated by the following JSON ABI:

[
  {
    "type": "event",
    "name": "ChannelBalanceDecreased",
    "inputs": [
      {
        "name": "channelId",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      },
      {
        "name": "newBalance",
        "type": "uint96",
        "indexed": false,
        "internalType": "HoprChannels.Balance"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "ChannelBalanceIncreased",
    "inputs": [
      {
        "name": "channelId",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      },
      {
        "name": "newBalance",
        "type": "uint96",
        "indexed": false,
        "internalType": "HoprChannels.Balance"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "ChannelClosed",
    "inputs": [
      {
        "name": "channelId",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "ChannelOpened",
    "inputs": [
      {
        "name": "source",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "destination",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "DomainSeparatorUpdated",
    "inputs": [
      {
        "name": "domainSeparator",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "OutgoingChannelClosureInitiated",
    "inputs": [
      {
        "name": "channelId",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      },
      {
        "name": "closureTime",
        "type": "uint32",
        "indexed": false,
        "internalType": "HoprChannels.Timestamp"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "TicketRedeemed",
    "inputs": [
      {
        "name": "channelId",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      },
      {
        "name": "newTicketIndex",
        "type": "uint48",
        "indexed": false,
        "internalType": "HoprChannels.TicketIndex"
      }
    ],
    "anonymous": false
  }
]

Structs

ChannelBalanceDecreased

Event with signature ChannelBalanceDecreased(bytes32,uint96) and selector 0x22e2a422a8860656a3a33cfa1daf771e76798ce5649747957235025de12e0b24.

ChannelBalanceIncreased

Event with signature ChannelBalanceIncreased(bytes32,uint96) and selector 0x5fa17246d3a5d68d42baa94cde33042180b783a399c02bf63ac2076e0f708738.

ChannelClosed

Event with signature ChannelClosed(bytes32) and selector 0xceeab2eef998c17fe96f30f83fbf3c55fc5047f6e40c55a0cf72d236e9d2ba72.

ChannelOpened

Event with signature ChannelOpened(address,address) and selector 0xdd90f938230335e59dc925c57ecb0e27a28c2d87356e31f00cd5554abd6c1b2d.

DomainSeparatorUpdated

Event with signature DomainSeparatorUpdated(bytes32) and selector 0x771f5240ae5fd8a7640d3fb82fa70aab2fb1dbf35f2ef464f8509946717664c5.

HoprChannelsEventsInstance

A HoprChannelsEvents instance.

OutgoingChannelClosureInitiated

Event with signature OutgoingChannelClosureInitiated(bytes32,uint32) and selector 0x07b5c950597fc3bed92e2ad37fa84f701655acb372982e486f5fad3607f04a5c.

TicketRedeemed

Event with signature TicketRedeemed(bytes32,uint48) and selector 0x7165e2ebc7ce35cc98cb7666f9945b3617f3f36326b76d18937ba5fecf18739a.

Enums

HoprChannelsEventsEvents

Container for all the HoprChannelsEvents events.

Statics

BYTECODE

The creation / init bytecode of the contract.

DEPLOYED_BYTECODE

The runtime bytecode of the contract, as deployed on the network.

Functions

deploy

Deploys this contract using the given provider and constructor arguments, if any.

deploy_builder

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

new

Creates a new wrapper around an on-chain HoprChannelsEvents contract instance.