hopr_bindings::hoprcrypto

Module HoprCrypto

Generated by the following Solidity interface…

interface HoprCrypto {
    error InvalidCurvePoint();
    error InvalidFieldElement();
    error InvalidPointWitness();
}

…which was generated by the following JSON ABI:

[
  {
    "type": "error",
    "name": "InvalidCurvePoint",
    "inputs": []
  },
  {
    "type": "error",
    "name": "InvalidFieldElement",
    "inputs": []
  },
  {
    "type": "error",
    "name": "InvalidPointWitness",
    "inputs": []
  }
]

Structs

HoprCryptoInstance

A HoprCrypto instance.

InvalidCurvePoint

Custom error with signature InvalidCurvePoint() and selector 0x72454a82.

InvalidFieldElement

Custom error with signature InvalidFieldElement() and selector 0x3ae4ed6b.

InvalidPointWitness

Custom error with signature InvalidPointWitness() and selector 0xedfdcd98.

Enums

HoprCryptoErrors

Container for all the HoprCrypto custom errors.

Statics

BYTECODE

The creation / init bytecode of the contract.

DEPLOYED_BYTECODE

The runtime bytecode of the contract, as deployed on the network.

Functions

deploy

Deploys this contract using the given provider and constructor arguments, if any.

deploy_builder

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

new

Creates a new wrapper around an on-chain HoprCrypto contract instance.