hopr_bindings::hoprdummyproxyfornetworkregistry

Module HoprDummyProxyForNetworkRegistry

Generated by the following Solidity interface…

interface HoprDummyProxyForNetworkRegistry {
    event AccountDeregistered(address indexed account);
    event AccountRegistered(address indexed account);
    event AllowAllAccountsEligible(bool isAllowed);
    event OwnershipTransferred(address indexed previousOwner, address indexed newOwner);

    constructor(address newOwner);

    function MAX_REGISTRATION_PER_ACCOUNT() external view returns (uint256);
    function canOperateFor(address, address) external pure returns (bool eligiblity);
    function isAllAllowed() external view returns (bool);
    function maxAllowedRegistrations(address account) external view returns (uint256);
    function owner() external view returns (address);
    function ownerAddAccount(address account) external;
    function ownerBatchAddAccounts(address[] memory accounts) external;
    function ownerBatchRemoveAccounts(address[] memory accounts) external;
    function ownerRemoveAccount(address account) external;
    function renounceOwnership() external;
    function transferOwnership(address newOwner) external;
    function updateAllowAll(bool _updatedAllow) external;
}

…which was generated by the following JSON ABI:

[
  {
    "type": "constructor",
    "inputs": [
      {
        "name": "newOwner",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "MAX_REGISTRATION_PER_ACCOUNT",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "canOperateFor",
    "inputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "eligiblity",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "pure"
  },
  {
    "type": "function",
    "name": "isAllAllowed",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "maxAllowedRegistrations",
    "inputs": [
      {
        "name": "account",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "owner",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "ownerAddAccount",
    "inputs": [
      {
        "name": "account",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "ownerBatchAddAccounts",
    "inputs": [
      {
        "name": "accounts",
        "type": "address[]",
        "internalType": "address[]"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "ownerBatchRemoveAccounts",
    "inputs": [
      {
        "name": "accounts",
        "type": "address[]",
        "internalType": "address[]"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "ownerRemoveAccount",
    "inputs": [
      {
        "name": "account",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "renounceOwnership",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "transferOwnership",
    "inputs": [
      {
        "name": "newOwner",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "updateAllowAll",
    "inputs": [
      {
        "name": "_updatedAllow",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "event",
    "name": "AccountDeregistered",
    "inputs": [
      {
        "name": "account",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "AccountRegistered",
    "inputs": [
      {
        "name": "account",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "AllowAllAccountsEligible",
    "inputs": [
      {
        "name": "isAllowed",
        "type": "bool",
        "indexed": false,
        "internalType": "bool"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "OwnershipTransferred",
    "inputs": [
      {
        "name": "previousOwner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "newOwner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  }
]

Structs

AccountDeregistered

Event with signature AccountDeregistered(address) and selector 0x0e63d629afe34b3ca5107c10f90abff5091b31551b371758bd50af76834dc074.

AccountRegistered

Event with signature AccountRegistered(address) and selector 0xcd822dc9688e20acea68724a2fbcfe4f3e526d20ecaa37b18fe3047ab377d6a5.

AllowAllAccountsEligible

Event with signature AllowAllAccountsEligible(bool) and selector 0xafab23a4bc8c49250ba37eeb0625b0a9b271f55d1501838d24f54508c3b17342.

HoprDummyProxyForNetworkRegistryInstance

A HoprDummyProxyForNetworkRegistry instance.

MAX_REGISTRATION_PER_ACCOUNTCall

Function with signature MAX_REGISTRATION_PER_ACCOUNT() and selector 0xf46c84b8.

MAX_REGISTRATION_PER_ACCOUNTReturn

Container type for the return parameters of the MAX_REGISTRATION_PER_ACCOUNT() function.

OwnershipTransferred

Event with signature OwnershipTransferred(address,address) and selector 0x8be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e0.

canOperateForCall

Function with signature canOperateFor(address,address) and selector 0xd85ad044.

canOperateForReturn

Container type for the return parameters of the canOperateFor(address,address) function.

constructorCall

Constructor`.

isAllAllowedCall

Function with signature isAllAllowed() and selector 0xf67f5e6f.

isAllAllowedReturn

Container type for the return parameters of the isAllAllowed() function.

maxAllowedRegistrationsCall

Function with signature maxAllowedRegistrations(address) and selector 0xb3544e82.

maxAllowedRegistrationsReturn

Container type for the return parameters of the maxAllowedRegistrations(address) function.

ownerAddAccountCall

Function with signature ownerAddAccount(address) and selector 0xa71b1b74.

ownerAddAccountReturn

Container type for the return parameters of the ownerAddAccount(address) function.

ownerBatchAddAccountsCall

Function with signature ownerBatchAddAccounts(address[]) and selector 0x486354e2.

ownerBatchAddAccountsReturn

Container type for the return parameters of the ownerBatchAddAccounts(address[]) function.

ownerBatchRemoveAccountsCall

Function with signature ownerBatchRemoveAccounts(address[]) and selector 0x5c5471e7.

ownerBatchRemoveAccountsReturn

Container type for the return parameters of the ownerBatchRemoveAccounts(address[]) function.

ownerCall

Function with signature owner() and selector 0x8da5cb5b.

ownerRemoveAccountCall

Function with signature ownerRemoveAccount(address) and selector 0x4051f257.

ownerRemoveAccountReturn

Container type for the return parameters of the ownerRemoveAccount(address) function.

ownerReturn

Container type for the return parameters of the owner() function.

renounceOwnershipCall

Function with signature renounceOwnership() and selector 0x715018a6.

renounceOwnershipReturn

Container type for the return parameters of the renounceOwnership() function.

transferOwnershipCall

Function with signature transferOwnership(address) and selector 0xf2fde38b.

transferOwnershipReturn

Container type for the return parameters of the transferOwnership(address) function.

updateAllowAllCall

Function with signature updateAllowAll(bool) and selector 0x089ffa47.

updateAllowAllReturn

Container type for the return parameters of the updateAllowAll(bool) function.

Enums

HoprDummyProxyForNetworkRegistryCalls

Container for all the HoprDummyProxyForNetworkRegistry function calls.

HoprDummyProxyForNetworkRegistryEvents

Container for all the HoprDummyProxyForNetworkRegistry events.

Statics

BYTECODE

The creation / init bytecode of the contract.

DEPLOYED_BYTECODE

The runtime bytecode of the contract, as deployed on the network.

Functions

deploy

Deploys this contract using the given provider and constructor arguments, if any.

deploy_builder

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

new

Creates a new wrapper around an on-chain HoprDummyProxyForNetworkRegistry contract instance.