hopr_bindings::hoprdummyproxyfornetworkregistry::HoprDummyProxyForNetworkRegistry

Struct HoprDummyProxyForNetworkRegistryInstance

pub struct HoprDummyProxyForNetworkRegistryInstance<P, N = Ethereum> { /* private fields */ }

A HoprDummyProxyForNetworkRegistry instance.

Contains type-safe methods for interacting with an on-chain instance of the HoprDummyProxyForNetworkRegistry contract located at a given address, using a given provider P.

If the contract bytecode is available (see the sol! documentation on how to provide it), the deploy and deploy_builder methods can be used to deploy a new instance of the contract.

See the module-level documentation for all the available methods.

Implementations

impl<P: Provider<N>, N: Network> HoprDummyProxyForNetworkRegistryInstance<P, N>

Instantiation and getters/setters.

pub const fn new(address: Address, provider: P) -> Self

Creates a new wrapper around an on-chain HoprDummyProxyForNetworkRegistry contract instance.

See the wrapper’s documentation for more details.

pub async fn deploy( provider: P, newOwner: Address, ) -> Result<HoprDummyProxyForNetworkRegistryInstance<P, N>>

Deploys this contract using the given provider and constructor arguments, if any.

Returns a new instance of the contract, if the deployment was successful.

For more fine-grained control over the deployment process, use deploy_builder instead.

pub fn deploy_builder(provider: P, newOwner: Address) -> RawCallBuilder<P, N>

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

This is a simple wrapper around creating a RawCallBuilder with the data set to the bytecode concatenated with the constructor’s ABI-encoded arguments.

pub const fn address(&self) -> &Address

Returns a reference to the address.

pub fn set_address(&mut self, address: Address)

Sets the address.

pub fn at(self, address: Address) -> Self

Sets the address and returns self.

pub const fn provider(&self) -> &P

Returns a reference to the provider.

impl<P: Clone, N> HoprDummyProxyForNetworkRegistryInstance<&P, N>

pub fn with_cloned_provider( self, ) -> HoprDummyProxyForNetworkRegistryInstance<P, N>

Clones the provider and returns a new instance with the cloned provider.

impl<P: Provider<N>, N: Network> HoprDummyProxyForNetworkRegistryInstance<P, N>

Function calls.

pub fn call_builder<C: SolCall>(&self, call: &C) -> SolCallBuilder<&P, C, N>

Creates a new call builder using this contract instance’s provider and address.

Note that the call can be any function call, not just those defined in this contract. Prefer using the other methods for building type-safe contract calls.

pub fn MAX_REGISTRATION_PER_ACCOUNT( &self, ) -> SolCallBuilder<&P, MAX_REGISTRATION_PER_ACCOUNTCall, N>

Creates a new call builder for the [MAX_REGISTRATION_PER_ACCOUNT] function.

pub fn canOperateFor( &self, _0: Address, _1: Address, ) -> SolCallBuilder<&P, canOperateForCall, N>

Creates a new call builder for the [canOperateFor] function.

pub fn isAllAllowed(&self) -> SolCallBuilder<&P, isAllAllowedCall, N>

Creates a new call builder for the [isAllAllowed] function.

pub fn maxAllowedRegistrations( &self, account: Address, ) -> SolCallBuilder<&P, maxAllowedRegistrationsCall, N>

Creates a new call builder for the [maxAllowedRegistrations] function.

pub fn owner(&self) -> SolCallBuilder<&P, ownerCall, N>

Creates a new call builder for the [owner] function.

pub fn ownerAddAccount( &self, account: Address, ) -> SolCallBuilder<&P, ownerAddAccountCall, N>

Creates a new call builder for the [ownerAddAccount] function.

pub fn ownerBatchAddAccounts( &self, accounts: Vec<Address>, ) -> SolCallBuilder<&P, ownerBatchAddAccountsCall, N>

Creates a new call builder for the [ownerBatchAddAccounts] function.

pub fn ownerBatchRemoveAccounts( &self, accounts: Vec<Address>, ) -> SolCallBuilder<&P, ownerBatchRemoveAccountsCall, N>

Creates a new call builder for the [ownerBatchRemoveAccounts] function.

pub fn ownerRemoveAccount( &self, account: Address, ) -> SolCallBuilder<&P, ownerRemoveAccountCall, N>

Creates a new call builder for the [ownerRemoveAccount] function.

pub fn renounceOwnership(&self) -> SolCallBuilder<&P, renounceOwnershipCall, N>

Creates a new call builder for the [renounceOwnership] function.

pub fn transferOwnership( &self, newOwner: Address, ) -> SolCallBuilder<&P, transferOwnershipCall, N>

Creates a new call builder for the [transferOwnership] function.

pub fn updateAllowAll( &self, _updatedAllow: bool, ) -> SolCallBuilder<&P, updateAllowAllCall, N>

Creates a new call builder for the [updateAllowAll] function.

impl<P: Provider<N>, N: Network> HoprDummyProxyForNetworkRegistryInstance<P, N>

Event filters.

pub fn event_filter<E: SolEvent>(&self) -> Event<&P, E, N>

Creates a new event filter using this contract instance’s provider and address.

Note that the type can be any event, not just those defined in this contract. Prefer using the other methods for building type-safe event filters.

pub fn AccountDeregistered_filter(&self) -> Event<&P, AccountDeregistered, N>

Creates a new event filter for the AccountDeregistered event.

pub fn AccountRegistered_filter(&self) -> Event<&P, AccountRegistered, N>

Creates a new event filter for the AccountRegistered event.

pub fn AllowAllAccountsEligible_filter( &self, ) -> Event<&P, AllowAllAccountsEligible, N>

Creates a new event filter for the AllowAllAccountsEligible event.

pub fn OwnershipTransferred_filter(&self) -> Event<&P, OwnershipTransferred, N>

Creates a new event filter for the OwnershipTransferred event.