hopr_bindings::hoprledger::HoprLedger

Enum HoprLedgerCalls

pub enum HoprLedgerCalls {
    LEDGER_VERSION(LEDGER_VERSIONCall),
    ledgerDomainSeparator(ledgerDomainSeparatorCall),
    updateLedgerDomainSeparator(updateLedgerDomainSeparatorCall),
}

Container for all the HoprLedger function calls.

Variants

LEDGER_VERSION(LEDGER_VERSIONCall)

ledgerDomainSeparator(ledgerDomainSeparatorCall)

updateLedgerDomainSeparator(updateLedgerDomainSeparatorCall)

Implementations

impl HoprLedgerCalls

pub const SELECTORS: &'static [[u8; 4]]

All the selectors of this enum.

Note that the selectors might not be in the same order as the variants. No guarantees are made about the order of the selectors.

Prefer using SolInterface methods instead.