hopr_bindings::hoprnetworkregistry::HoprNetworkRegistry

Enum HoprNetworkRegistryCalls

pub enum HoprNetworkRegistryCalls {

    DEFAULT_ADMIN_ROLE(DEFAULT_ADMIN_ROLECall),
    MANAGER_ROLE(MANAGER_ROLECall),
    countRegisterdNodesPerAccount(countRegisterdNodesPerAccountCall),
    disableRegistry(disableRegistryCall),
    enableRegistry(enableRegistryCall),
    enabled(enabledCall),
    getRoleAdmin(getRoleAdminCall),
    getRoleMember(getRoleMemberCall),
    getRoleMemberCount(getRoleMemberCountCall),
    grantRole(grantRoleCall),
    hasRole(hasRoleCall),
    isAccountEligible(isAccountEligibleCall),
    isNodeRegisteredAndEligible(isNodeRegisteredAndEligibleCall),
    isNodeRegisteredByAccount(isNodeRegisteredByAccountCall),
    managerDeregister(managerDeregisterCall),
    managerForceSync(managerForceSyncCall),
    managerRegister(managerRegisterCall),
    managerSync(managerSyncCall),
    maxAdditionalRegistrations(maxAdditionalRegistrationsCall),
    nodeRegisterdToAccount(nodeRegisterdToAccountCall),
    renounceRole(renounceRoleCall),
    requirementImplementation(requirementImplementationCall),
    revokeRole(revokeRoleCall),
    selfDeregister(selfDeregisterCall),
    selfRegister(selfRegisterCall),
    selfSync(selfSyncCall),
    supportsInterface(supportsInterfaceCall),
    updateRequirementImplementation(updateRequirementImplementationCall),

}

Container for all the HoprNetworkRegistry function calls.

Variants

DEFAULT_ADMIN_ROLE(DEFAULT_ADMIN_ROLECall)

MANAGER_ROLE(MANAGER_ROLECall)

countRegisterdNodesPerAccount(countRegisterdNodesPerAccountCall)

disableRegistry(disableRegistryCall)

enableRegistry(enableRegistryCall)

enabled(enabledCall)

getRoleAdmin(getRoleAdminCall)

getRoleMember(getRoleMemberCall)

getRoleMemberCount(getRoleMemberCountCall)

grantRole(grantRoleCall)

hasRole(hasRoleCall)

isAccountEligible(isAccountEligibleCall)

isNodeRegisteredAndEligible(isNodeRegisteredAndEligibleCall)

isNodeRegisteredByAccount(isNodeRegisteredByAccountCall)

managerDeregister(managerDeregisterCall)

managerForceSync(managerForceSyncCall)

managerRegister(managerRegisterCall)

managerSync(managerSyncCall)

maxAdditionalRegistrations(maxAdditionalRegistrationsCall)

nodeRegisterdToAccount(nodeRegisterdToAccountCall)

renounceRole(renounceRoleCall)

requirementImplementation(requirementImplementationCall)

revokeRole(revokeRoleCall)

selfDeregister(selfDeregisterCall)

selfRegister(selfRegisterCall)

selfSync(selfSyncCall)

supportsInterface(supportsInterfaceCall)

updateRequirementImplementation(updateRequirementImplementationCall)

Implementations

impl HoprNetworkRegistryCalls

pub const SELECTORS: &'static [[u8; 4]]

All the selectors of this enum.

Note that the selectors might not be in the same order as the variants. No guarantees are made about the order of the selectors.

Prefer using SolInterface methods instead.