hopr_bindings::hoprnetworkregistry::HoprNetworkRegistry

Struct HoprNetworkRegistryInstance

pub struct HoprNetworkRegistryInstance<P, N = Ethereum> { /* private fields */ }

A HoprNetworkRegistry instance.

Contains type-safe methods for interacting with an on-chain instance of the HoprNetworkRegistry contract located at a given address, using a given provider P.

If the contract bytecode is available (see the sol! documentation on how to provide it), the deploy and deploy_builder methods can be used to deploy a new instance of the contract.

See the module-level documentation for all the available methods.

Implementations

impl<P: Provider<N>, N: Network> HoprNetworkRegistryInstance<P, N>

Instantiation and getters/setters.

pub const fn new(address: Address, provider: P) -> Self

Creates a new wrapper around an on-chain HoprNetworkRegistry contract instance.

See the wrapper’s documentation for more details.

pub async fn deploy( provider: P, _requirementImplementation: Address, _newOwner: Address, _manager: Address, ) -> Result<HoprNetworkRegistryInstance<P, N>>

Deploys this contract using the given provider and constructor arguments, if any.

Returns a new instance of the contract, if the deployment was successful.

For more fine-grained control over the deployment process, use deploy_builder instead.

pub fn deploy_builder( provider: P, _requirementImplementation: Address, _newOwner: Address, _manager: Address, ) -> RawCallBuilder<P, N>

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

This is a simple wrapper around creating a RawCallBuilder with the data set to the bytecode concatenated with the constructor’s ABI-encoded arguments.

pub const fn address(&self) -> &Address

Returns a reference to the address.

pub fn set_address(&mut self, address: Address)

Sets the address.

pub fn at(self, address: Address) -> Self

Sets the address and returns self.

pub const fn provider(&self) -> &P

Returns a reference to the provider.

impl<P: Clone, N> HoprNetworkRegistryInstance<&P, N>

pub fn with_cloned_provider(self) -> HoprNetworkRegistryInstance<P, N>

Clones the provider and returns a new instance with the cloned provider.

impl<P: Provider<N>, N: Network> HoprNetworkRegistryInstance<P, N>

Function calls.

pub fn call_builder<C: SolCall>(&self, call: &C) -> SolCallBuilder<&P, C, N>

Creates a new call builder using this contract instance’s provider and address.

Note that the call can be any function call, not just those defined in this contract. Prefer using the other methods for building type-safe contract calls.

pub fn DEFAULT_ADMIN_ROLE( &self, ) -> SolCallBuilder<&P, DEFAULT_ADMIN_ROLECall, N>

Creates a new call builder for the [DEFAULT_ADMIN_ROLE] function.

pub fn MANAGER_ROLE(&self) -> SolCallBuilder<&P, MANAGER_ROLECall, N>

Creates a new call builder for the [MANAGER_ROLE] function.

pub fn countRegisterdNodesPerAccount( &self, _0: Address, ) -> SolCallBuilder<&P, countRegisterdNodesPerAccountCall, N>

Creates a new call builder for the [countRegisterdNodesPerAccount] function.

pub fn disableRegistry(&self) -> SolCallBuilder<&P, disableRegistryCall, N>

Creates a new call builder for the [disableRegistry] function.

pub fn enableRegistry(&self) -> SolCallBuilder<&P, enableRegistryCall, N>

Creates a new call builder for the [enableRegistry] function.

pub fn enabled(&self) -> SolCallBuilder<&P, enabledCall, N>

Creates a new call builder for the [enabled] function.

pub fn getRoleAdmin( &self, role: FixedBytes<32>, ) -> SolCallBuilder<&P, getRoleAdminCall, N>

Creates a new call builder for the [getRoleAdmin] function.

pub fn getRoleMember( &self, role: FixedBytes<32>, index: U256, ) -> SolCallBuilder<&P, getRoleMemberCall, N>

Creates a new call builder for the [getRoleMember] function.

pub fn getRoleMemberCount( &self, role: FixedBytes<32>, ) -> SolCallBuilder<&P, getRoleMemberCountCall, N>

Creates a new call builder for the [getRoleMemberCount] function.

pub fn grantRole( &self, role: FixedBytes<32>, account: Address, ) -> SolCallBuilder<&P, grantRoleCall, N>

Creates a new call builder for the [grantRole] function.

pub fn hasRole( &self, role: FixedBytes<32>, account: Address, ) -> SolCallBuilder<&P, hasRoleCall, N>

Creates a new call builder for the [hasRole] function.

pub fn isAccountEligible( &self, stakingAccount: Address, ) -> SolCallBuilder<&P, isAccountEligibleCall, N>

Creates a new call builder for the [isAccountEligible] function.

pub fn isNodeRegisteredAndEligible( &self, nodeAddress: Address, ) -> SolCallBuilder<&P, isNodeRegisteredAndEligibleCall, N>

Creates a new call builder for the [isNodeRegisteredAndEligible] function.

pub fn isNodeRegisteredByAccount( &self, nodeAddress: Address, account: Address, ) -> SolCallBuilder<&P, isNodeRegisteredByAccountCall, N>

Creates a new call builder for the [isNodeRegisteredByAccount] function.

pub fn managerDeregister( &self, nodeAddresses: Vec<Address>, ) -> SolCallBuilder<&P, managerDeregisterCall, N>

Creates a new call builder for the [managerDeregister] function.

pub fn managerForceSync( &self, stakingAccounts: Vec<Address>, eligibilities: Vec<bool>, ) -> SolCallBuilder<&P, managerForceSyncCall, N>

Creates a new call builder for the [managerForceSync] function.

pub fn managerRegister( &self, stakingAccounts: Vec<Address>, nodeAddresses: Vec<Address>, ) -> SolCallBuilder<&P, managerRegisterCall, N>

Creates a new call builder for the [managerRegister] function.

pub fn managerSync( &self, stakingAccounts: Vec<Address>, ) -> SolCallBuilder<&P, managerSyncCall, N>

Creates a new call builder for the [managerSync] function.

pub fn maxAdditionalRegistrations( &self, stakingAccount: Address, ) -> SolCallBuilder<&P, maxAdditionalRegistrationsCall, N>

Creates a new call builder for the [maxAdditionalRegistrations] function.

pub fn nodeRegisterdToAccount( &self, _0: Address, ) -> SolCallBuilder<&P, nodeRegisterdToAccountCall, N>

Creates a new call builder for the [nodeRegisterdToAccount] function.

pub fn renounceRole( &self, role: FixedBytes<32>, account: Address, ) -> SolCallBuilder<&P, renounceRoleCall, N>

Creates a new call builder for the [renounceRole] function.

pub fn requirementImplementation( &self, ) -> SolCallBuilder<&P, requirementImplementationCall, N>

Creates a new call builder for the [requirementImplementation] function.

pub fn revokeRole( &self, role: FixedBytes<32>, account: Address, ) -> SolCallBuilder<&P, revokeRoleCall, N>

Creates a new call builder for the [revokeRole] function.

pub fn selfDeregister( &self, nodeAddresses: Vec<Address>, ) -> SolCallBuilder<&P, selfDeregisterCall, N>

Creates a new call builder for the [selfDeregister] function.

pub fn selfRegister( &self, nodeAddresses: Vec<Address>, ) -> SolCallBuilder<&P, selfRegisterCall, N>

Creates a new call builder for the [selfRegister] function.

pub fn selfSync(&self) -> SolCallBuilder<&P, selfSyncCall, N>

Creates a new call builder for the [selfSync] function.

pub fn supportsInterface( &self, interfaceId: FixedBytes<4>, ) -> SolCallBuilder<&P, supportsInterfaceCall, N>

Creates a new call builder for the [supportsInterface] function.

pub fn updateRequirementImplementation( &self, _requirementImplementation: Address, ) -> SolCallBuilder<&P, updateRequirementImplementationCall, N>

Creates a new call builder for the [updateRequirementImplementation] function.

impl<P: Provider<N>, N: Network> HoprNetworkRegistryInstance<P, N>

Event filters.

pub fn event_filter<E: SolEvent>(&self) -> Event<&P, E, N>

Creates a new event filter using this contract instance’s provider and address.

Note that the type can be any event, not just those defined in this contract. Prefer using the other methods for building type-safe event filters.

pub fn Deregistered_filter(&self) -> Event<&P, Deregistered, N>

Creates a new event filter for the Deregistered event.

pub fn DeregisteredByManager_filter( &self, ) -> Event<&P, DeregisteredByManager, N>

Creates a new event filter for the DeregisteredByManager event.

pub fn EligibilityUpdated_filter(&self) -> Event<&P, EligibilityUpdated, N>

Creates a new event filter for the EligibilityUpdated event.

pub fn NetworkRegistryStatusUpdated_filter( &self, ) -> Event<&P, NetworkRegistryStatusUpdated, N>

Creates a new event filter for the NetworkRegistryStatusUpdated event.

pub fn Registered_filter(&self) -> Event<&P, Registered, N>

Creates a new event filter for the Registered event.

pub fn RegisteredByManager_filter(&self) -> Event<&P, RegisteredByManager, N>

Creates a new event filter for the RegisteredByManager event.

pub fn RequirementUpdated_filter(&self) -> Event<&P, RequirementUpdated, N>

Creates a new event filter for the RequirementUpdated event.

pub fn RoleAdminChanged_filter(&self) -> Event<&P, RoleAdminChanged, N>

Creates a new event filter for the RoleAdminChanged event.

pub fn RoleGranted_filter(&self) -> Event<&P, RoleGranted, N>

Creates a new event filter for the RoleGranted event.

pub fn RoleRevoked_filter(&self) -> Event<&P, RoleRevoked, N>

Creates a new event filter for the RoleRevoked event.