hopr_bindings::hoprnetworkregistryevents

Module HoprNetworkRegistryEvents

Generated by the following Solidity interface…

interface HoprNetworkRegistryEvents {
    event Deregistered(address indexed stakingAccount, address indexed nodeAddress);
    event DeregisteredByManager(address indexed stakingAccount, address indexed nodeAddress);
    event EligibilityUpdated(address indexed stakingAccount, bool indexed eligibility);
    event NetworkRegistryStatusUpdated(bool indexed isEnabled);
    event Registered(address indexed stakingAccount, address indexed nodeAddress);
    event RegisteredByManager(address indexed stakingAccount, address indexed nodeAddress);
    event RequirementUpdated(address indexed requirementImplementation);
}

…which was generated by the following JSON ABI:

[
  {
    "type": "event",
    "name": "Deregistered",
    "inputs": [
      {
        "name": "stakingAccount",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "nodeAddress",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "DeregisteredByManager",
    "inputs": [
      {
        "name": "stakingAccount",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "nodeAddress",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "EligibilityUpdated",
    "inputs": [
      {
        "name": "stakingAccount",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "eligibility",
        "type": "bool",
        "indexed": true,
        "internalType": "bool"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "NetworkRegistryStatusUpdated",
    "inputs": [
      {
        "name": "isEnabled",
        "type": "bool",
        "indexed": true,
        "internalType": "bool"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "Registered",
    "inputs": [
      {
        "name": "stakingAccount",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "nodeAddress",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "RegisteredByManager",
    "inputs": [
      {
        "name": "stakingAccount",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "nodeAddress",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "RequirementUpdated",
    "inputs": [
      {
        "name": "requirementImplementation",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  }
]

Structs

Deregistered

Event with signature Deregistered(address,address) and selector 0x159a8ae948fcc6e7f8b5495c0d73e91598440af8a62817246c09edc2c28dba8a.

DeregisteredByManager

Event with signature DeregisteredByManager(address,address) and selector 0x91760a1001f531ec2bee8602be55ad2b1159465edff5bbcd41d20703742a4d42.

EligibilityUpdated

Event with signature EligibilityUpdated(address,bool) and selector 0xe2994f8d6f600ad473dba82c0a890ab7affacb860d3365f474baa3dc04a2e557.

HoprNetworkRegistryEventsInstance

A HoprNetworkRegistryEvents instance.

NetworkRegistryStatusUpdated

Event with signature NetworkRegistryStatusUpdated(bool) and selector 0x8bc85f25e3e933978f099408b3bb3eb7630b76f7455f2e205bcbc0a780e17e72.

Registered

Event with signature Registered(address,address) and selector 0x0a31ee9d46a828884b81003c8498156ea6aa15b9b54bdd0ef0b533d9eba57e55.

RegisteredByManager

Event with signature RegisteredByManager(address,address) and selector 0x04e1eaa4da18f7e742a4edbbe9ede36cdf75f793bb03ab2158c0877f1caf2b93.

RequirementUpdated

Event with signature RequirementUpdated(address) and selector 0x8ac4b2eb7749f75c5b99b898e547fd615dd7a424e68356ea196b7dae742d6c32.

Enums

HoprNetworkRegistryEventsEvents

Container for all the HoprNetworkRegistryEvents events.

Statics

BYTECODE

The creation / init bytecode of the contract.

DEPLOYED_BYTECODE

The runtime bytecode of the contract, as deployed on the network.

Functions

deploy

Deploys this contract using the given provider and constructor arguments, if any.

deploy_builder

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

new

Creates a new wrapper around an on-chain HoprNetworkRegistryEvents contract instance.