hopr_bindings::hoprnodemanagementmodule::HoprNodeManagementModule

Function deploy_builder

pub fn deploy_builder<P: Provider<N>, N: Network>(
    provider: P,
) -> RawCallBuilder<P, N>

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

This is a simple wrapper around creating a RawCallBuilder with the data set to the bytecode concatenated with the constructor’s ABI-encoded arguments.