hopr_bindings::hoprnodemanagementmodule::HoprNodeManagementModule

Struct HoprNodeManagementModuleInstance

pub struct HoprNodeManagementModuleInstance<P, N = Ethereum> { /* private fields */ }

A HoprNodeManagementModule instance.

Contains type-safe methods for interacting with an on-chain instance of the HoprNodeManagementModule contract located at a given address, using a given provider P.

If the contract bytecode is available (see the sol! documentation on how to provide it), the deploy and deploy_builder methods can be used to deploy a new instance of the contract.

See the module-level documentation for all the available methods.

Implementations

impl<P: Provider<N>, N: Network> HoprNodeManagementModuleInstance<P, N>

Instantiation and getters/setters.

pub const fn new(address: Address, provider: P) -> Self

Creates a new wrapper around an on-chain HoprNodeManagementModule contract instance.

See the wrapper’s documentation for more details.

pub async fn deploy( provider: P, ) -> Result<HoprNodeManagementModuleInstance<P, N>>

Deploys this contract using the given provider and constructor arguments, if any.

Returns a new instance of the contract, if the deployment was successful.

For more fine-grained control over the deployment process, use deploy_builder instead.

pub fn deploy_builder(provider: P) -> RawCallBuilder<P, N>

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

This is a simple wrapper around creating a RawCallBuilder with the data set to the bytecode concatenated with the constructor’s ABI-encoded arguments.

pub const fn address(&self) -> &Address

Returns a reference to the address.

pub fn set_address(&mut self, address: Address)

Sets the address.

pub fn at(self, address: Address) -> Self

Sets the address and returns self.

pub const fn provider(&self) -> &P

Returns a reference to the provider.

impl<P: Clone, N> HoprNodeManagementModuleInstance<&P, N>

pub fn with_cloned_provider(self) -> HoprNodeManagementModuleInstance<P, N>

Clones the provider and returns a new instance with the cloned provider.

impl<P: Provider<N>, N: Network> HoprNodeManagementModuleInstance<P, N>

Function calls.

pub fn call_builder<C: SolCall>(&self, call: &C) -> SolCallBuilder<&P, C, N>

Creates a new call builder using this contract instance’s provider and address.

Note that the call can be any function call, not just those defined in this contract. Prefer using the other methods for building type-safe contract calls.

pub fn addChannelsAndTokenTarget( &self, defaultTarget: <Target as SolType>::RustType, ) -> SolCallBuilder<&P, addChannelsAndTokenTargetCall, N>

Creates a new call builder for the [addChannelsAndTokenTarget] function.

pub fn addNode( &self, nodeAddress: Address, ) -> SolCallBuilder<&P, addNodeCall, N>

Creates a new call builder for the [addNode] function.

pub fn decodeFunctionSigsAndPermissions( &self, encoded: FixedBytes<32>, length: U256, ) -> SolCallBuilder<&P, decodeFunctionSigsAndPermissionsCall, N>

Creates a new call builder for the [decodeFunctionSigsAndPermissions] function.

pub fn encodeFunctionSigsAndPermissions( &self, functionSigs: Vec<FixedBytes<4>>, permissions: Vec<<GranularPermission as SolType>::RustType>, ) -> SolCallBuilder<&P, encodeFunctionSigsAndPermissionsCall, N>

Creates a new call builder for the [encodeFunctionSigsAndPermissions] function.

pub fn execTransactionFromModule( &self, to: Address, value: U256, data: Bytes, operation: <Operation as SolType>::RustType, ) -> SolCallBuilder<&P, execTransactionFromModuleCall, N>

Creates a new call builder for the [execTransactionFromModule] function.

pub fn execTransactionFromModuleReturnData( &self, to: Address, value: U256, data: Bytes, operation: <Operation as SolType>::RustType, ) -> SolCallBuilder<&P, execTransactionFromModuleReturnDataCall, N>

Creates a new call builder for the [execTransactionFromModuleReturnData] function.

pub fn getGranularPermissions( &self, capabilityKey: FixedBytes<32>, pairId: FixedBytes<32>, ) -> SolCallBuilder<&P, getGranularPermissionsCall, N>

Creates a new call builder for the [getGranularPermissions] function.

pub fn getTargets(&self) -> SolCallBuilder<&P, getTargetsCall, N>

Creates a new call builder for the [getTargets] function.

pub fn includeNode( &self, nodeDefaultTarget: <Target as SolType>::RustType, ) -> SolCallBuilder<&P, includeNodeCall, N>

Creates a new call builder for the [includeNode] function.

pub fn initialize( &self, initParams: Bytes, ) -> SolCallBuilder<&P, initializeCall, N>

Creates a new call builder for the [initialize] function.

pub fn isHoprNodeManagementModule( &self, ) -> SolCallBuilder<&P, isHoprNodeManagementModuleCall, N>

Creates a new call builder for the [isHoprNodeManagementModule] function.

pub fn isNode(&self, nodeAddress: Address) -> SolCallBuilder<&P, isNodeCall, N>

Creates a new call builder for the [isNode] function.

pub fn multisend(&self) -> SolCallBuilder<&P, multisendCall, N>

Creates a new call builder for the [multisend] function.

pub fn owner(&self) -> SolCallBuilder<&P, ownerCall, N>

Creates a new call builder for the [owner] function.

pub fn proxiableUUID(&self) -> SolCallBuilder<&P, proxiableUUIDCall, N>

Creates a new call builder for the [proxiableUUID] function.

pub fn removeNode( &self, nodeAddress: Address, ) -> SolCallBuilder<&P, removeNodeCall, N>

Creates a new call builder for the [removeNode] function.

pub fn renounceOwnership(&self) -> SolCallBuilder<&P, renounceOwnershipCall, N>

Creates a new call builder for the [renounceOwnership] function.

pub fn revokeTarget( &self, targetAddress: Address, ) -> SolCallBuilder<&P, revokeTargetCall, N>

Creates a new call builder for the [revokeTarget] function.

pub fn scopeChannelsCapabilities( &self, targetAddress: Address, channelId: FixedBytes<32>, encodedSigsPermissions: FixedBytes<32>, ) -> SolCallBuilder<&P, scopeChannelsCapabilitiesCall, N>

Creates a new call builder for the [scopeChannelsCapabilities] function.

pub fn scopeSendCapability( &self, nodeAddress: Address, beneficiary: Address, permission: <GranularPermission as SolType>::RustType, ) -> SolCallBuilder<&P, scopeSendCapabilityCall, N>

Creates a new call builder for the [scopeSendCapability] function.

pub fn scopeTargetChannels( &self, defaultTarget: <Target as SolType>::RustType, ) -> SolCallBuilder<&P, scopeTargetChannelsCall, N>

Creates a new call builder for the [scopeTargetChannels] function.

pub fn scopeTargetSend( &self, defaultTarget: <Target as SolType>::RustType, ) -> SolCallBuilder<&P, scopeTargetSendCall, N>

Creates a new call builder for the [scopeTargetSend] function.

pub fn scopeTargetToken( &self, defaultTarget: <Target as SolType>::RustType, ) -> SolCallBuilder<&P, scopeTargetTokenCall, N>

Creates a new call builder for the [scopeTargetToken] function.

pub fn scopeTokenCapabilities( &self, nodeAddress: Address, targetAddress: Address, beneficiary: Address, encodedSigsPermissions: FixedBytes<32>, ) -> SolCallBuilder<&P, scopeTokenCapabilitiesCall, N>

Creates a new call builder for the [scopeTokenCapabilities] function.

pub fn setMultisend( &self, _multisend: Address, ) -> SolCallBuilder<&P, setMultisendCall, N>

Creates a new call builder for the [setMultisend] function.

pub fn transferOwnership( &self, newOwner: Address, ) -> SolCallBuilder<&P, transferOwnershipCall, N>

Creates a new call builder for the [transferOwnership] function.

pub fn tryGetTarget( &self, targetAddress: Address, ) -> SolCallBuilder<&P, tryGetTargetCall, N>

Creates a new call builder for the [tryGetTarget] function.

pub fn upgradeTo( &self, newImplementation: Address, ) -> SolCallBuilder<&P, upgradeToCall, N>

Creates a new call builder for the [upgradeTo] function.

pub fn upgradeToAndCall( &self, newImplementation: Address, data: Bytes, ) -> SolCallBuilder<&P, upgradeToAndCallCall, N>

Creates a new call builder for the [upgradeToAndCall] function.

impl<P: Provider<N>, N: Network> HoprNodeManagementModuleInstance<P, N>

Event filters.

pub fn event_filter<E: SolEvent>(&self) -> Event<&P, E, N>

Creates a new event filter using this contract instance’s provider and address.

Note that the type can be any event, not just those defined in this contract. Prefer using the other methods for building type-safe event filters.

pub fn AdminChanged_filter(&self) -> Event<&P, AdminChanged, N>

Creates a new event filter for the AdminChanged event.

pub fn BeaconUpgraded_filter(&self) -> Event<&P, BeaconUpgraded, N>

Creates a new event filter for the BeaconUpgraded event.

pub fn ExecutionFailure_filter(&self) -> Event<&P, ExecutionFailure, N>

Creates a new event filter for the ExecutionFailure event.

pub fn ExecutionSuccess_filter(&self) -> Event<&P, ExecutionSuccess, N>

Creates a new event filter for the ExecutionSuccess event.

pub fn Initialized_filter(&self) -> Event<&P, Initialized, N>

Creates a new event filter for the Initialized event.

pub fn NodeAdded_filter(&self) -> Event<&P, NodeAdded, N>

Creates a new event filter for the NodeAdded event.

pub fn NodeRemoved_filter(&self) -> Event<&P, NodeRemoved, N>

Creates a new event filter for the NodeRemoved event.

pub fn OwnershipTransferred_filter(&self) -> Event<&P, OwnershipTransferred, N>

Creates a new event filter for the OwnershipTransferred event.

pub fn SetMultisendAddress_filter(&self) -> Event<&P, SetMultisendAddress, N>

Creates a new event filter for the SetMultisendAddress event.

pub fn Upgraded_filter(&self) -> Event<&P, Upgraded, N>

Creates a new event filter for the Upgraded event.