hopr_bindings::hoprnodesaferegistry

Module HoprNodeSafeRegistry

Generated by the following Solidity interface…

interface HoprNodeSafeRegistry {
    error NodeAddressZero();
    error NodeHasSafe();
    error NodeIsContract();
    error NodeNotModuleMember();
    error NotValidSafe();
    error NotValidSignatureFromNode();
    error SafeAddressZero();

    event DergisteredNodeSafe(address indexed safeAddress, address indexed nodeAddress);
    event DomainSeparatorUpdated(bytes32 indexed domainSeparator);
    event RegisteredNodeSafe(address indexed safeAddress, address indexed nodeAddress);

    constructor();

    function NODE_SAFE_TYPEHASH() external view returns (bytes32);
    function VERSION() external view returns (string memory);
    function deregisterNodeBySafe(address nodeAddr) external;
    function domainSeparator() external view returns (bytes32);
    function isNodeSafeRegistered(address safeAddress, address nodeChainKeyAddress) external view returns (bool);
    function nodeSigNonce(address nodeAddress) external view returns (uint256);
    function nodeToSafe(address nodeAddress) external view returns (address);
    function registerSafeByNode(address safeAddr) external;
    function registerSafeWithNodeSig(address safeAddress, address nodeChainKeyAddress, bytes memory sig) external;
    function updateDomainSeparator() external;
}

…which was generated by the following JSON ABI:

[
  {
    "type": "constructor",
    "inputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "NODE_SAFE_TYPEHASH",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "VERSION",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "string",
        "internalType": "string"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "deregisterNodeBySafe",
    "inputs": [
      {
        "name": "nodeAddr",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "domainSeparator",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "isNodeSafeRegistered",
    "inputs": [
      {
        "name": "safeAddress",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "nodeChainKeyAddress",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "nodeSigNonce",
    "inputs": [
      {
        "name": "nodeAddress",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "nodeToSafe",
    "inputs": [
      {
        "name": "nodeAddress",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "registerSafeByNode",
    "inputs": [
      {
        "name": "safeAddr",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "registerSafeWithNodeSig",
    "inputs": [
      {
        "name": "safeAddress",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "nodeChainKeyAddress",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "sig",
        "type": "bytes",
        "internalType": "bytes"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "updateDomainSeparator",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "event",
    "name": "DergisteredNodeSafe",
    "inputs": [
      {
        "name": "safeAddress",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "nodeAddress",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "DomainSeparatorUpdated",
    "inputs": [
      {
        "name": "domainSeparator",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "RegisteredNodeSafe",
    "inputs": [
      {
        "name": "safeAddress",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "nodeAddress",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "error",
    "name": "NodeAddressZero",
    "inputs": []
  },
  {
    "type": "error",
    "name": "NodeHasSafe",
    "inputs": []
  },
  {
    "type": "error",
    "name": "NodeIsContract",
    "inputs": []
  },
  {
    "type": "error",
    "name": "NodeNotModuleMember",
    "inputs": []
  },
  {
    "type": "error",
    "name": "NotValidSafe",
    "inputs": []
  },
  {
    "type": "error",
    "name": "NotValidSignatureFromNode",
    "inputs": []
  },
  {
    "type": "error",
    "name": "SafeAddressZero",
    "inputs": []
  }
]

Structs

DergisteredNodeSafe

Event with signature DergisteredNodeSafe(address,address) and selector 0xe837f339639296bec9f4bc986154f5ff76bbbefc3149a512920e1311414e3da1.

DomainSeparatorUpdated

Event with signature DomainSeparatorUpdated(bytes32) and selector 0x771f5240ae5fd8a7640d3fb82fa70aab2fb1dbf35f2ef464f8509946717664c5.

HoprNodeSafeRegistryInstance

A HoprNodeSafeRegistry instance.

NODE_SAFE_TYPEHASHCall

Function with signature NODE_SAFE_TYPEHASH() and selector 0x84b0dec2.

NODE_SAFE_TYPEHASHReturn

Container type for the return parameters of the NODE_SAFE_TYPEHASH() function.

NodeAddressZero

Custom error with signature NodeAddressZero() and selector 0x7d0ed526.

NodeHasSafe

Custom error with signature NodeHasSafe() and selector 0x67dc3001.

NodeIsContract

Custom error with signature NodeIsContract() and selector 0x7e0a7c0c.

NodeNotModuleMember

Custom error with signature NodeNotModuleMember() and selector 0xba5d7eb0.

NotValidSafe

Custom error with signature NotValidSafe() and selector 0x56fe8554.

NotValidSignatureFromNode

Custom error with signature NotValidSignatureFromNode() and selector 0xb4c91ffb.

RegisteredNodeSafe

Event with signature RegisteredNodeSafe(address,address) and selector 0x776da40cbf2c223552661bd48502ceb72409b3362bb0c66d8989bd7e20e4497d.

SafeAddressZero

Custom error with signature SafeAddressZero() and selector 0x14e79221.

VERSIONCall

Function with signature VERSION() and selector 0xffa1ad74.

VERSIONReturn

Container type for the return parameters of the VERSION() function.

constructorCall

Constructor`.

deregisterNodeBySafeCall

Function with signature deregisterNodeBySafe(address) and selector 0x91607c4c.

deregisterNodeBySafeReturn

Container type for the return parameters of the deregisterNodeBySafe(address) function.

domainSeparatorCall

Function with signature domainSeparator() and selector 0xf698da25.

domainSeparatorReturn

Container type for the return parameters of the domainSeparator() function.

isNodeSafeRegisteredCall

Function with signature isNodeSafeRegistered(address,address) and selector 0x37480abd.

isNodeSafeRegisteredReturn

Container type for the return parameters of the isNodeSafeRegistered(address,address) function.

nodeSigNonceCall

Function with signature nodeSigNonce(address) and selector 0xaa90f2a5.

nodeSigNonceReturn

Container type for the return parameters of the nodeSigNonce(address) function.

nodeToSafeCall

Function with signature nodeToSafe(address) and selector 0x89978c40.

nodeToSafeReturn

Container type for the return parameters of the nodeToSafe(address) function.

registerSafeByNodeCall

Function with signature registerSafeByNode(address) and selector 0x7f935931.

registerSafeByNodeReturn

Container type for the return parameters of the registerSafeByNode(address) function.

registerSafeWithNodeSigCall

Function with signature registerSafeWithNodeSig(address,address,bytes) and selector 0x49d215e1.

registerSafeWithNodeSigReturn

Container type for the return parameters of the registerSafeWithNodeSig(address,address,bytes) function.

updateDomainSeparatorCall

Function with signature updateDomainSeparator() and selector 0x89ccfe89.

updateDomainSeparatorReturn

Container type for the return parameters of the updateDomainSeparator() function.

Enums

HoprNodeSafeRegistryCalls

Container for all the HoprNodeSafeRegistry function calls.

HoprNodeSafeRegistryErrors

Container for all the HoprNodeSafeRegistry custom errors.

HoprNodeSafeRegistryEvents

Container for all the HoprNodeSafeRegistry events.

Statics

BYTECODE

The creation / init bytecode of the contract.

DEPLOYED_BYTECODE

The runtime bytecode of the contract, as deployed on the network.

Functions

deploy

Deploys this contract using the given provider and constructor arguments, if any.

deploy_builder

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

new

Creates a new wrapper around an on-chain HoprNodeSafeRegistry contract instance.