hopr_bindings::hoprnodesaferegistryevents

Module HoprNodeSafeRegistryEvents

Generated by the following Solidity interface…

interface HoprNodeSafeRegistryEvents {
    event DergisteredNodeSafe(address indexed safeAddress, address indexed nodeAddress);
    event DomainSeparatorUpdated(bytes32 indexed domainSeparator);
    event RegisteredNodeSafe(address indexed safeAddress, address indexed nodeAddress);
}

…which was generated by the following JSON ABI:

[
  {
    "type": "event",
    "name": "DergisteredNodeSafe",
    "inputs": [
      {
        "name": "safeAddress",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "nodeAddress",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "DomainSeparatorUpdated",
    "inputs": [
      {
        "name": "domainSeparator",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "RegisteredNodeSafe",
    "inputs": [
      {
        "name": "safeAddress",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "nodeAddress",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  }
]

Structs

DergisteredNodeSafe

Event with signature DergisteredNodeSafe(address,address) and selector 0xe837f339639296bec9f4bc986154f5ff76bbbefc3149a512920e1311414e3da1.

DomainSeparatorUpdated

Event with signature DomainSeparatorUpdated(bytes32) and selector 0x771f5240ae5fd8a7640d3fb82fa70aab2fb1dbf35f2ef464f8509946717664c5.

HoprNodeSafeRegistryEventsInstance

A HoprNodeSafeRegistryEvents instance.

RegisteredNodeSafe

Event with signature RegisteredNodeSafe(address,address) and selector 0x776da40cbf2c223552661bd48502ceb72409b3362bb0c66d8989bd7e20e4497d.

Enums

HoprNodeSafeRegistryEventsEvents

Container for all the HoprNodeSafeRegistryEvents events.

Statics

BYTECODE

The creation / init bytecode of the contract.

DEPLOYED_BYTECODE

The runtime bytecode of the contract, as deployed on the network.

Functions

deploy

Deploys this contract using the given provider and constructor arguments, if any.

deploy_builder

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

new

Creates a new wrapper around an on-chain HoprNodeSafeRegistryEvents contract instance.