hopr_bindings::hoprnodesaferegistryevents::HoprNodeSafeRegistryEvents

Function deploy

pub fn deploy<P: Provider<N>, N: Network>(
    provider: P,
) -> impl Future<Output = Result<HoprNodeSafeRegistryEventsInstance<P, N>>>

Deploys this contract using the given provider and constructor arguments, if any.

Returns a new instance of the contract, if the deployment was successful.

For more fine-grained control over the deployment process, use deploy_builder instead.