hopr_bindings::hoprnodestakefactory

Module HoprNodeStakeFactory

Generated by the following Solidity interface…

interface HoprNodeStakeFactory {
    error TooFewOwners();

    event NewHoprNodeStakeModule(address indexed moduleImplementation, address instance);
    event NewHoprNodeStakeSafe(address instance);

    constructor();

    function clone(address moduleSingletonAddress, address[] memory admins, uint256 nonce, bytes32 defaultTarget) external returns (address, address payable);
    function predictDeterministicAddress(address implementation, bytes32 salt) external view returns (address predicted);
    function safeVersion() external pure returns (string memory);
}

…which was generated by the following JSON ABI:

[
  {
    "type": "constructor",
    "inputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "clone",
    "inputs": [
      {
        "name": "moduleSingletonAddress",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "admins",
        "type": "address[]",
        "internalType": "address[]"
      },
      {
        "name": "nonce",
        "type": "uint256",
        "internalType": "uint256"
      },
      {
        "name": "defaultTarget",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "",
        "type": "address",
        "internalType": "address payable"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "predictDeterministicAddress",
    "inputs": [
      {
        "name": "implementation",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "salt",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "outputs": [
      {
        "name": "predicted",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "safeVersion",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "string",
        "internalType": "string"
      }
    ],
    "stateMutability": "pure"
  },
  {
    "type": "event",
    "name": "NewHoprNodeStakeModule",
    "inputs": [
      {
        "name": "moduleImplementation",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "instance",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "NewHoprNodeStakeSafe",
    "inputs": [
      {
        "name": "instance",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "error",
    "name": "TooFewOwners",
    "inputs": []
  }
]

Structs

HoprNodeStakeFactoryInstance

A HoprNodeStakeFactory instance.

NewHoprNodeStakeModule

Event with signature NewHoprNodeStakeModule(address,address) and selector 0x813d391dc490d6c1dae7d3fdd555f337533d1da2c908c6efd36d4cf557a63206.

NewHoprNodeStakeSafe

Event with signature NewHoprNodeStakeSafe(address) and selector 0x8231d169f416b666ae7fa43faa24a18899738075a53f32c97617d173b189e386.

TooFewOwners

Custom error with signature TooFewOwners() and selector 0x0d7fd1aa.

cloneCall

Function with signature clone(address,address[],uint256,bytes32) and selector 0x930f426a.

cloneReturn

Container type for the return parameters of the clone(address,address[],uint256,bytes32) function.

constructorCall

Constructor`.

predictDeterministicAddressCall

Function with signature predictDeterministicAddress(address,bytes32) and selector 0x360d0fad.

predictDeterministicAddressReturn

Container type for the return parameters of the predictDeterministicAddress(address,bytes32) function.

safeVersionCall

Function with signature safeVersion() and selector 0x7ef2721f.

safeVersionReturn

Container type for the return parameters of the safeVersion() function.

Enums

HoprNodeStakeFactoryCalls

Container for all the HoprNodeStakeFactory function calls.

HoprNodeStakeFactoryErrors

Container for all the HoprNodeStakeFactory custom errors.

HoprNodeStakeFactoryEvents

Container for all the HoprNodeStakeFactory events.

Statics

BYTECODE

The creation / init bytecode of the contract.

DEPLOYED_BYTECODE

The runtime bytecode of the contract, as deployed on the network.

Functions

deploy

Deploys this contract using the given provider and constructor arguments, if any.

deploy_builder

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

new

Creates a new wrapper around an on-chain HoprNodeStakeFactory contract instance.