hopr_bindings::hoprnodestakefactoryevents

Module HoprNodeStakeFactoryEvents

Generated by the following Solidity interface…

interface HoprNodeStakeFactoryEvents {
    event NewHoprNodeStakeModule(address indexed moduleImplementation, address instance);
    event NewHoprNodeStakeSafe(address instance);
}

…which was generated by the following JSON ABI:

[
  {
    "type": "event",
    "name": "NewHoprNodeStakeModule",
    "inputs": [
      {
        "name": "moduleImplementation",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "instance",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "NewHoprNodeStakeSafe",
    "inputs": [
      {
        "name": "instance",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  }
]

Structs

HoprNodeStakeFactoryEventsInstance

A HoprNodeStakeFactoryEvents instance.

NewHoprNodeStakeModule

Event with signature NewHoprNodeStakeModule(address,address) and selector 0x813d391dc490d6c1dae7d3fdd555f337533d1da2c908c6efd36d4cf557a63206.

NewHoprNodeStakeSafe

Event with signature NewHoprNodeStakeSafe(address) and selector 0x8231d169f416b666ae7fa43faa24a18899738075a53f32c97617d173b189e386.

Enums

HoprNodeStakeFactoryEventsEvents

Container for all the HoprNodeStakeFactoryEvents events.

Statics

BYTECODE

The creation / init bytecode of the contract.

DEPLOYED_BYTECODE

The runtime bytecode of the contract, as deployed on the network.

Functions

deploy

Deploys this contract using the given provider and constructor arguments, if any.

deploy_builder

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

new

Creates a new wrapper around an on-chain HoprNodeStakeFactoryEvents contract instance.