hopr_bindings::hoprstakingproxyfornetworkregistry::HoprStakingProxyForNetworkRegistry

Enum HoprStakingProxyForNetworkRegistryCalls

pub enum HoprStakingProxyForNetworkRegistryCalls {

    canOperateFor(canOperateForCall),
    eligibleNftTypeAndRank(eligibleNftTypeAndRankCall),
    maxAllowedRegistrations(maxAllowedRegistrationsCall),
    maxRegistrationsPerSpecialNft(maxRegistrationsPerSpecialNftCall),
    owner(ownerCall),
    ownerAddNftTypeAndRank(ownerAddNftTypeAndRankCall),
    ownerBatchAddNftTypeAndRank(ownerBatchAddNftTypeAndRankCall),
    ownerBatchAddSpecialNftTypeAndRank(ownerBatchAddSpecialNftTypeAndRankCall),
    ownerBatchRemoveNftTypeAndRank(ownerBatchRemoveNftTypeAndRankCall),
    ownerBatchRemoveSpecialNftTypeAndRank(ownerBatchRemoveSpecialNftTypeAndRankCall),
    ownerRemoveNftTypeAndRank(ownerRemoveNftTypeAndRankCall),
    ownerUpdateThreshold(ownerUpdateThresholdCall),
    renounceOwnership(renounceOwnershipCall),
    specialNftTypeAndRank(specialNftTypeAndRankCall),
    stakeContract(stakeContractCall),
    stakeThreshold(stakeThresholdCall),
    transferOwnership(transferOwnershipCall),
    updateStakeContract(updateStakeContractCall),

}

Container for all the HoprStakingProxyForNetworkRegistry function calls.

Variants

canOperateFor(canOperateForCall)

eligibleNftTypeAndRank(eligibleNftTypeAndRankCall)

maxAllowedRegistrations(maxAllowedRegistrationsCall)

maxRegistrationsPerSpecialNft(maxRegistrationsPerSpecialNftCall)

owner(ownerCall)

ownerAddNftTypeAndRank(ownerAddNftTypeAndRankCall)

ownerBatchAddNftTypeAndRank(ownerBatchAddNftTypeAndRankCall)

ownerBatchAddSpecialNftTypeAndRank(ownerBatchAddSpecialNftTypeAndRankCall)

ownerBatchRemoveNftTypeAndRank(ownerBatchRemoveNftTypeAndRankCall)

ownerBatchRemoveSpecialNftTypeAndRank(ownerBatchRemoveSpecialNftTypeAndRankCall)

ownerRemoveNftTypeAndRank(ownerRemoveNftTypeAndRankCall)

ownerUpdateThreshold(ownerUpdateThresholdCall)

renounceOwnership(renounceOwnershipCall)

specialNftTypeAndRank(specialNftTypeAndRankCall)

stakeContract(stakeContractCall)

stakeThreshold(stakeThresholdCall)

transferOwnership(transferOwnershipCall)

updateStakeContract(updateStakeContractCall)

Implementations

impl HoprStakingProxyForNetworkRegistryCalls

pub const SELECTORS: &'static [[u8; 4]]

All the selectors of this enum.

Note that the selectors might not be in the same order as the variants. No guarantees are made about the order of the selectors.

Prefer using SolInterface methods instead.