hopr_bindings::hoprstakingproxyfornetworkregistry::HoprStakingProxyForNetworkRegistry

Struct HoprStakingProxyForNetworkRegistryInstance

pub struct HoprStakingProxyForNetworkRegistryInstance<P, N = Ethereum> { /* private fields */ }

A HoprStakingProxyForNetworkRegistry instance.

Contains type-safe methods for interacting with an on-chain instance of the HoprStakingProxyForNetworkRegistry contract located at a given address, using a given provider P.

If the contract bytecode is available (see the sol! documentation on how to provide it), the deploy and deploy_builder methods can be used to deploy a new instance of the contract.

See the module-level documentation for all the available methods.

Implementations

impl<P: Provider<N>, N: Network> HoprStakingProxyForNetworkRegistryInstance<P, N>

Instantiation and getters/setters.

pub const fn new(address: Address, provider: P) -> Self

Creates a new wrapper around an on-chain HoprStakingProxyForNetworkRegistry contract instance.

See the wrapper’s documentation for more details.

pub async fn deploy( provider: P, _stakeContract: Address, _newOwner: Address, _minStake: U256, ) -> Result<HoprStakingProxyForNetworkRegistryInstance<P, N>>

Deploys this contract using the given provider and constructor arguments, if any.

Returns a new instance of the contract, if the deployment was successful.

For more fine-grained control over the deployment process, use deploy_builder instead.

pub fn deploy_builder( provider: P, _stakeContract: Address, _newOwner: Address, _minStake: U256, ) -> RawCallBuilder<P, N>

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

This is a simple wrapper around creating a RawCallBuilder with the data set to the bytecode concatenated with the constructor’s ABI-encoded arguments.

pub const fn address(&self) -> &Address

Returns a reference to the address.

pub fn set_address(&mut self, address: Address)

Sets the address.

pub fn at(self, address: Address) -> Self

Sets the address and returns self.

pub const fn provider(&self) -> &P

Returns a reference to the provider.

impl<P: Clone, N> HoprStakingProxyForNetworkRegistryInstance<&P, N>

pub fn with_cloned_provider( self, ) -> HoprStakingProxyForNetworkRegistryInstance<P, N>

Clones the provider and returns a new instance with the cloned provider.

impl<P: Provider<N>, N: Network> HoprStakingProxyForNetworkRegistryInstance<P, N>

Function calls.

pub fn call_builder<C: SolCall>(&self, call: &C) -> SolCallBuilder<&P, C, N>

Creates a new call builder using this contract instance’s provider and address.

Note that the call can be any function call, not just those defined in this contract. Prefer using the other methods for building type-safe contract calls.

pub fn canOperateFor( &self, _0: Address, _1: Address, ) -> SolCallBuilder<&P, canOperateForCall, N>

Creates a new call builder for the [canOperateFor] function.

pub fn eligibleNftTypeAndRank( &self, _0: U256, ) -> SolCallBuilder<&P, eligibleNftTypeAndRankCall, N>

Creates a new call builder for the [eligibleNftTypeAndRank] function.

pub fn maxAllowedRegistrations( &self, account: Address, ) -> SolCallBuilder<&P, maxAllowedRegistrationsCall, N>

Creates a new call builder for the [maxAllowedRegistrations] function.

pub fn maxRegistrationsPerSpecialNft( &self, _0: U256, ) -> SolCallBuilder<&P, maxRegistrationsPerSpecialNftCall, N>

Creates a new call builder for the [maxRegistrationsPerSpecialNft] function.

pub fn owner(&self) -> SolCallBuilder<&P, ownerCall, N>

Creates a new call builder for the [owner] function.

pub fn ownerAddNftTypeAndRank( &self, nftType: U256, nftRank: String, ) -> SolCallBuilder<&P, ownerAddNftTypeAndRankCall, N>

Creates a new call builder for the [ownerAddNftTypeAndRank] function.

pub fn ownerBatchAddNftTypeAndRank( &self, nftTypes: Vec<U256>, nftRanks: Vec<String>, ) -> SolCallBuilder<&P, ownerBatchAddNftTypeAndRankCall, N>

Creates a new call builder for the [ownerBatchAddNftTypeAndRank] function.

pub fn ownerBatchAddSpecialNftTypeAndRank( &self, nftTypes: Vec<U256>, nftRanks: Vec<String>, maxRegistrations: Vec<U256>, ) -> SolCallBuilder<&P, ownerBatchAddSpecialNftTypeAndRankCall, N>

Creates a new call builder for the [ownerBatchAddSpecialNftTypeAndRank] function.

pub fn ownerBatchRemoveNftTypeAndRank( &self, nftTypes: Vec<U256>, nftRanks: Vec<String>, ) -> SolCallBuilder<&P, ownerBatchRemoveNftTypeAndRankCall, N>

Creates a new call builder for the [ownerBatchRemoveNftTypeAndRank] function.

pub fn ownerBatchRemoveSpecialNftTypeAndRank( &self, nftTypes: Vec<U256>, nftRanks: Vec<String>, ) -> SolCallBuilder<&P, ownerBatchRemoveSpecialNftTypeAndRankCall, N>

Creates a new call builder for the [ownerBatchRemoveSpecialNftTypeAndRank] function.

pub fn ownerRemoveNftTypeAndRank( &self, nftType: U256, nftRank: String, ) -> SolCallBuilder<&P, ownerRemoveNftTypeAndRankCall, N>

Creates a new call builder for the [ownerRemoveNftTypeAndRank] function.

pub fn ownerUpdateThreshold( &self, newThreshold: U256, ) -> SolCallBuilder<&P, ownerUpdateThresholdCall, N>

Creates a new call builder for the [ownerUpdateThreshold] function.

pub fn renounceOwnership(&self) -> SolCallBuilder<&P, renounceOwnershipCall, N>

Creates a new call builder for the [renounceOwnership] function.

pub fn specialNftTypeAndRank( &self, _0: U256, ) -> SolCallBuilder<&P, specialNftTypeAndRankCall, N>

Creates a new call builder for the [specialNftTypeAndRank] function.

pub fn stakeContract(&self) -> SolCallBuilder<&P, stakeContractCall, N>

Creates a new call builder for the [stakeContract] function.

pub fn stakeThreshold(&self) -> SolCallBuilder<&P, stakeThresholdCall, N>

Creates a new call builder for the [stakeThreshold] function.

pub fn transferOwnership( &self, newOwner: Address, ) -> SolCallBuilder<&P, transferOwnershipCall, N>

Creates a new call builder for the [transferOwnership] function.

pub fn updateStakeContract( &self, _stakeContract: Address, ) -> SolCallBuilder<&P, updateStakeContractCall, N>

Creates a new call builder for the [updateStakeContract] function.

impl<P: Provider<N>, N: Network> HoprStakingProxyForNetworkRegistryInstance<P, N>

Event filters.

pub fn event_filter<E: SolEvent>(&self) -> Event<&P, E, N>

Creates a new event filter using this contract instance’s provider and address.

Note that the type can be any event, not just those defined in this contract. Prefer using the other methods for building type-safe event filters.

pub fn NftTypeAndRankAdded_filter(&self) -> Event<&P, NftTypeAndRankAdded, N>

Creates a new event filter for the NftTypeAndRankAdded event.

pub fn NftTypeAndRankRemoved_filter( &self, ) -> Event<&P, NftTypeAndRankRemoved, N>

Creates a new event filter for the NftTypeAndRankRemoved event.

pub fn OwnershipTransferred_filter(&self) -> Event<&P, OwnershipTransferred, N>

Creates a new event filter for the OwnershipTransferred event.

pub fn SpecialNftTypeAndRankAdded_filter( &self, ) -> Event<&P, SpecialNftTypeAndRankAdded, N>

Creates a new event filter for the SpecialNftTypeAndRankAdded event.

pub fn SpecialNftTypeAndRankRemoved_filter( &self, ) -> Event<&P, SpecialNftTypeAndRankRemoved, N>

Creates a new event filter for the SpecialNftTypeAndRankRemoved event.

pub fn StakeContractUpdated_filter(&self) -> Event<&P, StakeContractUpdated, N>

Creates a new event filter for the StakeContractUpdated event.

pub fn ThresholdUpdated_filter(&self) -> Event<&P, ThresholdUpdated, N>

Creates a new event filter for the ThresholdUpdated event.