hopr_bindings::hoprticketpriceoracle

Module HoprTicketPriceOracle

Generated by the following Solidity interface…

interface HoprTicketPriceOracle {
    error TicketPriceMustNotBeSame();
    error TicketPriceMustNotBeZero();

    event OwnershipTransferred(address indexed previousOwner, address indexed newOwner);
    event TicketPriceUpdated(uint256, uint256);

    constructor(address _newOwner, uint256 _initialTicketPrice);

    function currentTicketPrice() external view returns (uint256);
    function owner() external view returns (address);
    function renounceOwnership() external;
    function setTicketPrice(uint256 _newTicketPrice) external;
    function transferOwnership(address newOwner) external;
}

…which was generated by the following JSON ABI:

[
  {
    "type": "constructor",
    "inputs": [
      {
        "name": "_newOwner",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "_initialTicketPrice",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "currentTicketPrice",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "owner",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "renounceOwnership",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "setTicketPrice",
    "inputs": [
      {
        "name": "_newTicketPrice",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "transferOwnership",
    "inputs": [
      {
        "name": "newOwner",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "event",
    "name": "OwnershipTransferred",
    "inputs": [
      {
        "name": "previousOwner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "newOwner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "TicketPriceUpdated",
    "inputs": [
      {
        "name": "",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      },
      {
        "name": "",
        "type": "uint256",
        "indexed": false,
        "internalType": "uint256"
      }
    ],
    "anonymous": false
  },
  {
    "type": "error",
    "name": "TicketPriceMustNotBeSame",
    "inputs": []
  },
  {
    "type": "error",
    "name": "TicketPriceMustNotBeZero",
    "inputs": []
  }
]

Structs

HoprTicketPriceOracleInstance

A HoprTicketPriceOracle instance.

OwnershipTransferred

Event with signature OwnershipTransferred(address,address) and selector 0x8be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e0.

TicketPriceMustNotBeSame

Custom error with signature TicketPriceMustNotBeSame() and selector 0x9b24e939.

TicketPriceMustNotBeZero

Custom error with signature TicketPriceMustNotBeZero() and selector 0x45b257ef.

TicketPriceUpdated

Event with signature TicketPriceUpdated(uint256,uint256) and selector 0xd4c5e06b1ae097ba02372652a7adaa6e4a8e00be527497a3ad0ebc3f761ef3fb.

constructorCall

Constructor`.

currentTicketPriceCall

Function with signature currentTicketPrice() and selector 0xe0159106.

currentTicketPriceReturn

Container type for the return parameters of the currentTicketPrice() function.

ownerCall

Function with signature owner() and selector 0x8da5cb5b.

ownerReturn

Container type for the return parameters of the owner() function.

renounceOwnershipCall

Function with signature renounceOwnership() and selector 0x715018a6.

renounceOwnershipReturn

Container type for the return parameters of the renounceOwnership() function.

setTicketPriceCall

Function with signature setTicketPrice(uint256) and selector 0x15981650.

setTicketPriceReturn

Container type for the return parameters of the setTicketPrice(uint256) function.

transferOwnershipCall

Function with signature transferOwnership(address) and selector 0xf2fde38b.

transferOwnershipReturn

Container type for the return parameters of the transferOwnership(address) function.

Enums

HoprTicketPriceOracleCalls

Container for all the HoprTicketPriceOracle function calls.

HoprTicketPriceOracleErrors

Container for all the HoprTicketPriceOracle custom errors.

HoprTicketPriceOracleEvents

Container for all the HoprTicketPriceOracle events.

Statics

BYTECODE

The creation / init bytecode of the contract.

DEPLOYED_BYTECODE

The runtime bytecode of the contract, as deployed on the network.

Functions

deploy

Deploys this contract using the given provider and constructor arguments, if any.

deploy_builder

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

new

Creates a new wrapper around an on-chain HoprTicketPriceOracle contract instance.