hopr_bindings::hoprwinningprobabilityoracle

Module HoprWinningProbabilityOracle

Generated by the following Solidity interface…

interface HoprWinningProbabilityOracle {
    type WinProb is uint56;

    error WinProbMustNotBeSame();

    event OwnershipTransferStarted(address indexed previousOwner, address indexed newOwner);
    event OwnershipTransferred(address indexed previousOwner, address indexed newOwner);
    event WinProbUpdated(WinProb oldWinProb, WinProb newWinProb);

    constructor(address _newOwner, WinProb _initialWinProb);

    function acceptOwnership() external;
    function currentWinProb() external view returns (WinProb);
    function owner() external view returns (address);
    function pendingOwner() external view returns (address);
    function renounceOwnership() external;
    function setWinProb(WinProb _newWinProb) external;
    function transferOwnership(address newOwner) external;
}

…which was generated by the following JSON ABI:

[
  {
    "type": "constructor",
    "inputs": [
      {
        "name": "_newOwner",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "_initialWinProb",
        "type": "uint56",
        "internalType": "WinProb"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "acceptOwnership",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "currentWinProb",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint56",
        "internalType": "WinProb"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "owner",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "pendingOwner",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "renounceOwnership",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "setWinProb",
    "inputs": [
      {
        "name": "_newWinProb",
        "type": "uint56",
        "internalType": "WinProb"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "transferOwnership",
    "inputs": [
      {
        "name": "newOwner",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "event",
    "name": "OwnershipTransferStarted",
    "inputs": [
      {
        "name": "previousOwner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "newOwner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "OwnershipTransferred",
    "inputs": [
      {
        "name": "previousOwner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      },
      {
        "name": "newOwner",
        "type": "address",
        "indexed": true,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "WinProbUpdated",
    "inputs": [
      {
        "name": "oldWinProb",
        "type": "uint56",
        "indexed": false,
        "internalType": "WinProb"
      },
      {
        "name": "newWinProb",
        "type": "uint56",
        "indexed": false,
        "internalType": "WinProb"
      }
    ],
    "anonymous": false
  },
  {
    "type": "error",
    "name": "WinProbMustNotBeSame",
    "inputs": []
  }
]

Structs

HoprWinningProbabilityOracleInstance

A HoprWinningProbabilityOracle instance.

OwnershipTransferStarted

Event with signature OwnershipTransferStarted(address,address) and selector 0x38d16b8cac22d99fc7c124b9cd0de2d3fa1faef420bfe791d8c362d765e22700.

OwnershipTransferred

Event with signature OwnershipTransferred(address,address) and selector 0x8be0079c531659141344cd1fd0a4f28419497f9722a3daafe3b4186f6b6457e0.

WinProb

WinProbMustNotBeSame

Custom error with signature WinProbMustNotBeSame() and selector 0x36bf1fcd.

WinProbUpdated

Event with signature WinProbUpdated(uint56,uint56) and selector 0x02728c4a80373f4765559dc3a046edea88186424827238bb46245668315d3800.

acceptOwnershipCall

Function with signature acceptOwnership() and selector 0x79ba5097.

acceptOwnershipReturn

Container type for the return parameters of the acceptOwnership() function.

constructorCall

Constructor`.

currentWinProbCall

Function with signature currentWinProb() and selector 0x168e9160.

currentWinProbReturn

Container type for the return parameters of the currentWinProb() function.

ownerCall

Function with signature owner() and selector 0x8da5cb5b.

ownerReturn

Container type for the return parameters of the owner() function.

pendingOwnerCall

Function with signature pendingOwner() and selector 0xe30c3978.

pendingOwnerReturn

Container type for the return parameters of the pendingOwner() function.

renounceOwnershipCall

Function with signature renounceOwnership() and selector 0x715018a6.

renounceOwnershipReturn

Container type for the return parameters of the renounceOwnership() function.

setWinProbCall

Function with signature setWinProb(uint56) and selector 0xfde46ff8.

setWinProbReturn

Container type for the return parameters of the setWinProb(uint56) function.

transferOwnershipCall

Function with signature transferOwnership(address) and selector 0xf2fde38b.

transferOwnershipReturn

Container type for the return parameters of the transferOwnership(address) function.

Enums

HoprWinningProbabilityOracleCalls

Container for all the HoprWinningProbabilityOracle function calls.

HoprWinningProbabilityOracleErrors

Container for all the HoprWinningProbabilityOracle custom errors.

HoprWinningProbabilityOracleEvents

Container for all the HoprWinningProbabilityOracle events.

Statics

BYTECODE

The creation / init bytecode of the contract.

DEPLOYED_BYTECODE

The runtime bytecode of the contract, as deployed on the network.

Functions

deploy

Deploys this contract using the given provider and constructor arguments, if any.

deploy_builder

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

new

Creates a new wrapper around an on-chain HoprWinningProbabilityOracle contract instance.