hopr_bindings::hoprwinningprobabilityoracleevents

Module HoprWinningProbabilityOracleEvents

Generated by the following Solidity interface…

interface HoprWinningProbabilityOracleEvents {
    type WinProb is uint56;

    event WinProbUpdated(WinProb oldWinProb, WinProb newWinProb);
}

…which was generated by the following JSON ABI:

[
  {
    "type": "event",
    "name": "WinProbUpdated",
    "inputs": [
      {
        "name": "oldWinProb",
        "type": "uint56",
        "indexed": false,
        "internalType": "WinProb"
      },
      {
        "name": "newWinProb",
        "type": "uint56",
        "indexed": false,
        "internalType": "WinProb"
      }
    ],
    "anonymous": false
  }
]

Structs

HoprWinningProbabilityOracleEventsInstance

A HoprWinningProbabilityOracleEvents instance.

WinProb

WinProbUpdated

Event with signature WinProbUpdated(uint56,uint56) and selector 0x02728c4a80373f4765559dc3a046edea88186424827238bb46245668315d3800.

Enums

HoprWinningProbabilityOracleEventsEvents

Container for all the HoprWinningProbabilityOracleEvents events.

Statics

BYTECODE

The creation / init bytecode of the contract.

DEPLOYED_BYTECODE

The runtime bytecode of the contract, as deployed on the network.

Functions

deploy

Deploys this contract using the given provider and constructor arguments, if any.

deploy_builder

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

new

Creates a new wrapper around an on-chain HoprWinningProbabilityOracleEvents contract instance.