hopr_bindings::hopr_announcements

Module HoprAnnouncements

Generated by the following Solidity interface…

interface HoprAnnouncements {
    type KeyId is uint32;
    struct KeyBindingWithSignature {
        bytes32 ed25519_sig_0;
        bytes32 ed25519_sig_1;
        bytes32 ed25519_pub_key;
        address chain_key;
    }

    error AddressEmptyCode(address target);
    error AlreadyInitialized();
    error ContractNotResponsible();
    error EmptyMultiaddr();
    error ExistingKeyBinding();
    error FailedCall();
    error InvalidSafeAddress();
    error KeyIdOutOfRange();
    error MultiSigUninitialized();
    error ZeroAddress(string reason);
    error ZeroInterval();

    event AddressAnnouncement(address node, string baseMultiaddr);
    event KeyBinding(bytes32 ed25519_sig_0, bytes32 ed25519_sig_1, bytes32 ed25519_pub_key, address chain_key);
    event LedgerDomainSeparatorUpdated(bytes32 indexed ledgerDomainSeparator);
    event RevokeAnnouncement(address node);

    constructor(address safeRegistry);

    function LEDGER_VERSION() external view returns (string memory);
    function SNAPSHOT_INTERVAL() external view returns (uint256);
    function announce(string memory baseMultiaddr) external;
    function announceSafe(address selfAddress, string memory baseMultiaddr) external;
    function bindKeys(bytes32 ed25519_sig_0, bytes32 ed25519_sig_1, bytes32 ed25519_pub_key) external;
    function bindKeysAnnounce(bytes32 ed25519_sig_0, bytes32 ed25519_sig_1, bytes32 ed25519_pub_key, string memory baseMultiaddr) external;
    function bindKeysAnnounceSafe(address selfAddress, bytes32 ed25519_sig_0, bytes32 ed25519_sig_1, bytes32 ed25519_pub_key, string memory baseMultiaddr) external;
    function bindKeysSafe(address selfAddress, bytes32 ed25519_sig_0, bytes32 ed25519_sig_1, bytes32 ed25519_pub_key) external;
    function getAllKeyBindings() external view returns (KeyBindingWithSignature[] memory);
    function getKeyBindingCount() external view returns (uint256);
    function getKeyBindingWithKeyId(KeyId keyId) external view returns (KeyBindingWithSignature memory);
    function getKeyIdRange() external pure returns (uint32 minKeyId, uint32 maxKeyId);
    function getKeyIdWithOffchainKey(bytes32 ed25519_pub_key) external view returns (bool, KeyId);
    function getOffchainKeyWithKeyId(KeyId keyId) external view returns (bytes32 ed25519_pub_key);
    function isOffchainKeyBound(bytes32 ed25519_pub_key) external view returns (bool);
    function latestRoot() external view returns (bytes28 rootHash, uint32 timestamp);
    function latestSnapshotRoot() external view returns (bytes28 rootHash, uint32 timestamp);
    function ledgerDomainSeparator() external view returns (bytes32);
    function multiaddrOf(address) external view returns (string memory);
    function multicall(bytes[] memory data) external returns (bytes[] memory results);
    function revoke() external;
    function revokeSafe(address selfAddress) external;
    function tryGetKeyBinding(bytes32 ed25519_pub_key) external view returns (bool, KeyId, KeyBindingWithSignature memory);
    function updateLedgerDomainSeparator() external;
}

…which was generated by the following JSON ABI:

[
  {
    "type": "constructor",
    "inputs": [
      {
        "name": "safeRegistry",
        "type": "address",
        "internalType": "contract HoprNodeSafeRegistry"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "LEDGER_VERSION",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "string",
        "internalType": "string"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "SNAPSHOT_INTERVAL",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "announce",
    "inputs": [
      {
        "name": "baseMultiaddr",
        "type": "string",
        "internalType": "string"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "announceSafe",
    "inputs": [
      {
        "name": "selfAddress",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "baseMultiaddr",
        "type": "string",
        "internalType": "string"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "bindKeys",
    "inputs": [
      {
        "name": "ed25519_sig_0",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "ed25519_sig_1",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "ed25519_pub_key",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "bindKeysAnnounce",
    "inputs": [
      {
        "name": "ed25519_sig_0",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "ed25519_sig_1",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "ed25519_pub_key",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "baseMultiaddr",
        "type": "string",
        "internalType": "string"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "bindKeysAnnounceSafe",
    "inputs": [
      {
        "name": "selfAddress",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "ed25519_sig_0",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "ed25519_sig_1",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "ed25519_pub_key",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "baseMultiaddr",
        "type": "string",
        "internalType": "string"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "bindKeysSafe",
    "inputs": [
      {
        "name": "selfAddress",
        "type": "address",
        "internalType": "address"
      },
      {
        "name": "ed25519_sig_0",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "ed25519_sig_1",
        "type": "bytes32",
        "internalType": "bytes32"
      },
      {
        "name": "ed25519_pub_key",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "getAllKeyBindings",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "tuple[]",
        "internalType": "struct KeyBindingWithSignature[]",
        "components": [
          {
            "name": "ed25519_sig_0",
            "type": "bytes32",
            "internalType": "bytes32"
          },
          {
            "name": "ed25519_sig_1",
            "type": "bytes32",
            "internalType": "bytes32"
          },
          {
            "name": "ed25519_pub_key",
            "type": "bytes32",
            "internalType": "bytes32"
          },
          {
            "name": "chain_key",
            "type": "address",
            "internalType": "address"
          }
        ]
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getKeyBindingCount",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "uint256",
        "internalType": "uint256"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getKeyBindingWithKeyId",
    "inputs": [
      {
        "name": "keyId",
        "type": "uint32",
        "internalType": "KeyId"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "tuple",
        "internalType": "struct KeyBindingWithSignature",
        "components": [
          {
            "name": "ed25519_sig_0",
            "type": "bytes32",
            "internalType": "bytes32"
          },
          {
            "name": "ed25519_sig_1",
            "type": "bytes32",
            "internalType": "bytes32"
          },
          {
            "name": "ed25519_pub_key",
            "type": "bytes32",
            "internalType": "bytes32"
          },
          {
            "name": "chain_key",
            "type": "address",
            "internalType": "address"
          }
        ]
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getKeyIdRange",
    "inputs": [],
    "outputs": [
      {
        "name": "minKeyId",
        "type": "uint32",
        "internalType": "uint32"
      },
      {
        "name": "maxKeyId",
        "type": "uint32",
        "internalType": "uint32"
      }
    ],
    "stateMutability": "pure"
  },
  {
    "type": "function",
    "name": "getKeyIdWithOffchainKey",
    "inputs": [
      {
        "name": "ed25519_pub_key",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      },
      {
        "name": "",
        "type": "uint32",
        "internalType": "KeyId"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "getOffchainKeyWithKeyId",
    "inputs": [
      {
        "name": "keyId",
        "type": "uint32",
        "internalType": "KeyId"
      }
    ],
    "outputs": [
      {
        "name": "ed25519_pub_key",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "isOffchainKeyBound",
    "inputs": [
      {
        "name": "ed25519_pub_key",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "latestRoot",
    "inputs": [],
    "outputs": [
      {
        "name": "rootHash",
        "type": "bytes28",
        "internalType": "bytes28"
      },
      {
        "name": "timestamp",
        "type": "uint32",
        "internalType": "uint32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "latestSnapshotRoot",
    "inputs": [],
    "outputs": [
      {
        "name": "rootHash",
        "type": "bytes28",
        "internalType": "bytes28"
      },
      {
        "name": "timestamp",
        "type": "uint32",
        "internalType": "uint32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "ledgerDomainSeparator",
    "inputs": [],
    "outputs": [
      {
        "name": "",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "multiaddrOf",
    "inputs": [
      {
        "name": "",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "string",
        "internalType": "string"
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "multicall",
    "inputs": [
      {
        "name": "data",
        "type": "bytes[]",
        "internalType": "bytes[]"
      }
    ],
    "outputs": [
      {
        "name": "results",
        "type": "bytes[]",
        "internalType": "bytes[]"
      }
    ],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "revoke",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "revokeSafe",
    "inputs": [
      {
        "name": "selfAddress",
        "type": "address",
        "internalType": "address"
      }
    ],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "function",
    "name": "tryGetKeyBinding",
    "inputs": [
      {
        "name": "ed25519_pub_key",
        "type": "bytes32",
        "internalType": "bytes32"
      }
    ],
    "outputs": [
      {
        "name": "",
        "type": "bool",
        "internalType": "bool"
      },
      {
        "name": "",
        "type": "uint32",
        "internalType": "KeyId"
      },
      {
        "name": "",
        "type": "tuple",
        "internalType": "struct KeyBindingWithSignature",
        "components": [
          {
            "name": "ed25519_sig_0",
            "type": "bytes32",
            "internalType": "bytes32"
          },
          {
            "name": "ed25519_sig_1",
            "type": "bytes32",
            "internalType": "bytes32"
          },
          {
            "name": "ed25519_pub_key",
            "type": "bytes32",
            "internalType": "bytes32"
          },
          {
            "name": "chain_key",
            "type": "address",
            "internalType": "address"
          }
        ]
      }
    ],
    "stateMutability": "view"
  },
  {
    "type": "function",
    "name": "updateLedgerDomainSeparator",
    "inputs": [],
    "outputs": [],
    "stateMutability": "nonpayable"
  },
  {
    "type": "event",
    "name": "AddressAnnouncement",
    "inputs": [
      {
        "name": "node",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      },
      {
        "name": "baseMultiaddr",
        "type": "string",
        "indexed": false,
        "internalType": "string"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "KeyBinding",
    "inputs": [
      {
        "name": "ed25519_sig_0",
        "type": "bytes32",
        "indexed": false,
        "internalType": "bytes32"
      },
      {
        "name": "ed25519_sig_1",
        "type": "bytes32",
        "indexed": false,
        "internalType": "bytes32"
      },
      {
        "name": "ed25519_pub_key",
        "type": "bytes32",
        "indexed": false,
        "internalType": "bytes32"
      },
      {
        "name": "chain_key",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "LedgerDomainSeparatorUpdated",
    "inputs": [
      {
        "name": "ledgerDomainSeparator",
        "type": "bytes32",
        "indexed": true,
        "internalType": "bytes32"
      }
    ],
    "anonymous": false
  },
  {
    "type": "event",
    "name": "RevokeAnnouncement",
    "inputs": [
      {
        "name": "node",
        "type": "address",
        "indexed": false,
        "internalType": "address"
      }
    ],
    "anonymous": false
  },
  {
    "type": "error",
    "name": "AddressEmptyCode",
    "inputs": [
      {
        "name": "target",
        "type": "address",
        "internalType": "address"
      }
    ]
  },
  {
    "type": "error",
    "name": "AlreadyInitialized",
    "inputs": []
  },
  {
    "type": "error",
    "name": "ContractNotResponsible",
    "inputs": []
  },
  {
    "type": "error",
    "name": "EmptyMultiaddr",
    "inputs": []
  },
  {
    "type": "error",
    "name": "ExistingKeyBinding",
    "inputs": []
  },
  {
    "type": "error",
    "name": "FailedCall",
    "inputs": []
  },
  {
    "type": "error",
    "name": "InvalidSafeAddress",
    "inputs": []
  },
  {
    "type": "error",
    "name": "KeyIdOutOfRange",
    "inputs": []
  },
  {
    "type": "error",
    "name": "MultiSigUninitialized",
    "inputs": []
  },
  {
    "type": "error",
    "name": "ZeroAddress",
    "inputs": [
      {
        "name": "reason",
        "type": "string",
        "internalType": "string"
      }
    ]
  },
  {
    "type": "error",
    "name": "ZeroInterval",
    "inputs": []
  }
]

Structs

AddressAnnouncement

Event with signature AddressAnnouncement(address,string) and selector 0xc4df5ba16814838ab2618829d68f8623bb897302f24dbdba2279dbe45adb3d14.

AddressEmptyCode

Custom error with signature AddressEmptyCode(address) and selector 0x9996b315.

AlreadyInitialized

Custom error with signature AlreadyInitialized() and selector 0x0dc149f0.

ContractNotResponsible

Custom error with signature ContractNotResponsible() and selector 0xacd5a823.

EmptyMultiaddr

Custom error with signature EmptyMultiaddr() and selector 0x09ceac70.

ExistingKeyBinding

Custom error with signature ExistingKeyBinding() and selector 0x0b098de0.

FailedCall

Custom error with signature FailedCall() and selector 0xd6bda275.

HoprAnnouncementsInstance

A HoprAnnouncements instance.

InvalidSafeAddress

Custom error with signature InvalidSafeAddress() and selector 0x8e9d7c5e.

KeyBinding

Event with signature KeyBinding(bytes32,bytes32,bytes32,address) and selector 0xbf49ccd41a4118c7e1e72143bad844e5f6ed12bf60caa642d88d19ca10244c36.

KeyBindingWithSignature

KeyId

KeyIdOutOfRange

Custom error with signature KeyIdOutOfRange() and selector 0x212a77fe.

LEDGER_VERSIONCall

Function with signature LEDGER_VERSION() and selector 0xddad1902.

LEDGER_VERSIONReturn

Container type for the return parameters of the LEDGER_VERSION() function.

LedgerDomainSeparatorUpdated

Event with signature LedgerDomainSeparatorUpdated(bytes32) and selector 0xa43fad83920fd09445855e854e73c9c532e17402c9ceb09993a2392843a5bdb9.

MultiSigUninitialized

Custom error with signature MultiSigUninitialized() and selector 0x454a20c8.

RevokeAnnouncement

Event with signature RevokeAnnouncement(address) and selector 0xa4de30a528becadf82649d1395c0e30dd18ae35b5a96ce71e9295bb14bc9f3bc.

SNAPSHOT_INTERVALCall

Function with signature SNAPSHOT_INTERVAL() and selector 0x6d2beef1.

SNAPSHOT_INTERVALReturn

Container type for the return parameters of the SNAPSHOT_INTERVAL() function.

ZeroAddress

Custom error with signature ZeroAddress(string) and selector 0xeac0d389.

ZeroInterval

Custom error with signature ZeroInterval() and selector 0x346ff607.

announceCall

Function with signature announce(string) and selector 0xea0a5237.

announceReturn

Container type for the return parameters of the announce(string) function.

announceSafeCall

Function with signature announceSafe(address,string) and selector 0xfad0e5a2.

announceSafeReturn

Container type for the return parameters of the announceSafe(address,string) function.

bindKeysAnnounceCall

Function with signature bindKeysAnnounce(bytes32,bytes32,bytes32,string) and selector 0xf884a9cb.

bindKeysAnnounceReturn

Container type for the return parameters of the bindKeysAnnounce(bytes32,bytes32,bytes32,string) function.

bindKeysAnnounceSafeCall

Function with signature bindKeysAnnounceSafe(address,bytes32,bytes32,bytes32,string) and selector 0xddb9dfc1.

bindKeysAnnounceSafeReturn

Container type for the return parameters of the bindKeysAnnounceSafe(address,bytes32,bytes32,bytes32,string) function.

bindKeysCall

Function with signature bindKeys(bytes32,bytes32,bytes32) and selector 0xdbb98d91.

bindKeysReturn

Container type for the return parameters of the bindKeys(bytes32,bytes32,bytes32) function.

bindKeysSafeCall

Function with signature bindKeysSafe(address,bytes32,bytes32,bytes32) and selector 0xcae2b434.

bindKeysSafeReturn

Container type for the return parameters of the bindKeysSafe(address,bytes32,bytes32,bytes32) function.

constructorCall

Constructor`.

getAllKeyBindingsCall

Function with signature getAllKeyBindings() and selector 0x10ab5297.

getAllKeyBindingsReturn

Container type for the return parameters of the getAllKeyBindings() function.

getKeyBindingCountCall

Function with signature getKeyBindingCount() and selector 0xf83e4292.

getKeyBindingCountReturn

Container type for the return parameters of the getKeyBindingCount() function.

getKeyBindingWithKeyIdCall

Function with signature getKeyBindingWithKeyId(uint32) and selector 0xa8b4eec9.

getKeyBindingWithKeyIdReturn

Container type for the return parameters of the getKeyBindingWithKeyId(uint32) function.

getKeyIdRangeCall

Function with signature getKeyIdRange() and selector 0x773b4a33.

getKeyIdRangeReturn

Container type for the return parameters of the getKeyIdRange() function.

getKeyIdWithOffchainKeyCall

Function with signature getKeyIdWithOffchainKey(bytes32) and selector 0x604634c9.

getKeyIdWithOffchainKeyReturn

Container type for the return parameters of the getKeyIdWithOffchainKey(bytes32) function.

getOffchainKeyWithKeyIdCall

Function with signature getOffchainKeyWithKeyId(uint32) and selector 0x4ac3e4f2.

getOffchainKeyWithKeyIdReturn

Container type for the return parameters of the getOffchainKeyWithKeyId(uint32) function.

isOffchainKeyBoundCall

Function with signature isOffchainKeyBound(bytes32) and selector 0xa969635a.

isOffchainKeyBoundReturn

Container type for the return parameters of the isOffchainKeyBound(bytes32) function.

latestRootCall

Function with signature latestRoot() and selector 0xd7b0fef1.

latestRootReturn

Container type for the return parameters of the latestRoot() function.

latestSnapshotRootCall

Function with signature latestSnapshotRoot() and selector 0x0df18f94.

latestSnapshotRootReturn

Container type for the return parameters of the latestSnapshotRoot() function.

ledgerDomainSeparatorCall

Function with signature ledgerDomainSeparator() and selector 0xc966c4fe.

ledgerDomainSeparatorReturn

Container type for the return parameters of the ledgerDomainSeparator() function.

multiaddrOfCall

Function with signature multiaddrOf(address) and selector 0x53665aaa.

multiaddrOfReturn

Container type for the return parameters of the multiaddrOf(address) function.

multicallCall

Function with signature multicall(bytes[]) and selector 0xac9650d8.

multicallReturn

Container type for the return parameters of the multicall(bytes[]) function.

revokeCall

Function with signature revoke() and selector 0xb6549f75.

revokeReturn

Container type for the return parameters of the revoke() function.

revokeSafeCall

Function with signature revokeSafe(address) and selector 0x308c712e.

revokeSafeReturn

Container type for the return parameters of the revokeSafe(address) function.

tryGetKeyBindingCall

Function with signature tryGetKeyBinding(bytes32) and selector 0xa2a07756.

tryGetKeyBindingReturn

Container type for the return parameters of the tryGetKeyBinding(bytes32) function.

updateLedgerDomainSeparatorCall

Function with signature updateLedgerDomainSeparator() and selector 0xdc96fd50.

updateLedgerDomainSeparatorReturn

Container type for the return parameters of the updateLedgerDomainSeparator() function.

Enums

HoprAnnouncementsCalls

Container for all the HoprAnnouncements function calls.

HoprAnnouncementsErrors

Container for all the HoprAnnouncements custom errors.

HoprAnnouncementsEvents

Container for all the HoprAnnouncements events.

Statics

BYTECODE

The creation / init bytecode of the contract.

DEPLOYED_BYTECODE

The runtime bytecode of the contract, as deployed on the network.

Functions

deploy

Deploys this contract using the given provider and constructor arguments, if any.

deploy_builder

Creates a RawCallBuilder for deploying this contract using the given provider and constructor arguments, if any.

new

Creates a new wrapper around an on-chain HoprAnnouncements contract instance.